rustc_middle::dep_graph::dep_node
Enum DepKindDefs

#[repr(u16)]
enum DepKindDefs {

    Null = 0,
    Red = 1,
    TraitSelect = 2,
    CompileCodegenUnit = 3,
    CompileMonoItem = 4,
    trigger_delayed_bug = 5,
    registered_tools = 6,
    early_lint_checks = 7,
    resolutions = 8,
    resolver_for_lowering_raw = 9,
    source_span = 10,
    hir_crate = 11,
    hir_crate_items = 12,
    hir_module_items = 13,
    local_def_id_to_hir_id = 14,
    hir_owner_parent = 15,
    opt_hir_owner_nodes = 16,
    hir_attrs = 17,
    const_param_default = 18,
    type_of = 19,
    type_of_opaque = 20,
    type_alias_is_lazy = 21,
    collect_return_position_impl_trait_in_trait_tys = 22,
    is_type_alias_impl_trait = 23,
    unsizing_params_for_adt = 24,
    analysis = 25,
    check_expectations = 26,
    generics_of = 27,
    predicates_of = 28,
    opaque_types_defined_by = 29,
    explicit_item_bounds = 30,
    explicit_item_super_predicates = 31,
    item_bounds = 32,
    item_super_predicates = 33,
    item_non_self_assumptions = 34,
    impl_super_outlives = 35,
    native_libraries = 36,
    shallow_lint_levels_on = 37,
    lint_expectations = 38,
    expn_that_defined = 39,
    is_panic_runtime = 40,
    representability = 41,
    representability_adt_ty = 42,
    params_in_repr = 43,
    thir_body = 44,
    mir_keys = 45,
    mir_const_qualif = 46,
    mir_built = 47,
    thir_abstract_const = 48,
    mir_drops_elaborated_and_const_checked = 49,
    mir_for_ctfe = 50,
    mir_promoted = 51,
    closure_typeinfo = 52,
    closure_saved_names_of_captured_variables = 53,
    mir_coroutine_witnesses = 54,
    check_coroutine_obligations = 55,
    optimized_mir = 56,
    coverage_attr_on = 57,
    coverage_ids_info = 58,
    promoted_mir = 59,
    erase_regions_ty = 60,
    wasm_import_module_map = 61,
    trait_explicit_predicates_and_bounds = 62,
    explicit_predicates_of = 63,
    inferred_outlives_of = 64,
    explicit_super_predicates_of = 65,
    explicit_implied_predicates_of = 66,
    explicit_supertraits_containing_assoc_item = 67,
    type_param_predicates = 68,
    trait_def = 69,
    adt_def = 70,
    adt_destructor = 71,
    adt_async_destructor = 72,
    adt_sized_constraint = 73,
    adt_dtorck_constraint = 74,
    constness = 75,
    asyncness = 76,
    is_promotable_const_fn = 77,
    coroutine_by_move_body_def_id = 78,
    coroutine_kind = 79,
    coroutine_for_closure = 80,
    crate_variances = 81,
    variances_of = 82,
    inferred_outlives_crate = 83,
    associated_item_def_ids = 84,
    associated_item = 85,
    associated_items = 86,
    impl_item_implementor_ids = 87,
    associated_types_for_impl_traits_in_associated_fn = 88,
    associated_type_for_effects = 89,
    associated_type_for_impl_trait_in_trait = 90,
    impl_trait_header = 91,
    self_ty_of_trait_impl_enabling_order_dep_trait_object_hack = 92,
    inherent_impls = 93,
    incoherent_impls = 94,
    check_unsafety = 95,
    assumed_wf_types = 96,
    assumed_wf_types_for_rpitit = 97,
    fn_sig = 98,
    lint_mod = 99,
    check_unused_traits = 100,
    check_mod_attrs = 101,
    check_mod_unstable_api_usage = 102,
    check_mod_const_bodies = 103,
    check_mod_loops = 104,
    check_mod_naked_functions = 105,
    check_mod_privacy = 106,
    check_liveness = 107,
    live_symbols_and_ignored_derived_traits = 108,
    check_mod_deathness = 109,
    check_mod_type_wf = 110,
    coerce_unsized_info = 111,
    typeck = 112,
    diagnostic_only_typeck = 113,
    used_trait_imports = 114,
    coherent_trait = 115,
    mir_borrowck = 116,
    crate_inherent_impls = 117,
    crate_inherent_impls_validity_check = 118,
    crate_inherent_impls_overlap_check = 119,
    orphan_check_impl = 120,
    mir_callgraph_reachable = 121,
    mir_inliner_callees = 122,
    tag_for_variant = 123,
    eval_to_allocation_raw = 124,
    eval_static_initializer = 125,
    eval_to_const_value_raw = 126,
    eval_to_valtree = 127,
    valtree_to_const_val = 128,
    destructure_const = 129,
    lit_to_const = 130,
    check_match = 131,
    effective_visibilities = 132,
    check_private_in_public = 133,
    reachable_set = 134,
    region_scope_tree = 135,
    mir_shims = 136,
    symbol_name = 137,
    def_kind = 138,
    def_span = 139,
    def_ident_span = 140,
    lookup_stability = 141,
    lookup_const_stability = 142,
    lookup_default_body_stability = 143,
    should_inherit_track_caller = 144,
    lookup_deprecation_entry = 145,
    is_doc_hidden = 146,
    is_doc_notable_trait = 147,
    item_attrs = 148,
    codegen_fn_attrs = 149,
    asm_target_features = 150,
    fn_arg_names = 151,
    rendered_const = 152,
    rendered_precise_capturing_args = 153,
    impl_parent = 154,
    is_ctfe_mir_available = 155,
    is_mir_available = 156,
    own_existential_vtable_entries = 157,
    vtable_entries = 158,
    first_method_vtable_slot = 159,
    supertrait_vtable_slot = 160,
    vtable_allocation = 161,
    codegen_select_candidate = 162,
    all_local_trait_impls = 163,
    trait_impls_of = 164,
    specialization_graph_of = 165,
    dyn_compatibility_violations = 166,
    is_dyn_compatible = 167,
    param_env = 168,
    param_env_reveal_all_normalized = 169,
    is_copy_raw = 170,
    is_sized_raw = 171,
    is_freeze_raw = 172,
    is_unpin_raw = 173,
    needs_drop_raw = 174,
    needs_async_drop_raw = 175,
    has_significant_drop_raw = 176,
    has_structural_eq_impl = 177,
    adt_drop_tys = 178,
    adt_significant_drop_tys = 179,
    layout_of = 180,
    fn_abi_of_fn_ptr = 181,
    fn_abi_of_instance = 182,
    dylib_dependency_formats = 183,
    dependency_formats = 184,
    is_compiler_builtins = 185,
    has_global_allocator = 186,
    has_alloc_error_handler = 187,
    has_panic_handler = 188,
    is_profiler_runtime = 189,
    has_ffi_unwind_calls = 190,
    required_panic_strategy = 191,
    panic_in_drop_strategy = 192,
    is_no_builtins = 193,
    symbol_mangling_version = 194,
    extern_crate = 195,
    specialization_enabled_in = 196,
    specializes = 197,
    in_scope_traits_map = 198,
    defaultness = 199,
    check_well_formed = 200,
    reachable_non_generics = 201,
    is_reachable_non_generic = 202,
    is_unreachable_local_definition = 203,
    upstream_monomorphizations = 204,
    upstream_monomorphizations_for = 205,
    upstream_drop_glue_for = 206,
    upstream_async_drop_glue_for = 207,
    foreign_modules = 208,
    clashing_extern_declarations = 209,
    entry_fn = 210,
    proc_macro_decls_static = 211,
    crate_hash = 212,
    crate_host_hash = 213,
    extra_filename = 214,
    crate_extern_paths = 215,
    implementations_of_trait = 216,
    crate_incoherent_impls = 217,
    native_library = 218,
    inherit_sig_for_delegation_item = 219,
    resolve_bound_vars = 220,
    named_variable_map = 221,
    is_late_bound_map = 222,
    object_lifetime_default = 223,
    late_bound_vars_map = 224,
    visibility = 225,
    inhabited_predicate_adt = 226,
    inhabited_predicate_type = 227,
    dep_kind = 228,
    crate_name = 229,
    module_children = 230,
    extern_mod_stmt_cnum = 231,
    lib_features = 232,
    stability_implications = 233,
    intrinsic_raw = 234,
    get_lang_items = 235,
    all_diagnostic_items = 236,
    defined_lang_items = 237,
    diagnostic_items = 238,
    missing_lang_items = 239,
    visible_parent_map = 240,
    trimmed_def_paths = 241,
    missing_extern_crate_item = 242,
    used_crate_source = 243,
    debugger_visualizers = 244,
    postorder_cnums = 245,
    is_private_dep = 246,
    allocator_kind = 247,
    alloc_error_handler_kind = 248,
    upvars_mentioned = 249,
    maybe_unused_trait_imports = 250,
    names_imported_by_glob_use = 251,
    stability_index = 252,
    crates = 253,
    used_crates = 254,
    traits = 255,
    trait_impls_in_crate = 256,
    exported_symbols = 257,
    collect_and_partition_mono_items = 258,
    is_codegened_item = 259,
    codegen_unit = 260,
    unused_generic_params = 261,
    backend_optimization_level = 262,
    output_filenames = 263,
    normalize_canonicalized_projection_ty = 264,
    normalize_canonicalized_weak_ty = 265,
    normalize_canonicalized_inherent_projection_ty = 266,
    try_normalize_generic_arg_after_erasing_regions = 267,
    implied_outlives_bounds_compat = 268,
    implied_outlives_bounds = 269,
    dropck_outlives = 270,
    evaluate_obligation = 271,
    type_op_ascribe_user_type = 272,
    type_op_prove_predicate = 273,
    type_op_normalize_ty = 274,
    type_op_normalize_clause = 275,
    type_op_normalize_poly_fn_sig = 276,
    type_op_normalize_fn_sig = 277,
    instantiate_and_check_impossible_predicates = 278,
    is_impossible_associated_item = 279,
    method_autoderef_steps = 280,
    supported_target_features = 281,
    implied_target_features = 282,
    features_query = 283,
    crate_for_resolver = 284,
    resolve_instance_raw = 285,
    reveal_opaque_types_in_bounds = 286,
    limits = 287,
    diagnostic_hir_wf_check = 288,
    global_backend_features = 289,
    check_validity_requirement = 290,
    compare_impl_const = 291,
    deduced_param_attrs = 292,
    doc_link_resolutions = 293,
    doc_link_traits_in_scope = 294,
    stripped_cfg_items = 295,
    generics_require_sized_self = 296,
    cross_crate_inlinable = 297,

}

This enum serves as an index into arrays built by make_dep_kind_array.

Variants

Null = 0

We use this for most things when incr. comp. is turned off.

Red = 1

We use this to create a forever-red node.

TraitSelect = 2

CompileCodegenUnit = 3

CompileMonoItem = 4

trigger_delayed_bug = 5

This exists purely for testing the interactions between delayed bugs and incremental.

registered_tools = 6

Collects the list of all tools registered using #![register_tool].

early_lint_checks = 7

[query description - consider adding a doc-comment!] perform lints prior to macro expansion

resolutions = 8

[query description - consider adding a doc-comment!] getting the resolver outputs

resolver_for_lowering_raw = 9

[query description - consider adding a doc-comment!] getting the resolver for lowering

source_span = 10

Return the span for a definition. Contrary to def_span below, this query returns the full absolute span of the definition. This span is meant for dep-tracking rather than diagnostics. It should not be used outside of rustc_middle::hir::source_map.

hir_crate = 11

Represents crate as a whole (as distinct from the top-level crate module). If you call hir_crate (e.g., indirectly by calling tcx.hir().krate()), we will have to assume that any change means that you need to be recompiled. This is because the hir_crate query gives you access to all other items. To avoid this fate, do not call tcx.hir().krate(); instead, prefer wrappers like tcx.visit_all_items_in_krate().

hir_crate_items = 12

All items in the crate.

hir_module_items = 13

The items in a module.

This can be conveniently accessed by tcx.hir().visit_item_likes_in_module. Avoid calling this query directly.

local_def_id_to_hir_id = 14

Returns HIR ID for the given LocalDefId.

hir_owner_parent = 15

Gives access to the HIR node’s parent for the HIR owner key.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

opt_hir_owner_nodes = 16

Gives access to the HIR nodes and bodies inside key if it’s a HIR owner.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

hir_attrs = 17

Gives access to the HIR attributes inside the HIR owner key.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

const_param_default = 18

Given the def_id of a const-generic parameter, computes the associated default const parameter. e.g. fn example<const N: usize=3> called on N would return 3.

type_of = 19

Returns the Ty of the given DefId. If the DefId points to an alias, it will “skip” this alias to return the aliased type.

type_of_opaque = 20

Specialized instance of type_of that detects cycles that are due to revealing opaque because of an auto trait bound. Unless CyclePlaceholder needs to be handled separately, call type_of instead.

type_alias_is_lazy = 21

[query description - consider adding a doc-comment!] computing whether {path} is a lazy type alias

collect_return_position_impl_trait_in_trait_tys = 22

[query description - consider adding a doc-comment!] comparing an impl and trait method signature, inferring any hidden impl Trait types in the process

is_type_alias_impl_trait = 23

[query description - consider adding a doc-comment!] determine whether the opaque is a type-alias impl trait

unsizing_params_for_adt = 24

[query description - consider adding a doc-comment!] determining what parameters of tcx.def_path_str(key) can participate in unsizing

analysis = 25

The root query triggering all analysis passes like typeck or borrowck.

check_expectations = 26

This query checks the fulfillment of collected lint expectations. All lint emitting queries have to be done before this is executed to ensure that all expectations can be fulfilled.

This is an extra query to enable other drivers (like rustdoc) to only execute a small subset of the analysis query, while allowing lints to be expected. In rustc, this query will be executed as part of the analysis query and doesn’t have to be called a second time.

Tools can additionally pass in a tool filter. That will restrict the expectations to only trigger for lints starting with the listed tool name. This is useful for cases were not all linting code from rustc was called. With the default None all registered lints will also be checked for expectation fulfillment.

generics_of = 27

Maps from the DefId of an item (trait/struct/enum/fn) to its associated generics.

predicates_of = 28

Maps from the DefId of an item (trait/struct/enum/fn) to the predicates (where-clauses) that must be proven true in order to reference it. This is almost always the “predicates query” that you want.

opaque_types_defined_by = 29

[query description - consider adding a doc-comment!] computing the opaque types defined by tcx.def_path_str(key.to_def_id())

explicit_item_bounds = 30

Returns the list of bounds that are required to be satisfied by a implementation or definition. For associated types, these must be satisfied for an implementation to be well-formed, and for opaque types, these are required to be satisfied by the hidden-type of the opaque.

Syntactially, these are the bounds written on the trait’s type definition, or those after the impl keyword for an opaque:

type X: Bound + 'lt
//      ^^^^^^^^^^^
impl Debug + Display
//   ^^^^^^^^^^^^^^^

key is the DefId of the associated type or opaque type.

Bounds from the parent (e.g. with nested impl trait) are not included.

explicit_item_super_predicates = 31

The set of item bounds (see TyCtxt::explicit_item_bounds) that share the Self type of the item. These are a subset of the bounds that may explicitly be used for things like closure signature deduction.

item_bounds = 32

Elaborated version of the predicates from explicit_item_bounds.

For example:

trait MyTrait {
    type MyAType: Eq + ?Sized;
}

explicit_item_bounds returns [<Self as MyTrait>::MyAType: Eq], and item_bounds returns

[
    <Self as Trait>::MyAType: Eq,
    <Self as Trait>::MyAType: PartialEq<<Self as Trait>::MyAType>
]

Bounds from the parent (e.g. with nested impl trait) are not included.

item_super_predicates = 33

[query description - consider adding a doc-comment!] elaborating item assumptions for tcx.def_path_str(key)

item_non_self_assumptions = 34

[query description - consider adding a doc-comment!] elaborating item assumptions for tcx.def_path_str(key)

impl_super_outlives = 35

[query description - consider adding a doc-comment!] elaborating supertrait outlives for trait of tcx.def_path_str(key)

native_libraries = 36

Look up all native libraries this crate depends on. These are assembled from the following places:

extern blocks (depending on their link attributes)

the libs (-l) option

shallow_lint_levels_on = 37

[query description - consider adding a doc-comment!] looking up lint levels for tcx.def_path_str(key)

lint_expectations = 38

[query description - consider adding a doc-comment!] computing #[expect]ed lints in this crate

expn_that_defined = 39

[query description - consider adding a doc-comment!] getting the expansion that defined tcx.def_path_str(key)

is_panic_runtime = 40

[query description - consider adding a doc-comment!] checking if the crate is_panic_runtime

representability = 41

Checks whether a type is representable or infinitely sized

representability_adt_ty = 42

An implementation detail for the representability query

params_in_repr = 43

Set of param indexes for type params that are in the type’s representation

thir_body = 44

Fetch the THIR for a given body. If typeck for that body failed, returns an empty Thir.

mir_keys = 45

Set of all the DefIds in this crate that have MIR associated with them. This includes all the body owners, but also things like struct constructors.

mir_const_qualif = 46

Maps DefId’s that have an associated mir::Body to the result of the MIR const-checking pass. This is the set of qualifs in the final value of a const.

mir_built = 47

Build the MIR for a given DefId and prepare it for const qualification.

See the rustc dev guide for more info.

thir_abstract_const = 48

Try to build an abstract representation of the given constant.

mir_drops_elaborated_and_const_checked = 49

[query description - consider adding a doc-comment!] elaborating drops for tcx.def_path_str(key)

mir_for_ctfe = 50

[query description - consider adding a doc-comment!] caching mir of tcx.def_path_str(key) for CTFE

mir_promoted = 51

[query description - consider adding a doc-comment!] promoting constants in MIR for tcx.def_path_str(key)

closure_typeinfo = 52

[query description - consider adding a doc-comment!] finding symbols for captures of closure tcx.def_path_str(key)

closure_saved_names_of_captured_variables = 53

Returns names of captured upvars for closures and coroutines.

Here are some examples:

name__field1__field2 when the upvar is captured by value.

_ref__name__field when the upvar is captured by reference.

For coroutines this only contains upvars that are shared by all states.

mir_coroutine_witnesses = 54

[query description - consider adding a doc-comment!] coroutine witness types for tcx.def_path_str(key)

check_coroutine_obligations = 55

[query description - consider adding a doc-comment!] verify auto trait bounds for coroutine interior type tcx.def_path_str(key)

optimized_mir = 56

MIR after our optimization passes have run. This is MIR that is ready for codegen. This is also the only query that can fetch non-local MIR, at present.

coverage_attr_on = 57

Checks for the nearest #[coverage(off)] or #[coverage(on)] on this def and any enclosing defs, up to the crate root.

Returns false if #[coverage(off)] was found, or true if either #[coverage(on)] or no coverage attribute was found.

coverage_ids_info = 58

Summarizes coverage IDs inserted by the InstrumentCoverage MIR pass (for compiler option -Cinstrument-coverage), after MIR optimizations have had a chance to potentially remove some of them.

promoted_mir = 59

The DefId is the DefId of the containing MIR body. Promoteds do not have their own DefId. This function returns all promoteds in the specified body. The body references promoteds by the DefId and the mir::Promoted index. This is necessary, because after inlining a body may refer to promoteds from other bodies. In that case you still need to use the DefId of the original body.

erase_regions_ty = 60

Erases regions from ty to yield a new type. Normally you would just use tcx.erase_regions(value), however, which uses this query as a kind of cache.

wasm_import_module_map = 61

[query description - consider adding a doc-comment!] getting wasm import module map

trait_explicit_predicates_and_bounds = 62

Returns everything that looks like a predicate written explicitly by the user on a trait item.

Traits are unusual, because predicates on associated types are converted into bounds on that type for backwards compatibility:

trait X where Self::U: Copy { type U; }

becomes

trait X { type U: Copy; }

explicit_predicates_of and explicit_item_bounds will then take the appropriate subsets of the predicates here.

explicit_predicates_of = 63

Returns the predicates written explicitly by the user.

You should probably use predicates_of unless you’re looking for predicates with explicit spans for diagnostics purposes.

inferred_outlives_of = 64

Returns the inferred outlives predicates (e.g., for struct Foo<'a, T> { x: &'a T }, this would return T: 'a).

explicit_super_predicates_of = 65

Maps from the DefId of a trait to the list of super-predicates of the trait, before elaboration (so it doesn’t contain transitive super-predicates). This is a subset of the full list of predicates. We store these in a separate map because we must evaluate them even during type conversion, often before the full predicates are available (note that super-predicates must not be cyclic).

explicit_implied_predicates_of = 66

The predicates of the trait that are implied during elaboration. This is a superset of the super-predicates of the trait, but a subset of the predicates of the trait. For regular traits, this includes all super-predicates and their associated type bounds. For trait aliases, currently, this includes all of the predicates of the trait alias.

explicit_supertraits_containing_assoc_item = 67

The Ident is the name of an associated type.The query returns only the subset of supertraits that define the given associated type. This is used to avoid cycles in resolving type-dependent associated item paths like T::Item.

type_param_predicates = 68

To avoid cycles within the predicates of a single item we compute per-type-parameter predicates for resolving T::AssocTy.

trait_def = 69

[query description - consider adding a doc-comment!] computing trait definition for tcx.def_path_str(key)

adt_def = 70

[query description - consider adding a doc-comment!] computing ADT definition for tcx.def_path_str(key)

adt_destructor = 71

[query description - consider adding a doc-comment!] computing Drop impl for tcx.def_path_str(key)

adt_async_destructor = 72

[query description - consider adding a doc-comment!] computing AsyncDrop impl for tcx.def_path_str(key)

adt_sized_constraint = 73

[query description - consider adding a doc-comment!] computing the Sized constraint for tcx.def_path_str(key)

adt_dtorck_constraint = 74

[query description - consider adding a doc-comment!] computing drop-check constraints for tcx.def_path_str(key)

constness = 75

Returns true if this is a const fn, use the is_const_fn to know whether your crate actually sees it as const fn (e.g., the const-fn-ness might be unstable and you might not have the feature gate active).

Do not call this function manually. It is only meant to cache the base data for the is_const_fn function. Consider using is_const_fn or is_const_fn_raw instead.

asyncness = 76

[query description - consider adding a doc-comment!] checking if the function is async: tcx.def_path_str(key)

is_promotable_const_fn = 77

Returns true if calls to the function may be promoted.

This is either because the function is e.g., a tuple-struct or tuple-variant constructor, or because it has the #[rustc_promotable] attribute. The attribute should be removed in the future in favour of some form of check which figures out whether the function does not inspect the bits of any of its arguments (so is essentially just a constructor function).

coroutine_by_move_body_def_id = 78

The body of the coroutine, modified to take its upvars by move rather than by ref.

This is used by coroutine-closures, which must return a different flavor of coroutine when called using AsyncFnOnce::call_once. It is produced by the ByMoveBody pass which is run right after building the initial MIR, and will only be populated for coroutines which come out of the async closure desugaring.

coroutine_kind = 79

Returns Some(coroutine_kind) if the node pointed to by def_id is a coroutine.

coroutine_for_closure = 80

[query description - consider adding a doc-comment!] Given a coroutine-closure def id, return the def id of the coroutine returned by it

crate_variances = 81

Gets a map with the variance of every item; use variances_of instead.

variances_of = 82

Maps from the DefId of a type or region parameter to its (inferred) variance.

inferred_outlives_crate = 83

Maps from thee DefId of a type to its (inferred) outlives.

associated_item_def_ids = 84

Maps from an impl/trait or struct/variant DefId to a list of the DefIds of its associated items or fields.

associated_item = 85

Maps from a trait/impl item to the trait/impl item “descriptor”.

associated_items = 86

Collects the associated items defined on a trait or impl.

impl_item_implementor_ids = 87

Maps from associated items on a trait to the corresponding associated item on the impl specified by impl_id.

For example, with the following code

struct Type {}
                        // DefId
trait Trait {           // trait_id
    fn f();             // trait_f
    fn g() {}           // trait_g
}

impl Trait for Type {   // impl_id
    fn f() {}           // impl_f
    fn g() {}           // impl_g
}

The map returned for tcx.impl_item_implementor_ids(impl_id) would be { trait_f: impl_f, trait_g: impl_g }

associated_types_for_impl_traits_in_associated_fn = 88

Given fn_def_id of a trait or of an impl that implements a given trait: if fn_def_id is the def id of a function defined inside a trait, then it creates and returns the associated items that correspond to each impl trait in return position for that trait. if fn_def_id is the def id of a function defined inside an impl that implements a trait, then it creates and returns the associated items that correspond to each impl trait in return position of the implemented trait.

associated_type_for_effects = 89

[query description - consider adding a doc-comment!] creating associated items for effects in tcx.def_path_str(def_id)

associated_type_for_impl_trait_in_trait = 90

Given an impl trait in trait opaque_ty_def_id, create and return the corresponding associated item.

impl_trait_header = 91

Given an impl_id, return the trait it implements along with some header information. Return None if this is an inherent impl.

self_ty_of_trait_impl_enabling_order_dep_trait_object_hack = 92

[query description - consider adding a doc-comment!] computing self type wrt issue #33140 tcx.def_path_str(key)

inherent_impls = 93

Maps a DefId of a type to a list of its inherent impls. Contains implementations of methods that are inherent to a type. Methods in these implementations don’t need to be exported.

incoherent_impls = 94

[query description - consider adding a doc-comment!] collecting all inherent impls for {:?}

check_unsafety = 95

Unsafety-check this LocalDefId.

assumed_wf_types = 96

Returns the types assumed to be well formed while “inside” of the given item.

Note that we’ve liberated the late bound regions of function signatures, so this can not be used to check whether these types are well formed.

assumed_wf_types_for_rpitit = 97

We need to store the assumed_wf_types for an RPITIT so that impls of foreign traits with return-position impl trait in traits can inherit the right wf types.

fn_sig = 98

Computes the signature of the function.

lint_mod = 99

Performs lint checking for the module.

check_unused_traits = 100

[query description - consider adding a doc-comment!] checking unused trait imports in crate

check_mod_attrs = 101

Checks the attributes in the module.

check_mod_unstable_api_usage = 102

Checks for uses of unstable APIs in the module.

check_mod_const_bodies = 103

Checks the const bodies in the module for illegal operations (e.g. if or loop).

check_mod_loops = 104

Checks the loops in the module.

check_mod_naked_functions = 105

[query description - consider adding a doc-comment!] checking naked functions in describe_as_module(key, tcx)

check_mod_privacy = 106

[query description - consider adding a doc-comment!] checking privacy in describe_as_module(key.to_local_def_id(), tcx)

check_liveness = 107

[query description - consider adding a doc-comment!] checking liveness of variables in tcx.def_path_str(key)

live_symbols_and_ignored_derived_traits = 108

Return the live symbols in the crate for dead code check.

The second return value maps from ADTs to ignored derived traits (e.g. Debug and Clone) and their respective impl (i.e., part of the derive macro)

check_mod_deathness = 109

[query description - consider adding a doc-comment!] checking deathness of variables in describe_as_module(key, tcx)

check_mod_type_wf = 110

[query description - consider adding a doc-comment!] checking that types are well-formed in describe_as_module(key, tcx)

coerce_unsized_info = 111

Caches CoerceUnsized kinds for impls on custom types.

typeck = 112

[query description - consider adding a doc-comment!] type-checking tcx.def_path_str(key)

diagnostic_only_typeck = 113

[query description - consider adding a doc-comment!] type-checking tcx.def_path_str(key)

used_trait_imports = 114

[query description - consider adding a doc-comment!] finding used_trait_imports tcx.def_path_str(key)

coherent_trait = 115

[query description - consider adding a doc-comment!] coherence checking all impls of trait tcx.def_path_str(def_id)

mir_borrowck = 116

Borrow-checks the function body. If this is a closure, returns additional requirements that the closure’s creator must verify.

crate_inherent_impls = 117

Gets a complete map from all types to their inherent impls. Not meant to be used directly outside of coherence.

crate_inherent_impls_validity_check = 118

Checks all types in the crate for overlap in their inherent impls. Reports errors. Not meant to be used directly outside of coherence.

crate_inherent_impls_overlap_check = 119

Checks all types in the crate for overlap in their inherent impls. Reports errors. Not meant to be used directly outside of coherence.

orphan_check_impl = 120

Checks whether all impls in the crate pass the overlap check, returning which impls fail it. If all impls are correct, the returned slice is empty.

mir_callgraph_reachable = 121

Check whether the function has any recursion that could cause the inliner to trigger a cycle. Returns the call stack causing the cycle. The call stack does not contain the current function, just all intermediate functions.

mir_inliner_callees = 122

Obtain all the calls into other local functions

tag_for_variant = 123

Computes the tag (if any) for a given type and variant.

eval_to_allocation_raw = 124

Evaluates a constant and returns the computed allocation.

Do not use this directly, use the eval_to_const_value or eval_to_valtree instead.

eval_static_initializer = 125

Evaluate a static’s initializer, returning the allocation of the initializer’s memory.

eval_to_const_value_raw = 126

Evaluates const items or anonymous constants (such as enum variant explicit discriminants or array lengths) into a representation suitable for the type system and const generics.

Do not use this directly, use one of the following wrappers: tcx.const_eval_poly, tcx.const_eval_resolve, tcx.const_eval_instance, or tcx.const_eval_global_id.

eval_to_valtree = 127

Evaluate a constant and convert it to a type level constant or return None if that is not possible.

valtree_to_const_val = 128

Converts a type level constant value into ConstValue

destructure_const = 129

Destructures array, ADT or tuple constants into the constants of their fields.

lit_to_const = 130

[query description - consider adding a doc-comment!] converting literal to const

check_match = 131

[query description - consider adding a doc-comment!] match-checking tcx.def_path_str(key)

effective_visibilities = 132

Performs part of the privacy check and computes effective visibilities.

check_private_in_public = 133

[query description - consider adding a doc-comment!] checking for private elements in public interfaces

reachable_set = 134

[query description - consider adding a doc-comment!] reachability

region_scope_tree = 135

Per-body region::ScopeTree. The DefId should be the owner DefId for the body; in the case of closures, this will be redirected to the enclosing function.

mir_shims = 136

Generates a MIR body for the shim.

symbol_name = 137

The symbol_name query provides the symbol name for calling a given instance from the local crate. In particular, it will also look up the correct symbol name of instances from upstream crates.

def_kind = 138

[query description - consider adding a doc-comment!] looking up definition kind of tcx.def_path_str(def_id)

def_span = 139

Gets the span for the definition.

def_ident_span = 140

Gets the span for the identifier of the definition.

lookup_stability = 141

[query description - consider adding a doc-comment!] looking up stability of tcx.def_path_str(def_id)

lookup_const_stability = 142

[query description - consider adding a doc-comment!] looking up const stability of tcx.def_path_str(def_id)

lookup_default_body_stability = 143

[query description - consider adding a doc-comment!] looking up default body stability of tcx.def_path_str(def_id)

should_inherit_track_caller = 144

[query description - consider adding a doc-comment!] computing should_inherit_track_caller of tcx.def_path_str(def_id)

lookup_deprecation_entry = 145

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is deprecated

is_doc_hidden = 146

Determines whether an item is annotated with doc(hidden).

is_doc_notable_trait = 147

Determines whether an item is annotated with doc(notable_trait).

item_attrs = 148

Returns the attributes on the item at def_id.

Do not use this directly, use tcx.get_attrs instead.

codegen_fn_attrs = 149

[query description - consider adding a doc-comment!] computing codegen attributes of tcx.def_path_str(def_id)

asm_target_features = 150

[query description - consider adding a doc-comment!] computing target features for inline asm of tcx.def_path_str(def_id)

fn_arg_names = 151

[query description - consider adding a doc-comment!] looking up function parameter names for tcx.def_path_str(def_id)

rendered_const = 152

Gets the rendered value of the specified constant or associated constant. Used by rustdoc.

rendered_precise_capturing_args = 153

Gets the rendered precise capturing args for an opaque for use in rustdoc.

impl_parent = 154

[query description - consider adding a doc-comment!] computing specialization parent impl of tcx.def_path_str(def_id)

is_ctfe_mir_available = 155

[query description - consider adding a doc-comment!] checking if item has CTFE MIR available: tcx.def_path_str(key)

is_mir_available = 156

[query description - consider adding a doc-comment!] checking if item has MIR available: tcx.def_path_str(key)

own_existential_vtable_entries = 157

[query description - consider adding a doc-comment!] finding all existential vtable entries for trait tcx.def_path_str(key)

vtable_entries = 158

[query description - consider adding a doc-comment!] finding all vtable entries for trait tcx.def_path_str(key.def_id())

first_method_vtable_slot = 159

[query description - consider adding a doc-comment!] finding the slot within the vtable of key.self_ty() for the implementation of key.print_only_trait_name()

supertrait_vtable_slot = 160

[query description - consider adding a doc-comment!] finding the slot within vtable for trait object key.1 vtable ptr during trait upcasting coercion from key.0 vtable

vtable_allocation = 161

[query description - consider adding a doc-comment!] vtable const allocation for < key.0 as key.1.map(| trait_ref | format! ("{trait_ref}")).unwrap_or("_".to_owned()) >

codegen_select_candidate = 162

[query description - consider adding a doc-comment!] computing candidate for key.1

all_local_trait_impls = 163

Return all impl blocks in the current crate.

trait_impls_of = 164

Given a trait trait_id, return all known impl blocks.

specialization_graph_of = 165

[query description - consider adding a doc-comment!] building specialization graph of trait tcx.def_path_str(trait_id)

dyn_compatibility_violations = 166

[query description - consider adding a doc-comment!] determining dyn-compatibility of trait tcx.def_path_str(trait_id)

is_dyn_compatible = 167

[query description - consider adding a doc-comment!] checking if trait tcx.def_path_str(trait_id) is dyn-compatible

param_env = 168

Gets the ParameterEnvironment for a given item; this environment will be in “user-facing” mode, meaning that it is suitable for type-checking etc, and it does not normalize specializable associated types. This is almost always what you want, unless you are doing MIR optimizations, in which case you might want to use reveal_all() method to change modes.

param_env_reveal_all_normalized = 169

Like param_env, but returns the ParamEnv in Reveal::All mode. Prefer this over tcx.param_env(def_id).with_reveal_all_normalized(tcx), as this method is more efficient.

is_copy_raw = 170

Trait selection queries. These are best used by invoking ty.is_copy_modulo_regions(), ty.is_copy(), etc, since that will prune the environment where possible.

is_sized_raw = 171

Query backing Ty::is_sized.

is_freeze_raw = 172

Query backing Ty::is_freeze.

is_unpin_raw = 173

Query backing Ty::is_unpin.

needs_drop_raw = 174

Query backing Ty::needs_drop.

needs_async_drop_raw = 175

Query backing Ty::needs_async_drop.

has_significant_drop_raw = 176

Query backing Ty::has_significant_drop_raw.

has_structural_eq_impl = 177

Query backing Ty::is_structural_eq_shallow.

This is only correct for ADTs. Call is_structural_eq_shallow to handle all types correctly.

adt_drop_tys = 178

A list of types where the ADT requires drop if and only if any of those types require drop. If the ADT is known to always need drop then Err(AlwaysRequiresDrop) is returned.

adt_significant_drop_tys = 179

A list of types where the ADT requires drop if and only if any of those types has significant drop. A type marked with the attribute rustc_insignificant_dtor is considered to not be significant. A drop is significant if it is implemented by the user or does anything that will have any observable behavior (other than freeing up memory). If the ADT is known to have a significant destructor then Err(AlwaysRequiresDrop) is returned.

layout_of = 180

Computes the layout of a type. Note that this implicitly executes in “reveal all” mode, and will normalize the input type.

fn_abi_of_fn_ptr = 181

Compute a FnAbi suitable for indirect calls, i.e. to fn pointers.

NB: this doesn’t handle virtual calls - those should use fn_abi_of_instance instead, where the instance is an InstanceKind::Virtual.

fn_abi_of_instance = 182

Compute a FnAbi suitable for declaring/defining an fn instance, and for direct calls to an fn.

NB: that includes virtual calls, which are represented by “direct calls” to an InstanceKind::Virtual instance (of <dyn Trait as Trait>::fn).

dylib_dependency_formats = 183

[query description - consider adding a doc-comment!] getting dylib dependency formats of crate

dependency_formats = 184

[query description - consider adding a doc-comment!] getting the linkage format of all dependencies

is_compiler_builtins = 185

[query description - consider adding a doc-comment!] checking if the crate is_compiler_builtins

has_global_allocator = 186

[query description - consider adding a doc-comment!] checking if the crate has_global_allocator

has_alloc_error_handler = 187

[query description - consider adding a doc-comment!] checking if the crate has_alloc_error_handler

has_panic_handler = 188

[query description - consider adding a doc-comment!] checking if the crate has_panic_handler

is_profiler_runtime = 189

[query description - consider adding a doc-comment!] checking if a crate is #![profiler_runtime]

has_ffi_unwind_calls = 190

[query description - consider adding a doc-comment!] checking if tcx.def_path_str(key) contains FFI-unwind calls

required_panic_strategy = 191

[query description - consider adding a doc-comment!] getting a crate’s required panic strategy

panic_in_drop_strategy = 192

[query description - consider adding a doc-comment!] getting a crate’s configured panic-in-drop strategy

is_no_builtins = 193

[query description - consider adding a doc-comment!] getting whether a crate has #![no_builtins]

symbol_mangling_version = 194

[query description - consider adding a doc-comment!] getting a crate’s symbol mangling version

extern_crate = 195

[query description - consider adding a doc-comment!] getting crate’s ExternCrateData

specialization_enabled_in = 196

[query description - consider adding a doc-comment!] checking whether the crate enabled specialization/min_specialization

specializes = 197

[query description - consider adding a doc-comment!] computing whether impls specialize one another

in_scope_traits_map = 198

[query description - consider adding a doc-comment!] getting traits in scope at a block

defaultness = 199

Returns whether the impl or associated function has the default keyword.

check_well_formed = 200

[query description - consider adding a doc-comment!] checking that tcx.def_path_str(key) is well-formed

reachable_non_generics = 201

[query description - consider adding a doc-comment!] looking up the exported symbols of a crate

is_reachable_non_generic = 202

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is an exported symbol

is_unreachable_local_definition = 203

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is reachable from outside the crate

upstream_monomorphizations = 204

The entire set of monomorphizations the local crate can safely link to because they are exported from upstream crates. Do not depend on this directly, as its value changes anytime a monomorphization gets added or removed in any upstream crate. Instead use the narrower upstream_monomorphizations_for, upstream_drop_glue_for, upstream_async_drop_glue_for, or, even better, Instance::upstream_monomorphization().

upstream_monomorphizations_for = 205

Returns the set of upstream monomorphizations available for the generic function identified by the given def_id. The query makes sure to make a stable selection if the same monomorphization is available in multiple upstream crates.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

upstream_drop_glue_for = 206

Returns the upstream crate that exports drop-glue for the given type (args is expected to be a single-item list containing the type one wants drop-glue for).

This is a subset of upstream_monomorphizations_for in order to increase dep-tracking granularity. Otherwise adding or removing any type with drop-glue in any upstream crate would invalidate all functions calling drop-glue of an upstream type.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

NOTE: This query could easily be extended to also support other common functions that have are large set of monomorphizations (like Clone::clone for example).

upstream_async_drop_glue_for = 207

Returns the upstream crate that exports async-drop-glue for the given type (args is expected to be a single-item list containing the type one wants async-drop-glue for).

This is a subset of upstream_monomorphizations_for in order to increase dep-tracking granularity. Otherwise adding or removing any type with async-drop-glue in any upstream crate would invalidate all functions calling async-drop-glue of an upstream type.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

NOTE: This query could easily be extended to also support other common functions that have are large set of monomorphizations (like Clone::clone for example).

foreign_modules = 208

Returns a list of all extern blocks of a crate.

clashing_extern_declarations = 209

Lint against extern fn declarations having incompatible types.

entry_fn = 210

Identifies the entry-point (e.g., the main function) for a given crate, returning None if there is no entry point (such as for library crates).

proc_macro_decls_static = 211

Finds the rustc_proc_macro_decls item of a crate.

crate_hash = 212

[query description - consider adding a doc-comment!] looking up the hash a crate

crate_host_hash = 213

Gets the hash for the host proc macro. Used to support -Z dual-proc-macro.

extra_filename = 214

Gets the extra data to put in each output filename for a crate. For example, compiling the foo crate with extra-filename=-a creates a libfoo-b.rlib file.

crate_extern_paths = 215

Gets the paths where the crate came from in the file system.

implementations_of_trait = 216

Given a crate and a trait, look up all impls of that trait in the crate. Return (impl_id, self_ty).

crate_incoherent_impls = 217

Collects all incoherent impls for the given crate and type.

Do not call this directly, but instead use the incoherent_impls query. This query is only used to get the data necessary for that query.

native_library = 218

Get the corresponding native library from the native_libraries query

inherit_sig_for_delegation_item = 219

[query description - consider adding a doc-comment!] inheriting delegation signature

resolve_bound_vars = 220

Does lifetime resolution on items. Importantly, we can’t resolve lifetimes directly on things like trait methods, because of trait params. See rustc_resolve::late::lifetimes for details.

named_variable_map = 221

[query description - consider adding a doc-comment!] looking up a named region inside tcx.def_path_str(owner_id)

is_late_bound_map = 222

[query description - consider adding a doc-comment!] testing if a region is late bound inside tcx.def_path_str(owner_id)

object_lifetime_default = 223

For a given item’s generic parameter, gets the default lifetimes to be used for each parameter if a trait object were to be passed for that parameter. For example, for T in struct Foo<'a, T>, this would be 'static. For T in struct Foo<'a, T: 'a>, this would instead be 'a. This query will panic if passed something that is not a type parameter.

late_bound_vars_map = 224

[query description - consider adding a doc-comment!] looking up late bound vars inside tcx.def_path_str(owner_id)

visibility = 225

Computes the visibility of the provided def_id.

If the item from the def_id doesn’t have a visibility, it will panic. For example a generic type parameter will panic if you call this method on it:

use std::fmt::Debug;

pub trait Foo<T: Debug> {}

In here, if you call visibility on T, it’ll panic.

inhabited_predicate_adt = 226

[query description - consider adding a doc-comment!] computing the uninhabited predicate of {:?}

inhabited_predicate_type = 227

Do not call this query directly: invoke Ty::inhabited_predicate instead.

dep_kind = 228

[query description - consider adding a doc-comment!] fetching what a dependency looks like

crate_name = 229

Gets the name of the crate.

module_children = 230

[query description - consider adding a doc-comment!] collecting child items of module tcx.def_path_str(def_id)

extern_mod_stmt_cnum = 231

[query description - consider adding a doc-comment!] computing crate imported by tcx.def_path_str(def_id)

lib_features = 232

[query description - consider adding a doc-comment!] calculating the lib features defined in a crate

stability_implications = 233

[query description - consider adding a doc-comment!] calculating the implications between #[unstable] features defined in a crate

intrinsic_raw = 234

Whether the function is an intrinsic

get_lang_items = 235

Returns the lang items defined in another crate by loading it from metadata.

all_diagnostic_items = 236

Returns all diagnostic items defined in all crates.

defined_lang_items = 237

Returns the lang items defined in another crate by loading it from metadata.

diagnostic_items = 238

Returns the diagnostic items defined in a crate.

missing_lang_items = 239

[query description - consider adding a doc-comment!] calculating the missing lang items in a crate

visible_parent_map = 240

The visible parent map is a map from every item to a visible parent. It prefers the shortest visible path to an item. Used for diagnostics, for example path trimming. The parents are modules, enums or traits.

trimmed_def_paths = 241

Collects the “trimmed”, shortest accessible paths to all items for diagnostics. See the provider docs for more info.

missing_extern_crate_item = 242

[query description - consider adding a doc-comment!] seeing if we’re missing an extern crate item for this crate

used_crate_source = 243

[query description - consider adding a doc-comment!] looking at the source for a crate

debugger_visualizers = 244

Returns the debugger visualizers defined for this crate. NOTE: This query has to be marked eval_always because it reads data directly from disk that is not tracked anywhere else. I.e. it represents a genuine input to the query system.

postorder_cnums = 245

[query description - consider adding a doc-comment!] generating a postorder list of CrateNums

is_private_dep = 246

Returns whether or not the crate with CrateNum ‘cnum’ is marked as a private dependency

allocator_kind = 247

[query description - consider adding a doc-comment!] getting the allocator kind for the current crate

alloc_error_handler_kind = 248

[query description - consider adding a doc-comment!] alloc error handler kind for the current crate

upvars_mentioned = 249

[query description - consider adding a doc-comment!] collecting upvars mentioned in tcx.def_path_str(def_id)

maybe_unused_trait_imports = 250

[query description - consider adding a doc-comment!] fetching potentially unused trait imports

names_imported_by_glob_use = 251

[query description - consider adding a doc-comment!] finding names imported by glob use for tcx.def_path_str(def_id)

stability_index = 252

[query description - consider adding a doc-comment!] calculating the stability index for the local crate

crates = 253

[query description - consider adding a doc-comment!] fetching all foreign CrateNum instances

used_crates = 254

[query description - consider adding a doc-comment!] fetching CrateNums for all crates loaded non-speculatively

traits = 255

A list of all traits in a crate, used by rustdoc and error reporting.

trait_impls_in_crate = 256

[query description - consider adding a doc-comment!] fetching all trait impls in a crate

exported_symbols = 257

The list of symbols exported from the given crate.

All names contained in exported_symbols(cnum) are guaranteed to correspond to a publicly visible symbol in cnum machine code.

The exported_symbols sets of different crates do not intersect.

collect_and_partition_mono_items = 258

[query description - consider adding a doc-comment!] collect_and_partition_mono_items

is_codegened_item = 259

[query description - consider adding a doc-comment!] determining whether tcx.def_path_str(def_id) needs codegen

codegen_unit = 260

[query description - consider adding a doc-comment!] getting codegen unit {sym}

unused_generic_params = 261

[query description - consider adding a doc-comment!] determining which generic parameters are unused by tcx.def_path_str(key.def_id())

backend_optimization_level = 262

[query description - consider adding a doc-comment!] optimization level used by backend

output_filenames = 263

Return the filenames where output artefacts shall be stored.

This query returns an &Arc because codegen backends need the value even after the TyCtxt has been destroyed.

normalize_canonicalized_projection_ty = 264

Do not call this query directly: Invoke normalize instead.

normalize_canonicalized_weak_ty = 265

Do not call this query directly: Invoke normalize instead.

normalize_canonicalized_inherent_projection_ty = 266

Do not call this query directly: Invoke normalize instead.

try_normalize_generic_arg_after_erasing_regions = 267

Do not call this query directly: invoke try_normalize_erasing_regions instead.

implied_outlives_bounds_compat = 268

[query description - consider adding a doc-comment!] computing implied outlives bounds for goal.value.value

implied_outlives_bounds = 269

[query description - consider adding a doc-comment!] computing implied outlives bounds v2 for goal.value.value

dropck_outlives = 270

Do not call this query directly: invoke DropckOutlives::new(dropped_ty)).fully_perform(typeck.infcx) instead.

evaluate_obligation = 271

Do not call this query directly: invoke infcx.predicate_may_hold() or infcx.predicate_must_hold() instead.

type_op_ascribe_user_type = 272

Do not call this query directly: part of the Eq type-op

type_op_prove_predicate = 273

Do not call this query directly: part of the ProvePredicate type-op

type_op_normalize_ty = 274

Do not call this query directly: part of the Normalize type-op

type_op_normalize_clause = 275

Do not call this query directly: part of the Normalize type-op

type_op_normalize_poly_fn_sig = 276

Do not call this query directly: part of the Normalize type-op

type_op_normalize_fn_sig = 277

Do not call this query directly: part of the Normalize type-op

instantiate_and_check_impossible_predicates = 278

[query description - consider adding a doc-comment!] checking impossible instantiated predicates: tcx.def_path_str(key.0)

is_impossible_associated_item = 279

[query description - consider adding a doc-comment!] checking if tcx.def_path_str(key.1) is impossible to reference within tcx.def_path_str(key.0)

method_autoderef_steps = 280

[query description - consider adding a doc-comment!] computing autoderef types for goal.value.value

supported_target_features = 281

[query description - consider adding a doc-comment!] looking up supported target features

implied_target_features = 282

[query description - consider adding a doc-comment!] looking up implied target features

features_query = 283

[query description - consider adding a doc-comment!] looking up enabled feature gates

crate_for_resolver = 284

[query description - consider adding a doc-comment!] the ast before macro expansion and name resolution

resolve_instance_raw = 285

Attempt to resolve the given DefId to an Instance, for the given generics args (GenericArgsRef), returning one of:

Ok(Some(instance)) on success

Ok(None) when the GenericArgsRef are still too generic, and therefore don’t allow finding the final Instance

Err(ErrorGuaranteed) when the Instance resolution process couldn’t complete due to errors elsewhere - this is distinct from Ok(None) to avoid misleading diagnostics when an error has already been/will be emitted, for the original cause.

reveal_opaque_types_in_bounds = 286

[query description - consider adding a doc-comment!] revealing opaque types in {:?}

limits = 287

[query description - consider adding a doc-comment!] looking up limits

diagnostic_hir_wf_check = 288

Performs an HIR-based well-formed check on the item with the given HirId. If we get an Unimplemented error that matches the provided Predicate, return the cause of the newly created obligation.

This is only used by error-reporting code to get a better cause (in particular, a better span) for an existing error. Therefore, it is best-effort, and may never handle all of the cases that the normal ty::Ty-based wfcheck does. This is fine, because the ty::Ty-based wfcheck is always run.

global_backend_features = 289

The list of backend features computed from CLI flags (-Ctarget-cpu, -Ctarget-feature, --target and similar).

check_validity_requirement = 290

[query description - consider adding a doc-comment!] checking validity requirement for key.1.value : key.0

compare_impl_const = 291

[query description - consider adding a doc-comment!] checking assoc const tcx.def_path_str(key.0) has the same type as trait item

deduced_param_attrs = 292

[query description - consider adding a doc-comment!] deducing parameter attributes for tcx.def_path_str(def_id)

doc_link_resolutions = 293

[query description - consider adding a doc-comment!] resolutions for documentation links for a module

doc_link_traits_in_scope = 294

[query description - consider adding a doc-comment!] traits in scope for documentation links for a module

stripped_cfg_items = 295

Get all item paths that were stripped by a #[cfg] in a particular crate. Should not be called for the local crate before the resolver outputs are created, as it is only fed there.

generics_require_sized_self = 296

[query description - consider adding a doc-comment!] check whether the item has a where Self: Sized bound

cross_crate_inlinable = 297

[query description - consider adding a doc-comment!] whether the item should be made inlinable across crates