Trait rustc_middle::dep_graph::dep_node::DepNodeParams

pub trait DepNodeParams<Ctxt>: Debugwhere
    Ctxt: DepContext,{
    fn fingerprint_style() -> FingerprintStyle;
    fn recover(
        tcx: Ctxt,
        dep_node: &DepNode<<Ctxt as DepContext>::DepKind>
    ) -> Option<Self>;

    fn to_fingerprint(&self, Ctxt) -> Fingerprint { ... }
    fn to_debug_str(&self, Ctxt) -> String { ... }
}

Required Methods

fn fingerprint_style() -> FingerprintStyle

fn recover(
tcx: Ctxt,
dep_node: &DepNode<<Ctxt as DepContext>::DepKind>
) -> Option<Self>

This method tries to recover the query key from the given DepNode, something which is needed when forcing DepNodes during red-green evaluation. The query system will only call this method if fingerprint_style() is not FingerprintStyle::Opaque. It is always valid to return None here, in which case incremental compilation will treat the query as having changed instead of forcing it.

Provided Methods

fn to_fingerprint(&self, Ctxt) -> Fingerprint

This method turns the parameters of a DepNodeConstructor into an opaque Fingerprint to be used in DepNode. Not all DepNodeParams support being turned into a Fingerprint (they don’t need to if the corresponding DepNode is anonymous).

fn to_debug_str(&self, Ctxt) -> String