rustc_middle

Module hir

HIR datatypes. See the rustc dev guide for more info.

Modules

map

This module used to contain a type called Map. That type has since been eliminated, and all its methods are now on TyCtxt. But the module name stays as map because there isn’t an obviously better name for it.

nested_filter

place

Structs

Hashes

Hashes computed by TyCtxt::hash_owner_nodes if necessary.

ModuleItems

Gather the LocalDefId for each item-like within a module, including items contained within bodies. The Ids are in visitor order. This is used to partition a pass between modules.

Functions

provide