Module rustc_middle::lint

Structs

LintExpectation

This struct represents a lint expectation and holds all required information to emit the unfulfilled_lint_expectations lint if it is unfulfilled after the LateLintPass has completed.

ShallowLintLevelMap

Return type for the shallow_lint_levels_on query.

Enums

LintLevelSource

How a lint level was set.

Functions

explain_lint_level_source

in_external_macro

Returns whether span originates in a foreign crate’s external macro.

reveal_actual_level

From an initial level and source, verify the effect of special annotations: warnings lint level and lint caps.

struct_lint_level

Type Definitions

LevelAndSource

A tuple of a lint level and its source.