rustc_middle::lint
Function reveal_actual_level

pub fn reveal_actual_level(
    level: Option<Level>,
    src: &mut LintLevelSource,
    sess: &Session,
    lint: LintId,
    probe_for_lint_level: impl FnOnce(LintId) -> (Option<Level>, LintLevelSource),
) -> Level

From an initial level and source, verify the effect of special annotations: warnings lint level and lint caps.

The return of this function is suitable for diagnostics.