Struct rustc_middle::lint::LintExpectation

pub struct LintExpectation {
    pub reason: Option<Symbol>,
    pub emission_span: Span,
    pub is_unfulfilled_lint_expectations: bool,
    pub lint_tool: Option<Symbol>,
}

This struct represents a lint expectation and holds all required information to emit the unfulfilled_lint_expectations lint if it is unfulfilled after the LateLintPass has completed.

Fields
reason: Option<Symbol>

The reason for this expectation that can optionally be added as part of the attribute. It will be displayed as part of the lint message.

emission_span: Span

The Span of the attribute that this expectation originated from.

is_unfulfilled_lint_expectations: bool

Lint messages for the unfulfilled_lint_expectations lint will be adjusted to include an additional note. Therefore, we have to track if the expectation is for the lint.

lint_tool: Option<Symbol>

This will hold the name of the tool that this lint belongs to. For the lint clippy::some_lint the tool would be clippy, the same goes for rustdoc. This will be None for rustc lints

Implementations

impl LintExpectation

pub fn new(
reason: Option<Symbol>,
emission_span: Span,
is_unfulfilled_lint_expectations: bool,
lint_tool: Option<Symbol>
) -> Self