Module rustc_middle::middle

Modules

codegen_fn_attrs

dependency_format

Type definitions for learning about the dependency formats of all upstream crates (rlibs/dylibs/oh my).

exported_symbols

lang_items

Detecting language items.

lib_features

limits

Registering limits:

privacy

A pass that checks to make sure private fields and methods aren’t used outside their scopes. This pass will also generate a set of exported items which are available for use externally when compiled as a library.

region

This file declares the ScopeTree type, which describes the parent links in the region hierarchy.

resolve_lifetime

Name resolution for lifetimes: type declarations.

stability

A pass that annotates every item and method with its stability level, propagating default levels lexically from parent to children ast nodes.

Functions

provide