rustc_middle::middle
Module exported_symbols

Structs

SymbolExportInfo

The SymbolExportInfo of a symbols specifies symbol-related information that is relevant to code generation and linking.

Enums

ExportedSymbol

SymbolExportKind

Kind of exported symbols.

SymbolExportLevel

The SymbolExportLevel of a symbols specifies from which kinds of crates the symbol will be exported. C symbols will be exported from any kind of crate, including cdylibs which export very few things. Rust will only be exported if the crate produced is a Rust dylib.

Functions

metadata_symbol_name