Module rustc_middle::middle::limits

Registering limits:

recursion_limit,

move_size_limit, and

type_length_limit

There are various parts of the compiler that must impose arbitrary limits on how deeply they recurse to prevent stack overflow. Users can override this via an attribute on the crate like #![recursion_limit="22"]. This pass just peeks and looks for that attribute.

Functions

get_limit 🔒

get_limit_size

get_recursion_limit

provide