Module rustc_middle::middle::stability

A pass that annotates every item and method with its stability level, propagating default levels lexically from parent to children ast nodes.

Re-exports

pub use self::StabilityLevel::*;

Structs

DeprecationEntry

An entry in the depr_map.

Index

A stability index, giving the stability level for items and methods.

Enums

AllowUnstable

An override option for eval_stability.

EvalResult

Result of TyCtxt::eval_stability.

StabilityLevel

Functions

deprecation_in_effect

Checks whether an item marked with deprecated(since="X") is currently deprecated (i.e., whether X is not greater than the current rustc version).

deprecation_lint 🔒

deprecation_message 🔒

deprecation_message_and_lint

deprecation_suggestion

early_report_deprecation

late_report_deprecation 🔒

report_unstable

skip_stability_check_due_to_privacy 🔒

suggestion_for_allocator_api 🔒