Struct rustc_middle::mir::basic_blocks::BasicBlocks

pub struct BasicBlocks<'tcx> {
    basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>,
    cache: Cache,
}

Fields
basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>cache: Cache
Implementations

impl<'tcx> BasicBlocks<'tcx>

pub fn new(basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>) -> Self

pub fn is_cfg_cyclic(&self) -> bool

Returns true if control-flow graph contains a cycle reachable from the START_BLOCK.

pub fn dominators(&self) -> &Dominators<BasicBlock>

pub fn predecessors(&self) -> &IndexVec<BasicBlock, SmallVec<[BasicBlock; 4]>>

Returns predecessors for each basic block.

pub fn reverse_postorder(&self) -> &[BasicBlock]

Returns basic blocks in a reverse postorder.

See traversal::reverse_postorder’s docs to learn what is preorder traversal.

pub fn switch_sources( &self, ) -> &FxHashMap<(BasicBlock, BasicBlock), SmallVec<[Option<u128>; 1]>>

switch_sources()[&(target, switch)] returns a list of switch values that lead to a target block from a switch block.

pub fn as_mut(&mut self) -> &mut IndexVec<BasicBlock, BasicBlockData<'tcx>>

Returns mutable reference to basic blocks. Invalidates CFG cache.

pub fn as_mut_preserves_cfg( &mut self, ) -> &mut IndexVec<BasicBlock, BasicBlockData<'tcx>>

Get mutable access to basic blocks without invalidating the CFG cache.

By calling this method instead of e.g. BasicBlocks::as_mut you promise not to change the CFG. This means that

The number of basic blocks remains unchanged

The set of successors of each terminator remains unchanged.

For each TerminatorKind::SwitchInt, the targets remains the same and the terminator kind is not changed.

If any of these conditions cannot be upheld, you should call BasicBlocks::invalidate_cfg_cache.

pub fn invalidate_cfg_cache(&mut self)

Invalidates cached information about the CFG.

You will only ever need this if you have also called BasicBlocks::as_mut_preserves_cfg. All other methods that allow you to mutate the basic blocks also call this method themselves, thereby avoiding any risk of accidentally cache invalidation.