Enum rustc_middle::mir::AssertKind

pub enum AssertKind<O> {
    BoundsCheck {
        len: O,
        index: O,
    },
    Overflow(BinOp, O, O),
    OverflowNeg(O),
    DivisionByZero(O),
    RemainderByZero(O),
    ResumedAfterReturn(CoroutineKind),
    ResumedAfterPanic(CoroutineKind),
    MisalignedPointerDereference {
        required: O,
        found: O,
    },
}

Information about an assertion failure.

Variants

BoundsCheck

Fields

len: O

index: O

Overflow(BinOp, O, O)

OverflowNeg(O)

DivisionByZero(O)

RemainderByZero(O)

ResumedAfterReturn(CoroutineKind)

ResumedAfterPanic(CoroutineKind)

MisalignedPointerDereference

Fields

required: O

found: O

Implementations

impl<O> AssertKind<O>

pub fn is_optional_overflow_check(&self) -> bool

Returns true if this an overflow checking assertion controlled by -C overflow-checks.

pub fn panic_function(&self) -> LangItem

Get the lang item that is invoked to print a static message when this assert fires.

The caller is expected to handle BoundsCheck and MisalignedPointerDereference by invoking the appropriate lang item (panic_bounds_check/panic_misaligned_pointer_dereference) instead of printing a static message. Those have dynamic arguments that aren’t present for the rest of the messages here.

pub fn fmt_assert_args<W: Write>(&self, f: &mut W) -> Result
where O: Debug,

Format the message arguments for the assert(cond, msg..) terminator in MIR printing.

Needs to be kept in sync with the run-time behavior (which is defined by AssertKind::panic_function and the lang items mentioned in its docs). Note that we deliberately show more details here than we do at runtime, such as the actual numbers that overflowed – it is much easier to do so here than at runtime.

pub fn diagnostic_message(&self) -> DiagMessage

Format the diagnostic message for use in a lint (e.g. when the assertion fails during const-eval).

Needs to be kept in sync with the run-time behavior (which is defined by AssertKind::panic_function and the lang items mentioned in its docs). Note that we deliberately show more details here than we do at runtime, such as the actual numbers that overflowed – it is much easier to do so here than at runtime.

pub fn add_args(self, adder: &mut dyn FnMut(DiagArgName, DiagArgValue))
where O: Debug,