Function rustc_middle::mir::graphviz::write_graph_label

fn write_graph_label<'tcx, W: Write>(
    tcx: TyCtxt<'tcx>,
    body: &Body<'_>,
    w: &mut W,
) -> Result

Write the graphviz DOT label for the overall graph. This is essentially a block of text that will appear below the graph, showing the type of the fn this MIR represents and the types of all the variables and temporaries.