Module rustc_middle::mir::interpret::allocation::init_mask

Structs

InitChunkIter

Yields InitChunks. See InitMask::range_as_init_chunks.

InitCopy

Run-length encoding of the uninit mask. Used to copy parts of a mask multiple times to another allocation.

InitMask

A bitmask where each bit refers to the byte with the same index. If the bit is true, the byte is initialized. If it is false the byte is uninitialized.

Enums

InitChunk

A contiguous chunk of initialized or uninitialized memory.

Type Definitions

Block 🔒