Struct rustc_middle::mir::interpret::allocation::init_mask::InitMask

pub struct InitMask {
    blocks: InitMaskBlocks,
    len: Size,
}

A bitmask where each bit refers to the byte with the same index. If the bit is true, the byte is initialized. If it is false the byte is uninitialized. The actual bits are only materialized when needed, and we try to keep this data lazy as long as possible. Currently, if all the blocks have the same value, then the mask represents either a fully initialized or fully uninitialized const allocation, so we can only store that single value.

Fields
blocks: InitMaskBlockslen: Size
Implementations

impl InitMask

pub fn new(size: Size, state: bool) -> Self

pub fn is_range_initialized(&self, range: AllocRange) -> Result<(), AllocRange>

Checks whether the range is entirely initialized.

Returns Ok(()) if it’s initialized. Otherwise returns a range of byte indexes for the first contiguous span of the uninitialized access.

pub fn set_range(&mut self, range: AllocRange, new_state: bool)

Sets a specified range to a value. If the range is out-of-bounds, the mask will grow to accommodate it entirely.

fn materialize_blocks(&mut self) -> &mut InitMaskMaterialized

Materializes this mask’s blocks when the mask is lazy.

pub fn get(&self, idx: Size) -> bool

Returns the initialization state at the specified in-bounds index.

impl InitMask

pub fn range_as_init_chunks(&self, range: AllocRange) -> InitChunkIter<'_>

Returns an iterator, yielding a range of byte indexes for each contiguous region of initialized or uninitialized bytes inside the range start..end (end-exclusive).

The iterator guarantees the following:

Chunks are nonempty.

Chunks are adjacent (each range’s start is equal to the previous range’s end).

Chunks span exactly start..end (the first starts at start, the last ends at end).

Chunks alternate between InitChunk::Init and InitChunk::Uninit.

impl InitMask

Transferring the initialization mask to other allocations.

pub fn prepare_copy(&self, range: AllocRange) -> InitCopy

Creates a run-length encoding of the initialization mask; panics if range is empty.

This is essentially a more space-efficient version of InitMask::range_as_init_chunks(...).collect::<Vec<_>>().

pub fn apply_copy(&mut self, defined: InitCopy, range: AllocRange, repeat: u64)

Applies multiple instances of the run-length encoding to the initialization mask.