rustc_middle::mir::interpret::allocation

Module provenance_map

Store the provenance for each byte in the range, with a more efficient representation for the common case where PTR_SIZE consecutive bytes have the same provenance.

Structs

ProvenanceCopy

A partial, owned list of provenance to transfer into another allocation.

ProvenanceMap

Stores the provenance information of pointers stored in memory.