Enum rustc_middle::mir::interpret::Scalar

pub enum Scalar<Prov = AllocId> {
    Int(ScalarInt),
    Ptr(Pointer<Prov>, u8),
}

A Scalar represents an immediate, primitive value existing outside of a memory::Allocation. It is in many ways like a small chunk of an Allocation, up to 16 bytes in size. Like a range of bytes in an Allocation, a Scalar can either represent the raw bytes of a simple value or a pointer into another Allocation

These variants would be private if there was a convenient way to achieve that in Rust. Do not match on a Scalar! Use the various to_* methods instead.

Variants

Int(ScalarInt)

The raw bytes of a simple value.

Ptr(Pointer<Prov>, u8)

A pointer.

We also store the size of the pointer, such that a Scalar always knows how big it is. The size is always the pointer size of the current target, but this is not information that we always have readily available.

Implementations

impl<Prov> Scalar<Prov>

pub fn from_pointer(ptr: Pointer<Prov>, cx: &impl HasDataLayout) -> Self

pub fn from_maybe_pointer(
ptr: Pointer<Option<Prov>>,
cx: &impl HasDataLayout
) -> Self

Create a Scalar from a pointer with an Option<_> provenance (where None represents a plain integer / “invalid” pointer).

pub fn null_ptr(cx: &impl HasDataLayout) -> Self

pub fn from_bool(b: bool) -> Self

pub fn from_char(c: char) -> Self

pub fn try_from_uint(i: impl Into<u128>, size: Size) -> Option<Self>

pub fn from_uint(i: impl Into<u128>, size: Size) -> Self

pub fn from_u8(i: u8) -> Self

pub fn from_u16(i: u16) -> Self

pub fn from_u32(i: u32) -> Self

pub fn from_u64(i: u64) -> Self

pub fn from_machine_usize(i: u64, cx: &impl HasDataLayout) -> Self

pub fn try_from_int(i: impl Into<i128>, size: Size) -> Option<Self>

pub fn from_int(i: impl Into<i128>, size: Size) -> Self

pub fn from_i32(i: i32) -> Self

pub fn from_i64(i: i64) -> Self

pub fn from_machine_isize(i: i64, cx: &impl HasDataLayout) -> Self

pub fn from_f32(f: Single) -> Self

pub fn from_f64(f: Double) -> Self

pub fn to_bits_or_ptr_internal(
self,
target_size: Size
) -> Result<Result<u128, Pointer<Prov>>, ScalarSizeMismatch>

This is almost certainly not the method you want! You should dispatch on the type and use to_{u8,u16,...}/scalar_to_ptr to perform ptr-to-int / int-to-ptr casts as needed.

This method only exists for the benefit of low-level operations that truly need to treat the scalar in whatever form it is.

This throws UB (instead of ICEing) on a size mismatch since size mismatches can arise in Miri when someone declares a function that we shim (such as malloc) with a wrong type.

impl<'tcx, Prov: Provenance> Scalar<Prov>

pub fn to_pointer(
self,
cx: &impl HasDataLayout
) -> InterpResult<'tcx, Pointer<Option<Prov>>>

pub fn try_to_int(self) -> Result<ScalarInt, Scalar<AllocId>>

Fundamental scalar-to-int (cast) operation. Many convenience wrappers exist below, that you likely want to use instead.

Will perform ptr-to-int casts if needed and possible. If that fails, we know the offset is relative, so we return an “erased” Scalar (which is useful for error messages but not much else).

pub fn assert_int(self) -> ScalarInt

pub fn to_bits(self, target_size: Size) -> InterpResult<'tcx, u128>

This throws UB (instead of ICEing) on a size mismatch since size mismatches can arise in Miri when someone declares a function that we shim (such as malloc) with a wrong type.

pub fn assert_bits(self, target_size: Size) -> u128

pub fn to_bool(self) -> InterpResult<'tcx, bool>

pub fn to_char(self) -> InterpResult<'tcx, char>

pub fn to_uint(self, size: Size) -> InterpResult<'tcx, u128>

Converts the scalar to produce an unsigned integer of the given size. Fails if the scalar is a pointer.

pub fn to_u8(self) -> InterpResult<'tcx, u8>

Converts the scalar to produce a u8. Fails if the scalar is a pointer.

pub fn to_u16(self) -> InterpResult<'tcx, u16>

Converts the scalar to produce a u16. Fails if the scalar is a pointer.

pub fn to_u32(self) -> InterpResult<'tcx, u32>

Converts the scalar to produce a u32. Fails if the scalar is a pointer.

pub fn to_u64(self) -> InterpResult<'tcx, u64>

Converts the scalar to produce a u64. Fails if the scalar is a pointer.

pub fn to_u128(self) -> InterpResult<'tcx, u128>

Converts the scalar to produce a u128. Fails if the scalar is a pointer.

pub fn to_machine_usize(self, cx: &impl HasDataLayout) -> InterpResult<'tcx, u64>

Converts the scalar to produce a machine-pointer-sized unsigned integer. Fails if the scalar is a pointer.

pub fn to_int(self, size: Size) -> InterpResult<'tcx, i128>

Converts the scalar to produce a signed integer of the given size. Fails if the scalar is a pointer.

pub fn to_i8(self) -> InterpResult<'tcx, i8>

Converts the scalar to produce an i8. Fails if the scalar is a pointer.

pub fn to_i16(self) -> InterpResult<'tcx, i16>

Converts the scalar to produce an i16. Fails if the scalar is a pointer.

pub fn to_i32(self) -> InterpResult<'tcx, i32>

Converts the scalar to produce an i32. Fails if the scalar is a pointer.

pub fn to_i64(self) -> InterpResult<'tcx, i64>

Converts the scalar to produce an i64. Fails if the scalar is a pointer.

pub fn to_i128(self) -> InterpResult<'tcx, i128>

Converts the scalar to produce an i128. Fails if the scalar is a pointer.

pub fn to_machine_isize(self, cx: &impl HasDataLayout) -> InterpResult<'tcx, i64>

Converts the scalar to produce a machine-pointer-sized signed integer. Fails if the scalar is a pointer.

pub fn to_f32(self) -> InterpResult<'tcx, Single>

pub fn to_f64(self) -> InterpResult<'tcx, Double>