Struct rustc_middle::mir::interpret::ConstAllocation

pub struct ConstAllocation<'tcx, Prov = AllocId, Extra = ()>(pub Interned<'tcx, Allocation<Prov, Extra>>);

Interned types generally have an Outer type and an Inner type, where Outer is a newtype around Interned<Inner>, and all the operations are done on Outer, because all occurrences are interned. E.g. Ty is an outer type and TyS is its inner type.

Here things are different because only const allocations are interned. This means that both the inner type (Allocation) and the outer type (ConstAllocation) are used quite a bit.

Tuple Fields
0: Interned<'tcx, Allocation<Prov, Extra>>
Implementations

impl<'tcx, Prov, Extra> ConstAllocation<'tcx, Prov, Extra>

pub fn inner(self) -> &'tcx Allocation<Prov, Extra>