Struct rustc_middle::mir::interpret::Pointer

pub struct Pointer<Prov = AllocId> {
    pub(super) offset: Size,
    pub provenance: Prov,
}

Represents a pointer in the Miri engine.

Pointers are “tagged” with provenance information; typically the AllocId they belong to.

Fields
offset: Sizeprovenance: Prov
Implementations

impl<Prov> Pointer<Option<Prov>>

pub fn into_pointer_or_addr(self) -> Result<Pointer<Prov>, Size>

Convert this pointer that might have a provenance into a pointer that definitely has a provenance, or an absolute address.

This is rarely what you want; call ptr_try_get_alloc_id instead.

pub fn addr(self) -> Sizewhere Prov: Provenance,

Returns the absolute address the pointer points to. Only works if Prov::OFFSET_IS_ADDR is true!

impl<Prov> Pointer<Option<Prov>>

pub fn from_addr_invalid(addr: u64) -> Self

Creates a pointer to the given address, with invalid provenance (i.e., cannot be used for any memory access).

pub fn null() -> Self

impl<'tcx, Prov> Pointer<Prov>

pub fn new(provenance: Prov, offset: Size) -> Self

pub fn into_parts(self) -> (Prov, Size)

Obtain the constituents of this pointer. Not that the meaning of the offset depends on the type Prov! This function must only be used in the implementation of Machine::ptr_get_alloc, and when a Pointer is taken apart to be stored efficiently in an Allocation.

pub fn map_provenance(self, f: impl FnOnce(Prov) -> Prov) -> Self

pub fn offset( self, i: Size, cx: &impl HasDataLayout ) -> InterpResult<'tcx, Self>

pub fn overflowing_offset( self, i: Size, cx: &impl HasDataLayout ) -> (Self, bool)

pub fn wrapping_offset(self, i: Size, cx: &impl HasDataLayout) -> Self

pub fn signed_offset( self, i: i64, cx: &impl HasDataLayout ) -> InterpResult<'tcx, Self>

pub fn overflowing_signed_offset( self, i: i64, cx: &impl HasDataLayout ) -> (Self, bool)

pub fn wrapping_signed_offset(self, i: i64, cx: &impl HasDataLayout) -> Self