Module rustc_middle::mir::interpret::value

Structs

ConstAlloc

Represents the result of const evaluation via the eval_to_allocation query.

Enums

ConstValue

Represents a constant value in Rust. Scalar and Slice are optimizations for array length computations, enum discriminants and the pattern matching logic.

Scalar

A Scalar represents an immediate, primitive value existing outside of a memory::Allocation. It is in many ways like a small chunk of an Allocation, up to 16 bytes in size. Like a range of bytes in an Allocation, a Scalar can either represent the raw bytes of a simple value or a pointer into another Allocation

Functions

get_slice_bytes

Gets the bytes of a constant slice value.