Struct rustc_middle::mir::mono::CodegenUnit

pub struct CodegenUnit<'tcx> {
    name: Symbol,
    items: FxHashMap<MonoItem<'tcx>, (Linkage, Visibility)>,
    size_estimate: Option<usize>,
    primary: bool,
    is_code_coverage_dead_code_cgu: bool,
}

Fields
name: Symbol

A name for this CGU. Incremental compilation requires that name be unique amongst all crates. Therefore, it should contain something unique to this crate (e.g., a module path) as well as the crate name and disambiguator.

items: FxHashMap<MonoItem<'tcx>, (Linkage, Visibility)>size_estimate: Option<usize>primary: boolis_code_coverage_dead_code_cgu: bool

True if this is CGU is used to hold code coverage information for dead code, false otherwise.

Implementations

impl<'tcx> CodegenUnit<'tcx>

pub fn new(name: Symbol) -> CodegenUnit<'tcx>

pub fn name(&self) -> Symbol

pub fn set_name(&mut self, name: Symbol)

pub fn is_primary(&self) -> bool

pub fn make_primary(&mut self)

pub fn items(&self) -> &FxHashMap<MonoItem<'tcx>, (Linkage, Visibility)>

pub fn items_mut(
&mut self
) -> &mut FxHashMap<MonoItem<'tcx>, (Linkage, Visibility)>

pub fn is_code_coverage_dead_code_cgu(&self) -> bool

pub fn make_code_coverage_dead_code_cgu(&mut self)

Marks this CGU as the one used to contain code coverage information for dead code.

pub fn mangle_name(human_readable_name: &str) -> String

pub fn estimate_size(&mut self, tcx: TyCtxt<'tcx>)

pub fn size_estimate(&self) -> usize

pub fn modify_size_estimate(&mut self, delta: usize)

pub fn contains_item(&self, item: &MonoItem<'tcx>) -> bool

pub fn work_product_id(&self) -> WorkProductId

pub fn previous_work_product(&self, tcx: TyCtxt<'_>) -> WorkProduct

pub fn items_in_deterministic_order(
&self,
tcx: TyCtxt<'tcx>
) -> Vec<(MonoItem<'tcx>, (Linkage, Visibility))>

pub fn codegen_dep_node(&self, tcx: TyCtxt<'tcx>) -> DepNode