Module rustc_middle::mir::spanview

Structs

SpanViewable

Metadata to highlight the span of a MIR BasicBlock, Statement, or Terminator.

Constants

ANNOTATION_LEFT_BRACKET ğŸ”’

ANNOTATION_RIGHT_BRACKET ğŸ”’

CARET ğŸ”’

FOOTER ğŸ”’

HEADER ğŸ”’

NEW_LINE_SPAN ğŸ”’

START_BODY ğŸ”’

STYLE_SECTION ğŸ”’

TOOLTIP_INDENT

Functions

block_span_viewable ğŸ”’

compute_block_span ğŸ”’

escape_attr ğŸ”’

escape_html ğŸ”’

fn_span ğŸ”’

hir_body ğŸ”’

make_html_snippet ğŸ”’

source_range_no_file

Format a string showing the start line and column, and end line and column within a file.

statement_kind_name

statement_span_viewable ğŸ”’

terminator_kind_name

terminator_span_viewable ğŸ”’

tooltip ğŸ”’

trim_span ğŸ”’

trim_span_hi ğŸ”’

trim_span_lo ğŸ”’

write_coverage_gap ğŸ”’

write_document

Generate a spanview HTML+CSS document for the given local function def_id, and a pre-generated list SpanViewables.

write_mir_fn_spanview

Write a spanview HTML+CSS file to analyze MIR element spans.

write_next_viewable_with_overlaps ğŸ”’

Recursively process each ordered span. Spans that overlap will have progressively varying styles, such as increased padding for each overlap. Non-overlapping adjacent spans will have alternating style choices, to help distinguish between them if, visually adjacent. The layer is incremented for each overlap, and the alt bool alternates between true and false, for each adjacent non-overlapping span. Source code between the spans (code that is not in any coverage region) has neutral styling.

write_span ğŸ”’