rustc_middle::mir

Module statement

Functionality for statements, operands, places, and things that appear in them.

Structs

PlaceRef

PlaceTy

Statement

A statement in a basic block, including information about its source code.

Enums

RvalueInitializationState

Rvalues

Type Aliases

ProjectionKind

Alias for projections as they appear in UserTypeProjection, where we need neither the V parameter for Index nor the T for Field.