rustc_middle::mir

Module traversal

Structs

MonoReachable

Postorder

Postorder traversal of a graph.

Preorder

Preorder traversal of a graph.

Functions

mono_reachable

Traversal of a Body that tries to avoid unreachable blocks in a monomorphized Instance.

mono_reachable_as_bitset

MonoReachable internally accumulates a DenseBitSet of visited blocks. This is just a convenience function to run that traversal then extract its set of reached blocks.

mono_reachable_reverse_postorder

postorder

Postorder traversal of a graph.

preorder

Preorder traversal of a graph.

reachable

Returns an iterator over all basic blocks reachable from the START_BLOCK in no particular order.

reachable_as_bitset

Returns a DenseBitSet containing all basic blocks reachable from the START_BLOCK.

reverse_postorder

Reverse postorder traversal of a graph.