Function rustc_middle::mir::traversal::reachable

pub fn reachable<'a, 'tcx>(
    body: &'a Body<'tcx>
) -> impl 'a + Iterator<Item = (BasicBlock, &'a BasicBlockData<'tcx>)>

Returns an iterator over all basic blocks reachable from the START_BLOCK in no particular order.

This is clearer than writing preorder in cases where the order doesn’t matter.