Type Alias rustc_middle::query::queries::associated_items::ProvidedValue

pub type ProvidedValue<'tcx> = AssocItems;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Implementations

impl AssocItems

pub fn new(items_in_def_order: impl IntoIterator<Item = AssocItem>) -> Self

Constructs an AssociatedItems map from a series of ty::AssocItems in definition order.

pub fn in_definition_order(&self) -> impl '_ + Iterator<Item = &AssocItem>

Returns a slice of associated items in the order they were defined.

New code should avoid relying on definition order. If you need a particular associated item for a known trait, make that trait a lang item instead of indexing this array.

pub fn len(&self) -> usize

pub fn filter_by_name_unhygienic( &self, name: Symbol ) -> impl '_ + Iterator<Item = &AssocItem>

Returns an iterator over all associated items with the given name, ignoring hygiene.

pub fn find_by_name_and_kind( &self, tcx: TyCtxt<'_>, ident: Ident, kind: AssocKind, parent_def_id: DefId ) -> Option<&AssocItem>

Returns the associated item with the given name and AssocKind, if one exists.

pub fn find_by_name_and_kinds( &self, tcx: TyCtxt<'_>, ident: Ident, kinds: &[AssocKind], parent_def_id: DefId ) -> Option<&AssocItem>

Returns the associated item with the given name and any of AssocKind, if one exists.

pub fn find_by_name_and_namespace( &self, tcx: TyCtxt<'_>, ident: Ident, ns: Namespace, parent_def_id: DefId ) -> Option<&AssocItem>

Returns the associated item with the given name in the given Namespace, if one exists.