Type Alias rustc_middle::query::queries::closure_saved_names_of_captured_variables::ProvidedValue

pub type ProvidedValue<'tcx> = IndexVec<FieldIdx, Symbol>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.