Function rustc_middle::query::queries::collect_return_position_impl_trait_in_trait_tys::provided_to_erased

pub fn provided_to_erased<'tcx>(
    _tcx: TyCtxt<'tcx>,
    value: ProvidedValue<'tcx>,
) -> Erase<Value<'tcx>>

This function takes ProvidedValue and coverts it to an erased Value by allocating it on an arena if the query has the arena_cache modifier. The value is then erased and returned. This will happen when computing the query using a provider or decoding a stored result.