Type Alias rustc_middle::query::queries::const_caller_location::Value

pub type Value<'tcx> = ConstValue<'tcx>;

Aliased Type

enum Value<'tcx> {
    Scalar(Scalar<AllocId>),
    ZeroSized,
    Slice {
        data: ConstAllocation<'tcx>,
        meta: u64,
    },
    Indirect {
        alloc_id: AllocId,
        offset: Size,
    },
}

Variants

Scalar(Scalar<AllocId>)

Used for types with layout::abi::Scalar ABI.

Not using the enum Value to encode that this must not be Uninit.

ZeroSized

Only for ZSTs.

Slice

Fields

data: ConstAllocation<'tcx>

The allocation storing the slice contents. This always points to the beginning of the allocation.

meta: u64

The metadata field of the reference. This is a “target usize”, so we use u64 as in the interpreter.

Used for references to unsized types with slice tail.

This is worth an optimized representation since Rust has literals of type &str and &[u8]. Not having to indirect those through an AllocId (or two, if we used Indirect) has shown measurable performance improvements on stress tests. We then reuse this optimization for slice-tail types more generally during valtree-to-constval conversion.

Indirect

Fields

alloc_id: AllocId

The backing memory of the value. May contain more memory than needed for just the value if this points into some other larger ConstValue.

We use an AllocId here instead of a ConstAllocation<'tcx> to make sure that when a raw constant (which is basically just an AllocId) is turned into a ConstValue and back, we can preserve the original AllocId.

offset: Size

Offset into alloc

A value not representable by the other variants; needs to be stored in-memory.

Must not be used for scalars or ZST, but having &str or other slices in this variant is fine.

Implementations

impl<'tcx> ConstValue<'tcx>

pub fn try_to_scalar(&self) -> Option<Scalar<AllocId>>

pub fn try_to_scalar_int(&self) -> Option<ScalarInt>

pub fn try_to_bits(&self, size: Size) -> Option<u128>

pub fn try_to_bool(&self) -> Option<bool>

pub fn try_to_target_usize(&self, tcx: TyCtxt<'tcx>) -> Option<u64>

pub fn try_to_bits_for_ty( &self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, ty: Ty<'tcx> ) -> Option<u128>

pub fn from_bool(b: bool) -> Self

pub fn from_u64(i: u64) -> Self

pub fn from_u128(i: u128) -> Self

pub fn from_target_usize(i: u64, cx: &impl HasDataLayout) -> Self

pub fn try_get_slice_bytes_for_diagnostics( &self, tcx: TyCtxt<'tcx> ) -> Option<&'tcx [u8]>

Must only be called on constants of type &str or &[u8]!