Type Alias rustc_middle::query::queries::def_span::ProvidedValue

pub type ProvidedValue<'tcx> = Span;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx> {
    lo_or_index: u32,
    len_with_tag_or_marker: u16,
    ctxt_or_parent_or_marker: u16,
}

Fields
lo_or_index: u32len_with_tag_or_marker: u16ctxt_or_parent_or_marker: u16