Type Alias rustc_middle::query::queries::explicit_predicates_of::ProvidedValue

pub type ProvidedValue<'tcx> = GenericPredicates<'tcx>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx> {
    pub parent: Option<DefId>,
    pub predicates: &'tcx [(Clause<'tcx>, Span)],
    pub effects_min_tys: &'tcx RawList<(), Ty<'tcx>>,
}

Fields
parent: Option<DefId>predicates: &'tcx [(Clause<'tcx>, Span)]effects_min_tys: &'tcx RawList<(), Ty<'tcx>>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 32 bytes