Type Alias rustc_middle::query::queries::expn_that_defined::ProvidedValue

pub type ProvidedValue<'tcx> = ExpnId;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx> {
    pub krate: CrateNum,
    pub local_id: ExpnIndex,
}

Fields
krate: CrateNumlocal_id: ExpnIndex