Type Alias rustc_middle::query::queries::generics_of::ProvidedValue

pub type ProvidedValue<'tcx> = Generics;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Implementations

impl<'tcx> Generics

pub fn param_def_id_to_index( &self, tcx: TyCtxt<'tcx>, def_id: DefId ) -> Option<u32>

Looks through the generics and all parents to find the index of the given param def-id. This is in comparison to the param_def_id_to_index struct member, which only stores information about this item’s own generics.

pub fn count(&self) -> usize

pub fn own_counts(&self) -> GenericParamCount

pub fn own_defaults(&self) -> GenericParamCount

pub fn requires_monomorphization(&self, tcx: TyCtxt<'tcx>) -> bool

pub fn own_requires_monomorphization(&self) -> bool

pub fn param_at( &'tcx self, param_index: usize, tcx: TyCtxt<'tcx> ) -> &'tcx GenericParamDef

Returns the GenericParamDef with the given index.

pub fn params_to( &'tcx self, param_index: usize, tcx: TyCtxt<'tcx> ) -> &'tcx [GenericParamDef]

pub fn region_param( &'tcx self, param: &EarlyBoundRegion, tcx: TyCtxt<'tcx> ) -> &'tcx GenericParamDef

Returns the GenericParamDef associated with this EarlyBoundRegion.

pub fn type_param( &'tcx self, param: &ParamTy, tcx: TyCtxt<'tcx> ) -> &'tcx GenericParamDef

Returns the GenericParamDef associated with this ParamTy.

pub fn const_param( &'tcx self, param: &ParamConst, tcx: TyCtxt<'tcx> ) -> &GenericParamDef

Returns the GenericParamDef associated with this ParamConst.

pub fn has_impl_trait(&'tcx self) -> bool

Returns true if params has impl Trait.

pub fn own_args_no_defaults( &'tcx self, tcx: TyCtxt<'tcx>, args: &'tcx [GenericArg<'tcx>] ) -> &'tcx [GenericArg<'tcx>]

Returns the args corresponding to the generic parameters of this item, excluding Self.

This should only be used for diagnostics purposes.

pub fn own_args( &'tcx self, args: &'tcx [GenericArg<'tcx>] ) -> &'tcx [GenericArg<'tcx>]

Returns the args corresponding to the generic parameters of this item, excluding Self.

This should only be used for diagnostics purposes.