Type Alias rustc_middle::query::queries::hir_module_items::ProvidedValue

pub type ProvidedValue<'tcx> = ModuleItems;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Implementations

impl ModuleItems

pub fn items(&self) -> impl Iterator<Item = ItemId> + '_

pub fn trait_items(&self) -> impl Iterator<Item = TraitItemId> + '_

pub fn impl_items(&self) -> impl Iterator<Item = ImplItemId> + '_

pub fn foreign_items(&self) -> impl Iterator<Item = ForeignItemId> + '_

pub fn owners(&self) -> impl Iterator<Item = OwnerId> + '_

pub fn definitions(&self) -> impl Iterator<Item = LocalDefId> + '_

pub fn par_items(&self, f: impl Fn(ItemId) + DynSend + DynSync)

pub fn par_trait_items(&self, f: impl Fn(TraitItemId) + DynSend + DynSync)

pub fn par_impl_items(&self, f: impl Fn(ImplItemId) + DynSend + DynSync)

pub fn par_foreign_items(&self, f: impl Fn(ForeignItemId) + DynSend + DynSync)