Type Alias rustc_middle::query::queries::impl_polarity::ProvidedValue

pub type ProvidedValue<'tcx> = ImplPolarity;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

enum ProvidedValue<'tcx> {
    Positive,
    Negative,
    Reservation,
}

Variants

Positive

impl Trait for Type

Negative

impl !Trait for Type

Reservation

#[rustc_reservation_impl] impl Trait for Type

This is a “stability hack”, not a real Rust feature. See #64631 for details.

Implementations

impl ImplPolarity

pub fn flip(&self) -> Option<ImplPolarity>

Flips polarity by turning Positive into Negative and Negative into Positive.