Type Alias rustc_middle::query::queries::inhabited_predicate_adt::ProvidedValue

pub type ProvidedValue<'tcx> = InhabitedPredicate<'tcx>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

enum ProvidedValue<'tcx> {
    True,
    False,
    ConstIsZero(Const<'tcx>),
    NotInModule(DefId),
    GenericType(Ty<'tcx>),
    And(&'tcx [InhabitedPredicate<'tcx>; 2]),
    Or(&'tcx [InhabitedPredicate<'tcx>; 2]),
}

Variants

True

Inhabited

False

Uninhabited

ConstIsZero(Const<'tcx>)

Uninhabited when a const value is non-zero. This occurs when there is an array of uninhabited items, but the array is inhabited if it is empty.

NotInModule(DefId)

Uninhabited if within a certain module. This occurs when an uninhabited type has restricted visibility.

GenericType(Ty<'tcx>)

Inhabited if some generic type is inhabited. These are replaced by calling Self::instantiate.

And(&'tcx [InhabitedPredicate<'tcx>; 2])

A AND B

Or(&'tcx [InhabitedPredicate<'tcx>; 2])

A OR B

Implementations

impl<'tcx> InhabitedPredicate<'tcx>

pub fn apply( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, module_def_id: DefId ) -> bool

Returns true if the corresponding type is inhabited in the given ParamEnv and module

pub fn apply_any_module( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> Option<bool>

Same as apply, but returns None if self contains a module predicate

pub fn apply_ignore_module( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> bool

Same as apply, but NotInModule(_) predicates yield false. That is, privately uninhabited types are considered always uninhabited.

fn apply_inner<E>( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, in_module: &impl Fn(DefId) -> Result<bool, E> ) -> Result<bool, E>

pub fn and(self, tcx: TyCtxt<'tcx>, other: Self) -> Self

pub fn or(self, tcx: TyCtxt<'tcx>, other: Self) -> Self

pub fn all(tcx: TyCtxt<'tcx>, iter: impl IntoIterator<Item = Self>) -> Self

pub fn any(tcx: TyCtxt<'tcx>, iter: impl IntoIterator<Item = Self>) -> Self

fn reduce_and(self, tcx: TyCtxt<'tcx>, other: Self) -> Option<Self>

fn reduce_or(self, tcx: TyCtxt<'tcx>, other: Self) -> Option<Self>

pub fn instantiate(self, tcx: TyCtxt<'tcx>, args: GenericArgsRef<'tcx>) -> Self

Replaces generic types with its corresponding predicate

fn instantiate_opt( self, tcx: TyCtxt<'tcx>, args: GenericArgsRef<'tcx> ) -> Option<Self>