Type Alias rustc_middle::query::queries::live_symbols_and_ignored_derived_traits::ProvidedValue

pub type ProvidedValue<'tcx> = (UnordSet<LocalDefId>, UnordMap<LocalDefId, Vec<(DefId, DefId), Global>>);

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.