Type Alias rustc_middle::query::queries::method_autoderef_steps::ProvidedValue

pub type ProvidedValue<'tcx> = MethodAutoderefStepsResult<'tcx>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx> {
    pub steps: &'tcx [CandidateStep<'tcx>],
    pub opt_bad_ty: Option<&'tcx MethodAutoderefBadTy<'tcx>>,
    pub reached_recursion_limit: bool,
}

Fields
steps: &'tcx [CandidateStep<'tcx>]

The valid autoderef steps that could be found.

opt_bad_ty: Option<&'tcx MethodAutoderefBadTy<'tcx>>

If Some(T), a type autoderef reported an error on.

reached_recursion_limit: bool

If true, steps has been truncated due to reaching the recursion limit.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 32 bytes