rustc_middle::query::queries::mir_const_qualif

Type Alias ProvidedValue

pub type ProvidedValue<'tcx> = ConstQualifs;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as ArenaCached>::Provided instead.

Aliased Type

pub struct ProvidedValue<'tcx> {
    pub has_mut_interior: bool,
    pub needs_drop: bool,
    pub needs_non_const_drop: bool,
    pub tainted_by_errors: Option<ErrorGuaranteed>,
}

Fields
has_mut_interior: boolneeds_drop: boolneeds_non_const_drop: booltainted_by_errors: Option<ErrorGuaranteed>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 4 bytes