Type Alias rustc_middle::query::queries::mir_generator_witnesses::ProvidedValue

pub type ProvidedValue<'tcx> = Option<GeneratorLayout<'tcx>>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.