Type Alias rustc_middle::query::queries::mir_shims::ProvidedValue

pub type ProvidedValue<'tcx> = Body<'tcx>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Implementations

impl<'tcx> Body<'tcx>

pub fn new( source: MirSource<'tcx>, basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>>, source_scopes: IndexVec<SourceScope, SourceScopeData<'tcx>>, local_decls: IndexVec<Local, LocalDecl<'tcx>>, user_type_annotations: CanonicalUserTypeAnnotations<'tcx>, arg_count: usize, var_debug_info: Vec<VarDebugInfo<'tcx>>, span: Span, generator_kind: Option<GeneratorKind>, tainted_by_errors: Option<ErrorGuaranteed> ) -> Self

pub fn new_cfg_only( basic_blocks: IndexVec<BasicBlock, BasicBlockData<'tcx>> ) -> Self

Returns a partially initialized MIR body containing only a list of basic blocks.

The returned MIR contains no LocalDecls (even for the return place) or source scopes. It is only useful for testing but cannot be #[cfg(test)] because it is used in a different crate.

pub fn basic_blocks_mut( &mut self ) -> &mut IndexVec<BasicBlock, BasicBlockData<'tcx>>

pub fn local_kind(&self, local: Local) -> LocalKind

pub fn mut_vars_iter<'a>( &'a self ) -> impl Iterator<Item = Local> + Captures<'tcx> + 'a

Returns an iterator over all user-declared mutable locals.

pub fn mut_vars_and_args_iter<'a>( &'a self ) -> impl Iterator<Item = Local> + Captures<'tcx> + 'a

Returns an iterator over all user-declared mutable arguments and locals.

pub fn args_iter(&self) -> impl Iterator<Item = Local> + ExactSizeIterator

Returns an iterator over all function arguments.

pub fn vars_and_temps_iter( &self ) -> impl DoubleEndedIterator<Item = Local> + ExactSizeIterator

Returns an iterator over all user-defined variables and compiler-generated temporaries (all locals that are neither arguments nor the return place).

pub fn drain_vars_and_temps<'a>( &'a mut self ) -> impl Iterator<Item = LocalDecl<'tcx>> + 'a

pub fn source_info(&self, location: Location) -> &SourceInfo

Returns the source info associated with location.

pub fn return_ty(&self) -> Ty<'tcx>

Returns the return type; it always return first element from local_decls array.

pub fn bound_return_ty(&self) -> EarlyBinder<Ty<'tcx>>

Returns the return type; it always return first element from local_decls array.

pub fn terminator_loc(&self, bb: BasicBlock) -> Location

Gets the location of the terminator for the given block.

pub fn stmt_at( &self, location: Location ) -> Either<&Statement<'tcx>, &Terminator<'tcx>>

pub fn yield_ty(&self) -> Option<Ty<'tcx>>

pub fn generator_layout(&self) -> Option<&GeneratorLayout<'tcx>>

pub fn generator_drop(&self) -> Option<&Body<'tcx>>

pub fn generator_kind(&self) -> Option<GeneratorKind>

pub fn should_skip(&self) -> bool

pub fn is_custom_mir(&self) -> bool

pub fn post_mono_checks( &self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, normalize_const: impl Fn(Const<'tcx>) -> Result<Const<'tcx>, ErrorHandled> ) -> Result<(), ErrorHandled>

Must be called once the full substitution for this body is known, to ensure that the body is indeed fit for code generation or consumption more generally.

Sadly there’s no nice way to represent an “arbitrary normalizer”, so we take one for constants specifically. (Option<GenericArgsRef> could be used for that, but the fact that Instance::args_for_mir_body is private and instead instance exposes normalization functions makes it seem like exposing the generic args is not the intended strategy.)

Also sadly, CTFE doesn’t even know whether it runs on MIR that is already polymorphic or still monomorphic, so we cannot just immediately ICE on TooGeneric.

Returns Ok(()) if everything went fine, and Err if a problem occurred and got reported.