Type Alias rustc_middle::query::queries::normalize_projection_ty::ProvidedValue

pub type ProvidedValue<'tcx> = Result<&'tcx Canonical<'tcx, QueryResponse<'tcx, NormalizationResult<'tcx>>>, NoSolution>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

enum ProvidedValue<'tcx> {
    Ok(&'tcx Canonical<'tcx, QueryResponse<'tcx, NormalizationResult<'tcx>>>),
    Err(NoSolution),
}

Variants

Ok(&'tcx Canonical<'tcx, QueryResponse<'tcx, NormalizationResult<'tcx>>>)

Contains the success value

Err(NoSolution)

Contains the error value