Type Alias rustc_middle::query::queries::param_env_reveal_all_normalized::Value

pub type Value<'tcx> = ParamEnv<'tcx>;

Aliased Type

struct Value<'tcx> {
    packed: CopyTaggedPtr<&'tcx List<Clause<'tcx>>, ParamTag, true>,
}

Fields
packed: CopyTaggedPtr<&'tcx List<Clause<'tcx>>, ParamTag, true>

This packs both caller bounds and the reveal enum into one pointer.

Caller bounds are Obligations that the caller must satisfy. This is basically the set of bounds on the in-scope type parameters, translated into Obligations, and elaborated and normalized.

Use the caller_bounds() method to access.

Typically, this is Reveal::UserFacing, but during codegen we want Reveal::All.

Note: This is packed, use the reveal() method to access it.

Implementations

impl<'tcx> ParamEnv<'tcx>

pub fn empty() -> Self

Construct a trait environment suitable for contexts where there are no where-clauses in scope. Hidden types (like impl Trait) are left hidden, so this is suitable for ordinary type-checking.

pub fn caller_bounds(self) -> &'tcx List<Clause<'tcx>>

pub fn reveal(self) -> Reveal

pub fn reveal_all() -> Self

Construct a trait environment with no where-clauses in scope where the values of all impl Trait and other hidden types are revealed. This is suitable for monomorphized, post-typeck environments like codegen or doing optimizations.

N.B., if you want to have predicates in scope, use ParamEnv::new, or invoke param_env.with_reveal_all().

pub fn new(caller_bounds: &'tcx List<Clause<'tcx>>, reveal: Reveal) -> Self

Construct a trait environment with the given set of predicates.

pub fn with_user_facing(self) -> Self

pub fn with_reveal_all_normalized(self, tcx: TyCtxt<'tcx>) -> Self

Returns a new parameter environment with the same clauses, but which “reveals” the true results of projections in all cases (even for associated types that are specializable). This is the desired behavior during codegen and certain other special contexts; normally though we want to use Reveal::UserFacing, which is the default. All opaque types in the caller_bounds of the ParamEnv will be normalized to their underlying types. See PR #65989 and issue #65918 for more details

pub fn without_caller_bounds(self) -> Self

Returns this same environment but with no caller bounds.

pub fn and<T: TypeVisitable<TyCtxt<'tcx>>>( self, value: T ) -> ParamEnvAnd<'tcx, T>

Creates a suitable environment in which to perform trait queries on the given value. When type-checking, this is simply the pair of the environment plus value. But when reveal is set to All, then if value does not reference any type parameters, we will pair it with the empty environment. This improves caching and is generally invisible.

N.B., we preserve the environment when type-checking because it is possible for the user to have wacky where-clauses like where Box<u32>: Copy, which are clearly never satisfiable. We generally want to behave as if they were true, although the surrounding function is never reachable.