Type Alias rustc_middle::query::queries::proc_macro_decls_static::ProvidedValue

pub type ProvidedValue<'tcx> = Option<LocalDefId>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

enum ProvidedValue<'tcx> {
    None,
    Some(LocalDefId),
}

Variants

None

No value.

Some(LocalDefId)

Some value of type T.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 4 bytes

Size for each variant:

None: 0 bytes

Some: 4 bytes