Type Alias rustc_middle::query::queries::representability_adt_ty::ProvidedValue

pub type ProvidedValue<'tcx> = Representability;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

enum ProvidedValue<'tcx> {
    Representable,
    Infinite,
}

Variants

Representable

Infinite