rustc_middle::query::queries::resolve_instance_raw

Type Alias ProvidedValue

pub type ProvidedValue<'tcx> = Result<Option<Instance<'tcx>>, ErrorGuaranteed>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as ArenaCached>::Provided instead.

Aliased Type

pub enum ProvidedValue<'tcx> {
    Ok(Option<Instance<'tcx>>),
    Err(ErrorGuaranteed),
}

Variants

Ok(Option<Instance<'tcx>>)

Contains the success value

Err(ErrorGuaranteed)

Contains the error value

Layout

Note: Unable to compute type layout, possibly due to this type having generic parameters. Layout can only be computed for concrete, fully-instantiated types.