Type Alias rustc_middle::query::queries::super_predicates_that_define_assoc_item::ProvidedValue

pub type ProvidedValue<'tcx> = GenericPredicates<'tcx>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx> {
    pub parent: Option<DefId>,
    pub predicates: &'tcx [(Clause<'tcx>, Span)],
}

Fields
parent: Option<DefId>predicates: &'tcx [(Clause<'tcx>, Span)]
Implementations

impl<'tcx> GenericPredicates<'tcx>

pub fn instantiate( &self, tcx: TyCtxt<'tcx>, args: GenericArgsRef<'tcx> ) -> InstantiatedPredicates<'tcx>

pub fn instantiate_own( &self, tcx: TyCtxt<'tcx>, args: GenericArgsRef<'tcx> ) -> impl Iterator<Item = (Clause<'tcx>, Span)> + DoubleEndedIterator + ExactSizeIterator

fn instantiate_into( &self, tcx: TyCtxt<'tcx>, instantiated: &mut InstantiatedPredicates<'tcx>, args: GenericArgsRef<'tcx> )

pub fn instantiate_identity( &self, tcx: TyCtxt<'tcx> ) -> InstantiatedPredicates<'tcx>

fn instantiate_identity_into( &self, tcx: TyCtxt<'tcx>, instantiated: &mut InstantiatedPredicates<'tcx> )