Type Alias rustc_middle::query::queries::symbol_name::ProvidedValue

pub type ProvidedValue<'tcx> = SymbolName<'tcx>;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx> {
    pub name: &'tcx str,
}

Fields
name: &'tcx str

&str gives a consistent ordering, which ensures reproducible builds.

Implementations

impl<'tcx> SymbolName<'tcx>

pub fn new(tcx: TyCtxt<'tcx>, name: &str) -> SymbolName<'tcx>