Type Alias rustc_middle::query::queries::trait_def::ProvidedValue

pub type ProvidedValue<'tcx> = TraitDef;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Implementations

impl<'tcx> TraitDef

pub fn ancestors( &self, tcx: TyCtxt<'tcx>, of_impl: DefId ) -> Result<Ancestors<'tcx>, ErrorGuaranteed>