Type Alias rustc_middle::query::queries::trait_impls_in_crate::ProvidedValue

pub type ProvidedValue<'tcx> = &'tcx [DefId];

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.