Type Alias rustc_middle::query::queries::unused_generic_params::ProvidedValue

pub type ProvidedValue<'tcx> = UnusedGenericParams;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx>(FiniteBitSet<u32>);

Fields
0: FiniteBitSet<u32>
Implementations

impl UnusedGenericParams

pub fn new_all_unused(amount: u32) -> Self

pub fn new_all_used() -> Self

pub fn mark_used(&mut self, idx: u32)

pub fn is_unused(&self, idx: u32) -> bool

pub fn is_used(&self, idx: u32) -> bool

pub fn all_used(&self) -> bool

pub fn bits(&self) -> u32

pub fn from_bits(bits: u32) -> UnusedGenericParams