Type Alias rustc_middle::query::queries::visibility::Value

pub type Value<'tcx> = Visibility<DefId>;

Aliased Type

enum Value<'tcx> {
    Public,
    Restricted(DefId),
}

Variants

Public

Visible everywhere (including in other crates).

Restricted(DefId)

Visible only in the given crate-local module.

Implementations

impl<Id> Visibility<Id>

pub fn is_public(self) -> bool

pub fn map_id<OutId>(self, f: impl FnOnce(Id) -> OutId) -> Visibility<OutId>

impl<Id: Into<DefId>> Visibility<Id>

pub fn to_def_id(self) -> Visibility<DefId>

pub fn is_accessible_from( self, module: impl Into<DefId>, tcx: TyCtxt<'_> ) -> bool

Returns true if an item with this visibility is accessible from the given module.

pub fn is_at_least( self, vis: Visibility<impl Into<DefId>>, tcx: TyCtxt<'_> ) -> bool

Returns true if this visibility is at least as accessible as the given visibility

impl Visibility<DefId>

pub fn expect_local(self) -> Visibility

pub fn is_visible_locally(self) -> bool

Returns true if this item is visible anywhere in the local crate.