Type Alias rustc_middle::query::queries::vtable_allocation::ProvidedValue

pub type ProvidedValue<'tcx> = AllocId;

This type alias specifies the type returned from query providers and the type used for decoding. For regular queries this is the declared returned type V, but arena_cache will use <V as Deref>::Target instead.

Aliased Type

struct ProvidedValue<'tcx>(pub NonZeroU64);

Fields
0: NonZeroU64