Type Alias rustc_middle::query::queries::vtable_entries::LocalKey

pub type LocalKey<'tcx> = PolyTraitRef<'tcx>;

Aliased Type

struct LocalKey<'tcx> {
    value: TraitRef<'tcx>,
    bound_vars: &'tcx List<BoundVariableKind>,
}

Fields
value: TraitRef<'tcx>bound_vars: &'tcx List<BoundVariableKind>
Implementations

impl<'tcx> Binder<'tcx, TraitRef<'tcx>>

pub fn print_only_trait_path( self ) -> Binder<'tcx, TraitRefPrintOnlyTraitPath<'tcx>>

impl<'tcx> Binder<'tcx, TraitRef<'tcx>>

pub fn self_ty(&self) -> Binder<'tcx, Ty<'tcx>>

pub fn def_id(&self) -> DefId

impl<'tcx, T> Binder<'tcx, T>where T: TypeVisitable<TyCtxt<'tcx>>,

pub fn dummy(value: T) -> Binder<'tcx, T>

Wraps value in a binder, asserting that value does not contain any bound vars that would be bound by the binder. This is commonly used to ‘inject’ a value T into a different binding level.

pub fn bind_with_vars( value: T, bound_vars: &'tcx List<BoundVariableKind> ) -> Binder<'tcx, T>

impl<'tcx, T> Binder<'tcx, T>

pub fn skip_binder(self) -> T

Skips the binder and returns the “bound” value. This is a risky thing to do because it’s easy to get confused about De Bruijn indices and the like. It is usually better to discharge the binder using no_bound_vars or replace_late_bound_regions or something like that. skip_binder is only valid when you are either extracting data that has nothing to do with bound vars, you are doing some sort of test that does not involve bound regions, or you are being very careful about your depth accounting.

Some examples where skip_binder is reasonable:

extracting the DefId from a PolyTraitRef;

comparing the self type of a PolyTraitRef to see if it is equal to a type parameter X, since the type X does not reference any regions

pub fn bound_vars(&self) -> &'tcx List<BoundVariableKind>

pub fn as_ref(&self) -> Binder<'tcx, &T>

pub fn as_deref(&self) -> Binder<'tcx, &T::Target>where T: Deref,

pub fn map_bound_ref_unchecked<F, U>(&self, f: F) -> Binder<'tcx, U>where F: FnOnce(&T) -> U,

pub fn map_bound_ref<F, U: TypeVisitable<TyCtxt<'tcx>>>( &self, f: F ) -> Binder<'tcx, U>where F: FnOnce(&T) -> U,

pub fn map_bound<F, U: TypeVisitable<TyCtxt<'tcx>>>( self, f: F ) -> Binder<'tcx, U>where F: FnOnce(T) -> U,

pub fn try_map_bound<F, U: TypeVisitable<TyCtxt<'tcx>>, E>( self, f: F ) -> Result<Binder<'tcx, U>, E>where F: FnOnce(T) -> Result<U, E>,

pub fn rebind<U>(&self, value: U) -> Binder<'tcx, U>where U: TypeVisitable<TyCtxt<'tcx>>,

Wraps a value in a binder, using the same bound variables as the current Binder. This should not be used if the new value changes the bound variables. Note: the (old or new) value itself does not necessarily need to name all the bound variables.

This currently doesn’t do anything different than bind, because we don’t actually track bound vars. However, semantically, it is different because bound vars aren’t allowed to change here, whereas they are in bind. This may be (debug) asserted in the future.

pub fn no_bound_vars(self) -> Option<T>where T: TypeVisitable<TyCtxt<'tcx>>,

Unwraps and returns the value within, but only if it contains no bound vars at all. (In other words, if this binder – and indeed any enclosing binder – doesn’t bind anything at all.) Otherwise, returns None.

(One could imagine having a method that just unwraps a single binder, but permits late-bound vars bound by enclosing binders, but that would require adjusting the debruijn indices, and given the shallow binding structure we often use, would not be that useful.)

pub fn split<U, V, F>(self, f: F) -> (Binder<'tcx, U>, Binder<'tcx, V>)where F: FnOnce(T) -> (U, V),

Splits the contents into two things that share the same binder level as the original, returning two distinct binders.

f should consider bound regions at depth 1 to be free, and anything it produces with bound regions at depth 1 will be bound in the resulting return values.

impl<'tcx, T: IntoIterator> Binder<'tcx, T>

pub fn iter(self) -> impl Iterator<Item = Binder<'tcx, T::Item>>