Enum rustc_middle::thir::PatKind

pub enum PatKind<'tcx> {
    Wild,
    AscribeUserType {
        ascription: Ascription<'tcx>,
        subpattern: Box<Pat<'tcx>>,
    },
    Binding {
        mutability: Mutability,
        name: Symbol,
        mode: BindingMode,
        var: LocalVarId,
        ty: Ty<'tcx>,
        subpattern: Option<Box<Pat<'tcx>>>,
        is_primary: bool,
    },
    Variant {
        adt_def: AdtDef<'tcx>,
        substs: SubstsRef<'tcx>,
        variant_index: VariantIdx,
        subpatterns: Vec<FieldPat<'tcx>>,
    },
    Leaf {
        subpatterns: Vec<FieldPat<'tcx>>,
    },
    Deref {
        subpattern: Box<Pat<'tcx>>,
    },
    Constant {
        value: ConstantKind<'tcx>,
    },
    Range(Box<PatRange<'tcx>>),
    Slice {
        prefix: Box<[Box<Pat<'tcx>>]>,
        slice: Option<Box<Pat<'tcx>>>,
        suffix: Box<[Box<Pat<'tcx>>]>,
    },
    Array {
        prefix: Box<[Box<Pat<'tcx>>]>,
        slice: Option<Box<Pat<'tcx>>>,
        suffix: Box<[Box<Pat<'tcx>>]>,
    },
    Or {
        pats: Box<[Box<Pat<'tcx>>]>,
    },
}

Variants

Wild

A wildcard pattern: _.

AscribeUserType

Fields

ascription: Ascription<'tcx>

subpattern: Box<Pat<'tcx>>

Binding

Fields

mutability: Mutability

name: Symbol

mode: BindingMode

var: LocalVarId

ty: Ty<'tcx>

subpattern: Option<Box<Pat<'tcx>>>

is_primary: bool

Is this the leftmost occurrence of the binding, i.e., is var the HirId of this pattern?

x, ref x, x @ P, etc.

Variant

Fields

adt_def: AdtDef<'tcx>

substs: SubstsRef<'tcx>

variant_index: VariantIdx

subpatterns: Vec<FieldPat<'tcx>>

Foo(...) or Foo{...} or Foo, where Foo is a variant name from an ADT with multiple variants.

Leaf

Fields

subpatterns: Vec<FieldPat<'tcx>>

(...), Foo(...), Foo{...}, or Foo, where Foo is a variant name from an ADT with a single variant.

Deref

Fields

subpattern: Box<Pat<'tcx>>

box P, &P, &mut P, etc.

Constant

Fields

value: ConstantKind<'tcx>

One of the following:

&str, which will be handled as a string pattern and thus exhaustiveness checking will detect if you use the same string twice in different patterns.

integer, bool, char or float, which will be handled by exhaustiveness to cover exactly its own value, similar to &str, but these values are much simpler.

Opaque constants, that must not be matched structurally. So anything that does not derive PartialEq and Eq.

Range(Box<PatRange<'tcx>>)

Slice

Fields

prefix: Box<[Box<Pat<'tcx>>]>

slice: Option<Box<Pat<'tcx>>>

suffix: Box<[Box<Pat<'tcx>>]>

Matches against a slice, checking the length and extracting elements. irrefutable when there is a slice pattern and both prefix and suffix are empty. e.g., &[ref xs @ ..].

Array

Fields

prefix: Box<[Box<Pat<'tcx>>]>

slice: Option<Box<Pat<'tcx>>>

suffix: Box<[Box<Pat<'tcx>>]>

Fixed match against an array; irrefutable.

Or

Fields

pats: Box<[Box<Pat<'tcx>>]>

An or-pattern, e.g. p | q. Invariant: pats.len() >= 2.