rustc_middle::thir

Module visit

Traits

Visitor

Every walk_* method uses deconstruction to access fields of structs and enums. This will result in a compile error if a field is added, which makes it more likely the appropriate visit call will be added for it.

Functions

for_each_immediate_subpat 🔒

Invokes callback on each immediate subpattern of pat, if any. A building block for assembling THIR pattern visitors.

walk_arm

walk_block

walk_expr

walk_pat

walk_stmt