Module rustc_middle::traits

Trait Resolution. See the rustc dev guide for more information on how this works.

Re-exports

pub use self::select::EvaluationCache;

pub use self::select::EvaluationResult;

pub use self::select::OverflowError;

pub use self::select::SelectionCache;

pub use self::ObligationCauseCode::*;

Modules

chalk 🔒

Types required for Chalk-related queries

query

Experimental types for the trait query interface. The methods defined in this module are all based on canonicalization, which makes a canonical query by replacing unbound inference variables and regions, so that results can be reused more broadly. The providers for the queries defined here can be found in rustc_traits.

select

Candidate selection. See the rustc dev guide for more information on how this works.

specialization_graph

structural_impls 🔒

util

Structs

ChalkEnvironmentAndGoal

A chalk environment and goal.

ChalkRustInterner

DerivedObligationCause

IfExpressionCause

ImplDerivedObligationCause

ImplSourceAutoImplData

ImplSourceBuiltinData

ImplSourceClosureData

ImplSourceConstDestructData

ImplSourceFnPointerData

ImplSourceFutureData

ImplSourceGeneratorData

ImplSourceObjectData

ImplSourceTraitAliasData

ImplSourceTraitUpcastingData

ImplSourceUserDefinedData

Identifies a particular impl in the source, along with a set of substitutions from the impl’s type/lifetime parameters. The nested vector corresponds to the nested obligations attached to the impl’s type parameters.

InternedObligationCauseCode

MatchExpressionArmCause

ObligationCause

The reason why we incurred this obligation; used for error reporting.

UnifyReceiverContext

Enums

CodegenObligationError

These are the error cases for codegen_select_candidate.

ImplSource

Given the successful resolution of an obligation, the ImplSource indicates where the impl comes from.

MethodViolationCode

Reasons a method might not be object-safe.

ObjectSafetyViolation

ObligationCauseCode

Reveal

Depending on the stage of compilation, we want projection to be more or less conservative.

SelectionError

StatementAsExpression

WellFormedLoc

The ‘location’ at which we try to perform HIR-based wf checking. This information is used to obtain an hir::Ty, which we can walk in order to obtain precise spans for any ‘nested’ types (e.g. Foo in Option<Foo>).

Type Definitions

CanonicalChalkEnvironmentAndGoal

SelectionResult

When performing resolution, it is typically the case that there can be one of three outcomes: