Enum rustc_middle::traits::ImplSource

pub enum ImplSource<'tcx, N> {
    UserDefined(ImplSourceUserDefinedData<'tcx, N>),
    AutoImpl(ImplSourceAutoImplData<N>),
    Param(Vec<N>, BoundConstness),
    Object(ImplSourceObjectData<'tcx, N>),
    Builtin(ImplSourceBuiltinData<N>),
    TraitUpcasting(ImplSourceTraitUpcastingData<'tcx, N>),
    Closure(ImplSourceClosureData<'tcx, N>),
    FnPointer(ImplSourceFnPointerData<'tcx, N>),
    Generator(ImplSourceGeneratorData<'tcx, N>),
    Future(ImplSourceFutureData<'tcx, N>),
    TraitAlias(ImplSourceTraitAliasData<'tcx, N>),
    ConstDestruct(ImplSourceConstDestructData<N>),
}

Given the successful resolution of an obligation, the ImplSource indicates where the impl comes from.

For example, the obligation may be satisfied by a specific impl (case A), or it may be relative to some bound that is in scope (case B).

impl<T:Clone> Clone<T> for Option<T> { ... } // Impl_1
impl<T:Clone> Clone<T> for Box<T> { ... }    // Impl_2
impl Clone for i32 { ... }                   // Impl_3

fn foo<T: Clone>(concrete: Option<Box<i32>>, param: T, mixed: Option<T>) {
    // Case A: ImplSource points at a specific impl. Only possible when
    // type is concretely known. If the impl itself has bounded
    // type parameters, ImplSource will carry resolutions for those as well:
    concrete.clone(); // ImplSource(Impl_1, [ImplSource(Impl_2, [ImplSource(Impl_3)])])

    // Case B: ImplSource must be provided by caller. This applies when
    // type is a type parameter.
    param.clone();    // ImplSource::Param

    // Case C: A mix of cases A and B.
    mixed.clone();    // ImplSource(Impl_1, [ImplSource::Param])
}

The type parameter N

See explanation on ImplSourceUserDefinedData.

Variants

UserDefined(ImplSourceUserDefinedData<'tcx, N>)

ImplSource identifying a particular impl.

AutoImpl(ImplSourceAutoImplData<N>)

ImplSource for auto trait implementations. This carries the information and nested obligations with regards to an auto implementation for a trait Trait. The nested obligations ensure the trait implementation holds for all the constituent types.

Param(Vec<N>, BoundConstness)

Successful resolution to an obligation provided by the caller for some type parameter. The Vec<N> represents the obligations incurred from normalizing the where-clause (if any).

Object(ImplSourceObjectData<'tcx, N>)

Virtual calls through an object.

Builtin(ImplSourceBuiltinData<N>)

Successful resolution for a builtin trait.

TraitUpcasting(ImplSourceTraitUpcastingData<'tcx, N>)

ImplSource for trait upcasting coercion

Closure(ImplSourceClosureData<'tcx, N>)

ImplSource automatically generated for a closure. The DefId is the ID of the closure expression. This is an ImplSource::UserDefined in spirit, but the impl is generated by the compiler and does not appear in the source.

FnPointer(ImplSourceFnPointerData<'tcx, N>)

Same as above, but for a function pointer type with the given signature.

Generator(ImplSourceGeneratorData<'tcx, N>)

ImplSource automatically generated for a generator.

Future(ImplSourceFutureData<'tcx, N>)

ImplSource automatically generated for a generator backing an async future.

TraitAlias(ImplSourceTraitAliasData<'tcx, N>)

ImplSource for a trait alias.

ConstDestruct(ImplSourceConstDestructData<N>)

ImplSource for a const Drop implementation.

Implementations

impl<'tcx, N> ImplSource<'tcx, N>

pub fn nested_obligations(self) -> Vec<N>

pub fn borrow_nested_obligations(&self) -> &[N]

pub fn map<M, F>(self, f: F) -> ImplSource<'tcx, M>where
F: FnMut(N) -> M,