Type Alias rustc_middle::traits::query::CanonicalTypeOpNormalizeGoal

pub type CanonicalTypeOpNormalizeGoal<'tcx, T> = Canonical<'tcx, ParamEnvAnd<'tcx, Normalize<T>>>;

Aliased Type

struct CanonicalTypeOpNormalizeGoal<'tcx, T> {
    pub value: ParamEnvAnd<'tcx, Normalize<T>>,
    pub max_universe: UniverseIndex,
    pub variables: &'tcx List<CanonicalVarInfo<'tcx>>,
}

Fields
value: ParamEnvAnd<'tcx, Normalize<T>>max_universe: UniverseIndexvariables: &'tcx List<CanonicalVarInfo<'tcx>>
Implementations

impl<'tcx, V> Canonical<'tcx, V>

pub fn unchecked_map<W>(self, map_op: impl FnOnce(V) -> W) -> Canonical<'tcx, W>

Allows you to map the value of a canonical while keeping the same set of bound variables.

WARNING: This function is very easy to mis-use, hence the name! In particular, the new value W must use all the same type/region variables in precisely the same order as the original! (The ordering is defined by the TypeFoldable implementation of the type in question.)

An example of a correct use of this:

let a: Canonical<'_, T> = ...;
let b: Canonical<'_, (T,)> = a.unchecked_map(|v| (v, ));

An example of an incorrect use of this:

let a: Canonical<'tcx, T> = ...;
let ty: Ty<'tcx> = ...;
let b: Canonical<'tcx, (T, Ty<'tcx>)> = a.unchecked_map(|v| (v, ty));

pub fn unchecked_rebind<W>(self, value: W) -> Canonical<'tcx, W>

Allows you to map the value of a canonical while keeping the same set of bound variables.

WARNING: This function is very easy to mis-use, hence the name! See the comment of Canonical::unchecked_map for more details.