Struct rustc_middle::traits::solve::EvaluationCache

pub struct EvaluationCache<'tcx> {
    map: Lock<FxHashMap<CanonicalInput<'tcx>, CacheEntry<'tcx>>>,
}

The trait solver cache used by -Ztrait-solver=next.

FIXME(@lcnr): link to some official documentation of how this works.

Fields
map: Lock<FxHashMap<CanonicalInput<'tcx>, CacheEntry<'tcx>>>
Implementations

impl<'tcx> EvaluationCache<'tcx>

pub fn insert( &self, key: CanonicalInput<'tcx>, reached_depth: usize, did_overflow: bool, cycle_participants: FxHashSet<CanonicalInput<'tcx>>, dep_node: DepNodeIndex, result: QueryResult<'tcx> )

Insert a final result into the global cache.

pub fn get( &self, tcx: TyCtxt<'tcx>, key: CanonicalInput<'tcx>, cycle_participant_in_stack: impl FnOnce(&FxHashSet<CanonicalInput<'tcx>>) -> bool, available_depth: Limit ) -> Option<CacheData<'tcx>>

Try to fetch a cached result, checking the recursion limit and handling root goals of coinductive cycles.

If this returns Some the cache result can be used.