Struct rustc_middle::traits::ObligationCause

pub struct ObligationCause<'tcx> {
    pub span: Span,
    pub body_id: HirId,
    code: InternedObligationCauseCode<'tcx>,
}

The reason why we incurred this obligation; used for error reporting.

Non-misc ObligationCauseCodes are stored on the heap. This gives the best trade-off between keeping the type small (which makes copies cheaper) while not doing too many heap allocations.

We do not want to intern this as there are a lot of obligation causes which only live for a short period of time.

Fields
span: Spanbody_id: HirId

The ID of the fn body that triggered this obligation. This is used for region obligations to determine the precise environment in which the region obligation should be evaluated (in particular, closures can add new assumptions). See the field region_obligations of the FulfillmentContext for more information.

code: InternedObligationCauseCode<'tcx>
Implementations

impl<'tcx> ObligationCause<'tcx>

pub fn new(
span: Span,
body_id: HirId,
code: ObligationCauseCode<'tcx>
) -> ObligationCause<'tcx>

pub fn misc(span: Span, body_id: HirId) -> ObligationCause<'tcx>

pub fn dummy() -> ObligationCause<'tcx>

pub fn dummy_with_span(span: Span) -> ObligationCause<'tcx>

pub fn span(&self) -> Span

pub fn code(&self) -> &ObligationCauseCode<'tcx>

pub fn map_code(
&mut self,
f: impl FnOnce(InternedObligationCauseCode<'tcx>) -> ObligationCauseCode<'tcx>
)

pub fn derived_cause(
self,
parent_trait_pred: PolyTraitPredicate<'tcx>,
variant: impl FnOnce(DerivedObligationCause<'tcx>) -> ObligationCauseCode<'tcx>
) -> ObligationCause<'tcx>

Creates a cause for obligations that are derived from obligation by a recursive search (e.g., for a builtin bound, or eventually a auto trait Foo). If obligation is itself a derived obligation, this is just a clone, but otherwise we create a “derived obligation” cause so as to keep track of the original root obligation for error reporting.