Struct rustc_middle::ty::adt::AdtDefData

pub struct AdtDefData {
    pub did: DefId,
    variants: IndexVec<VariantIdx, VariantDef>,
    flags: AdtFlags,
    repr: ReprOptions,
}

The definition of a user-defined type, e.g., a struct, enum, or union.

These are all interned (by alloc_adt_def) into the global arena.

The initialism ADT stands for an algebraic data type (ADT). This is slightly wrong because unions are not ADTs. Moreover, Rust only allows recursive data types through indirection.

Recursive types

It may seem impossible to represent recursive types using Ty, since TyKind::Adt includes AdtDef, which includes its fields, creating a cycle. However, AdtDef does not actually include the types of its fields; it includes just their DefIds.

For example, the following type:

struct S { x: Box<S> }

is essentially represented with Ty as the following pseudocode:

struct S { x }

where x here represents the DefId of S.x. Then, the DefId can be used with TyCtxt::type_of() to get the type of the field.

Fields
did: DefId

The DefId of the struct, enum or union item.

variants: IndexVec<VariantIdx, VariantDef>

Variants of the ADT. If this is a struct or union, then there will be a single variant.

flags: AdtFlags

Flags of the ADT (e.g., is this a struct? is this non-exhaustive?).

repr: ReprOptions

Repr options provided by the user.

Implementations

impl AdtDefData

pub(super) fn new(
tcx: TyCtxt<'_>,
did: DefId,
kind: AdtKind,
variants: IndexVec<VariantIdx, VariantDef>,
repr: ReprOptions
) -> Self

Creates a new AdtDefData.