rustc_middle::ty::consts

Module int

Macros

from_scalar_int_for_x 🔒

from_scalar_int_for_x_signed 🔒

from_x_for_scalar_int 🔒

from_x_for_scalar_int_signed 🔒

Structs

CharTryFromScalarInt

Error returned when a conversion from ScalarInt to char fails.

ConstInt

A type for representing any integer. Only used for printing.

ScalarInt

The raw bytes of a simple value.

Enums

AtomicOrdering

An enum to represent the compiler-side view of intrinsics::AtomicOrdering. This lives here because there’s a method in this file that needs it and it is entirely unclear where else to put this…