rustc_middle::ty::consts::valtree

Type Alias ConstToValTreeResult

pub type ConstToValTreeResult<'tcx> = Result<Result<ValTree<'tcx>, Ty<'tcx>>, ErrorHandled>;

Ok(Err(ty)) indicates the constant was fine, but the valtree couldn’t be constructed because the value contains something of type ty that is not valtree-compatible. The caller can then show an appropriate error; the query does not have the necessary context to give good user-facing errors for this case.

Aliased Type

pub enum ConstToValTreeResult<'tcx> {
    Ok(Result<ValTree<'tcx>, Ty<'tcx>>),
    Err(ErrorHandled),
}

Variants

Ok(Result<ValTree<'tcx>, Ty<'tcx>>)

Contains the success value

Err(ErrorHandled)

Contains the error value

Layout

Note: Unable to compute type layout, possibly due to this type having generic parameters. Layout can only be computed for concrete, fully-instantiated types.