Struct rustc_middle::ty::context::GlobalCtxt

pub struct GlobalCtxt<'tcx> {

    pub arena: &'tcx WorkerLocal<Arena<'tcx>>,
    pub hir_arena: &'tcx WorkerLocal<Arena<'tcx>>,
    interners: CtxtInterners<'tcx>,
    pub sess: &'tcx Session,
    pub lint_store: Lrc<dyn Any + Sync + Send>,
    pub dep_graph: DepGraph,
    pub prof: SelfProfilerRef,
    pub types: CommonTypes<'tcx>,
    pub lifetimes: CommonLifetimes<'tcx>,
    pub consts: CommonConsts<'tcx>,
    untracked: Untracked,
    pub(crate) untracked_resolutions: ResolverGlobalCtxt,
    pub untracked_crate: Steal<Lrc<Crate>>,
    pub on_disk_cache: Option<&'tcx dyn OnDiskCache<'tcx>>,
    pub queries: &'tcx dyn QueryEngine<'tcx>,
    pub query_caches: QueryCaches<'tcx>,
    pub(crate) query_kinds: &'tcx [DepKindStruct<'tcx>],
    pub ty_rcache: Lock<FxHashMap<CReaderCacheKey, Ty<'tcx>>>,
    pub pred_rcache: Lock<FxHashMap<CReaderCacheKey, Predicate<'tcx>>>,
    pub selection_cache: SelectionCache<'tcx>,
    pub evaluation_cache: EvaluationCache<'tcx>,
    crate_name: Symbol,
    pub data_layout: TargetDataLayout,
    pub(crate) alloc_map: Lock<AllocMap<'tcx>>,
    output_filenames: Arc<OutputFilenames>,

}

Fields
arena: &'tcx WorkerLocal<Arena<'tcx>>hir_arena: &'tcx WorkerLocal<Arena<'tcx>>interners: CtxtInterners<'tcx>sess: &'tcx Sessionlint_store: Lrc<dyn Any + Sync + Send>

This only ever stores a LintStore but we don’t want a dependency on that type here.

FIXME(Centril): consider dyn LintStoreMarker once we can upcast to Any for some additional type safety.

dep_graph: DepGraphprof: SelfProfilerReftypes: CommonTypes<'tcx>

Common types, pre-interned for your convenience.

lifetimes: CommonLifetimes<'tcx>

Common lifetimes, pre-interned for your convenience.

consts: CommonConsts<'tcx>

Common consts, pre-interned for your convenience.

untracked: Untrackeduntracked_resolutions: ResolverGlobalCtxt

Output of the resolver.

untracked_crate: Steal<Lrc<Crate>>

The entire crate as AST. This field serves as the input for the hir_crate query, which lowers it from AST to HIR. It must not be read or used by anything else.

on_disk_cache: Option<&'tcx dyn OnDiskCache<'tcx>>

This provides access to the incremental compilation on-disk cache for query results. Do not access this directly. It is only meant to be used by DepGraph::try_mark_green() and the query infrastructure. This is None if we are not incremental compilation mode

queries: &'tcx dyn QueryEngine<'tcx>query_caches: QueryCaches<'tcx>query_kinds: &'tcx [DepKindStruct<'tcx>]ty_rcache: Lock<FxHashMap<CReaderCacheKey, Ty<'tcx>>>pred_rcache: Lock<FxHashMap<CReaderCacheKey, Predicate<'tcx>>>selection_cache: SelectionCache<'tcx>

Caches the results of trait selection. This cache is used for things that do not have to do with the parameters in scope.

evaluation_cache: EvaluationCache<'tcx>

Caches the results of trait evaluation. This cache is used for things that do not have to do with the parameters in scope. Merge this with selection_cache?

crate_name: Symbol

The definite name of the current crate after taking into account attributes, commandline parameters, etc.

data_layout: TargetDataLayout

Data layout specification for the current target.

alloc_map: Lock<AllocMap<'tcx>>

Stores memory for globals (statics/consts).

output_filenames: Arc<OutputFilenames>