Enum rustc_middle::ty::BoundConstness

pub enum BoundConstness {
    NotConst,
    ConstIfConst,
}

Variants

NotConst

T: Trait

ConstIfConst

T: ~const Trait

Requires resolving to const only when we are in a const context.

Implementations

impl BoundConstness

pub fn and(&mut self, constness: Constness) -> Constness

Reduce self and constness to two possible combined states instead of four.