Enum rustc_middle::ty::error::TypeError

pub enum TypeError<'tcx> {

    Mismatch,
    ConstnessMismatch(ExpectedFound<BoundConstness>),
    PolarityMismatch(ExpectedFound<ImplPolarity>),
    UnsafetyMismatch(ExpectedFound<Unsafety>),
    AbiMismatch(ExpectedFound<Abi>),
    Mutability,
    ArgumentMutability(usize),
    TupleSize(ExpectedFound<usize>),
    FixedArraySize(ExpectedFound<u64>),
    ArgCount,
    FieldMisMatch(Symbol, Symbol),
    RegionsDoesNotOutlive(Region<'tcx>, Region<'tcx>),
    RegionsInsufficientlyPolymorphic(BoundRegionKind, Region<'tcx>),
    RegionsPlaceholderMismatch,
    Sorts(ExpectedFound<Ty<'tcx>>),
    ArgumentSorts(ExpectedFound<Ty<'tcx>>, usize),
    IntMismatch(ExpectedFound<IntVarValue>),
    FloatMismatch(ExpectedFound<FloatTy>),
    Traits(ExpectedFound<DefId>),
    VariadicMismatch(ExpectedFound<bool>),
    CyclicTy(Ty<'tcx>),
    CyclicConst(Const<'tcx>),
    ProjectionMismatched(ExpectedFound<DefId>),
    ExistentialMismatch(ExpectedFound<&'tcx List<PolyExistentialPredicate<'tcx>>>),
    ConstMismatch(ExpectedFound<Const<'tcx>>),
    IntrinsicCast,
    TargetFeatureCast(DefId),

}

Variants

Mismatch

ConstnessMismatch(ExpectedFound<BoundConstness>)

PolarityMismatch(ExpectedFound<ImplPolarity>)

UnsafetyMismatch(ExpectedFound<Unsafety>)

AbiMismatch(ExpectedFound<Abi>)

Mutability

ArgumentMutability(usize)

TupleSize(ExpectedFound<usize>)

FixedArraySize(ExpectedFound<u64>)

ArgCount

FieldMisMatch(Symbol, Symbol)

RegionsDoesNotOutlive(Region<'tcx>, Region<'tcx>)

RegionsInsufficientlyPolymorphic(BoundRegionKind, Region<'tcx>)

RegionsPlaceholderMismatch

Sorts(ExpectedFound<Ty<'tcx>>)

ArgumentSorts(ExpectedFound<Ty<'tcx>>, usize)

IntMismatch(ExpectedFound<IntVarValue>)

FloatMismatch(ExpectedFound<FloatTy>)

Traits(ExpectedFound<DefId>)

VariadicMismatch(ExpectedFound<bool>)

CyclicTy(Ty<'tcx>)

Instantiating a type variable with the given type would have created a cycle (because it appears somewhere within that type).

CyclicConst(Const<'tcx>)

ProjectionMismatched(ExpectedFound<DefId>)

ExistentialMismatch(ExpectedFound<&'tcx List<PolyExistentialPredicate<'tcx>>>)

ConstMismatch(ExpectedFound<Const<'tcx>>)

IntrinsicCast

TargetFeatureCast(DefId)

Safe #[target_feature] functions are not assignable to safe function pointers.

Implementations

impl TypeError<'_>

pub fn involves_regions(self) -> bool

impl<'tcx> TypeError<'tcx>

Explains the source of a type err in a short, human readable way. This is meant to be placed in parentheses after some larger message. You should also invoke note_and_explain_type_err() afterwards to present additional details, particularly when it comes to lifetime-related errors.

pub fn to_string(self, tcx: TyCtxt<'tcx>) -> Cow<'static, str>

impl<'tcx> TypeError<'tcx>

pub fn must_include_note(self) -> bool