Struct rustc_middle::ty::flags::FlagComputation

pub struct FlagComputation {
    pub flags: TypeFlags,
    pub outer_exclusive_binder: DebruijnIndex,
}

Fields
flags: TypeFlagsouter_exclusive_binder: DebruijnIndex

see Ty::outer_exclusive_binder for details

Implementations

impl FlagComputation

fn new() -> FlagComputation

pub fn for_kind(kind: &TyKind<'_>) -> FlagComputation

pub fn for_predicate(binder: Binder<'_, PredicateKind<'_>>) -> FlagComputation

pub fn for_const(c: Const<'_>) -> TypeFlags

fn add_flags(&mut self, flags: TypeFlags)

fn add_bound_var(&mut self, binder: DebruijnIndex)

indicates that self refers to something at binding level binder

fn add_exclusive_binder(&mut self, exclusive_binder: DebruijnIndex)

indicates that self refers to something inside binding level binder – not bound by binder, but bound by the next binder internal to it

fn bound_computation<T, F>(&mut self, value: Binder<'_, T>, f: F)where F: FnOnce(&mut Self, T),

Adds the flags/depth from a set of types that appear within the current type, but within a region binder.

fn add_kind(&mut self, kind: &TyKind<'_>)

fn add_predicate(&mut self, binder: Binder<'_, PredicateKind<'_>>)

fn add_predicate_atom(&mut self, atom: PredicateKind<'_>)

fn add_ty(&mut self, ty: Ty<'_>)

fn add_tys(&mut self, tys: &[Ty<'_>])

fn add_region(&mut self, r: Region<'_>)

fn add_const(&mut self, c: Const<'_>)

fn add_existential_projection(&mut self, projection: &ExistentialProjection<'_>)

fn add_alias_ty(&mut self, alias_ty: AliasTy<'_>)

fn add_args(&mut self, args: &[GenericArg<'_>])

fn add_term(&mut self, term: Term<'_>)