Struct rustc_middle::ty::fold::RegionFolder

pub struct RegionFolder<'a, 'tcx> {
    tcx: TyCtxt<'tcx>,
    current_index: DebruijnIndex,
    fold_region_fn: &'a mut (dyn FnMut(Region<'tcx>, DebruijnIndex) -> Region<'tcx> + 'a),
}

Folds over the substructure of a type, visiting its component types and all regions that occur free within it.

That is, Ty can contain function or method types that bind regions at the call site (ReLateBound), and occurrences of regions (aka “lifetimes”) that are bound within a type are not visited by this folder; only regions that occur free will be visited by fld_r.

Fields
tcx: TyCtxt<'tcx>current_index: DebruijnIndex

Stores the index of a binder just outside the stuff we have visited. So this begins as INNERMOST; when we pass through a binder, it is incremented (via shift_in).

fold_region_fn: &'a mut (dyn FnMut(Region<'tcx>, DebruijnIndex) -> Region<'tcx> + 'a)

Callback invokes for each free region. The DebruijnIndex points to the binder just outside the ones we have passed through.

Implementations

impl<'a, 'tcx> RegionFolder<'a, 'tcx>

pub fn new(
tcx: TyCtxt<'tcx>,
fold_region_fn: &'a mut dyn FnMut(Region<'tcx>, DebruijnIndex) -> Region<'tcx>
) -> RegionFolder<'a, 'tcx>