Trait rustc_middle::ty::fold::TypeFolder

pub trait TypeFolder<'tcx>: FallibleTypeFolder<'tcx, Error = !> {
    fn tcx<'a>(&'a self) -> TyCtxt<'tcx>;

    fn fold_binder<T>(&mut self, t: Binder<'tcx, T>) -> Binder<'tcx, T>
    where
        T: TypeFoldable<'tcx>,
    { ... }
    fn fold_ty(&mut self, t: Ty<'tcx>) -> Ty<'tcx> { ... }
    fn fold_region(&mut self, r: Region<'tcx>) -> Region<'tcx> { ... }
    fn fold_const(&mut self, c: Const<'tcx>) -> Const<'tcx> { ... }
    fn fold_predicate(&mut self, p: Predicate<'tcx>) -> Predicate<'tcx> { ... }
}

This trait is implemented for every infallible folding traversal. There is a fold method defined for every type of interest. Each such method has a default that does an “identity” fold. Implementations of these methods often fall back to a super_fold_with method if the primary argument doesn’t satisfy a particular condition.

A blanket implementation of FallibleTypeFolder will defer to the infallible methods of this trait to ensure that the two APIs are coherent.

Required Methods

fn tcx<'a>(&'a self) -> TyCtxt<'tcx>

Provided Methods

fn fold_binder<T>(&mut self, t: Binder<'tcx, T>) -> Binder<'tcx, T>where
T: TypeFoldable<'tcx>,

fn fold_ty(&mut self, t: Ty<'tcx>) -> Ty<'tcx>

fn fold_region(&mut self, r: Region<'tcx>) -> Region<'tcx>

fn fold_const(&mut self, c: Const<'tcx>) -> Const<'tcx>

fn fold_predicate(&mut self, p: Predicate<'tcx>) -> Predicate<'tcx>