Struct rustc_middle::ty::generic_args::GenericArg

pub struct GenericArg<'tcx> {
    ptr: NonNull<()>,
    marker: PhantomData<(Ty<'tcx>, Region<'tcx>, Const<'tcx>)>,
}

An entity in the Rust type system, which can be one of several kinds (types, lifetimes, and consts). To reduce memory usage, a GenericArg is an interned pointer, with the lowest 2 bits being reserved for a tag to indicate the type (Ty, Region, or Const) it points to.

Note: the PartialEq, Eq and Hash derives are only valid because Ty, Region and Const are all interned.

Fields
ptr: NonNull<()>marker: PhantomData<(Ty<'tcx>, Region<'tcx>, Const<'tcx>)>
Implementations

impl<'tcx> GenericArg<'tcx>

pub fn walk(self) -> TypeWalker<'tcx>

Iterator that walks self and any types reachable from self, in depth-first order. Note that just walks the types that appear in self, it does not descend into the fields of structs or variants. For example:

isize => { isize }
Foo<Bar<isize>> => { Foo<Bar<isize>>, Bar<isize>, isize }
[isize] => { [isize], isize }

pub fn walk_shallow( self, visited: &mut SsoHashSet<GenericArg<'tcx>> ) -> impl Iterator<Item = GenericArg<'tcx>>

Iterator that walks the immediate children of self. Hence Foo<Bar<i32>, u32> yields the sequence [Bar<i32>, u32] (but not i32, like walk).

Iterator only walks items once. It accepts visited set, updates it with all visited types and skips any types that are already there.

impl<'tcx> GenericArg<'tcx>

pub fn unpack(self) -> GenericArgKind<'tcx>

pub fn as_type(self) -> Option<Ty<'tcx>>

pub fn as_region(self) -> Option<Region<'tcx>>

pub fn as_const(self) -> Option<Const<'tcx>>

pub fn expect_region(self) -> Region<'tcx>

Unpack the GenericArg as a region when it is known certainly to be a region.

pub fn expect_ty(self) -> Ty<'tcx>

Unpack the GenericArg as a type when it is known certainly to be a type. This is true in cases where GenericArgs is used in places where the kinds are known to be limited (e.g. in tuples, where the only parameters are type parameters).

pub fn expect_const(self) -> Const<'tcx>

Unpack the GenericArg as a const when it is known certainly to be a const.

pub fn is_non_region_infer(self) -> bool