rustc_middle::ty
Module inherent

Set of traits which are used to emulate the inherent impls that are present in rustc_middle. It is customary to glob-import rustc_type_ir::inherent::* to bring all of these traits into scope when programming in interner-agnostic settings, and to avoid importing any of these directly elsewhere (i.e. specify the full path for an implementation downstream).

Traits

Abi

AdtDef

BoundExistentialPredicates

BoundVarLike

Clause

Const

DefId

ExprConst

Features

GenericArg

GenericArgs

GenericsOf

IntoKind

ParamEnv

ParamLike

PlaceholderLike

Common capabilities of placeholder kinds

Predicate

Region

Safety

SliceLike

Span

Term

Ty

Tys