rustc_middle::ty
Module list

Macros

impl_list_empty 🔒

Structs

ListSkeleton 🔒

A RawList without the unsized tail. This type is used for layout computation and constructing empty lists.

RawList

A generic type that can be used to prepend a List with some header.

TypeInfo

The additional info that is stored in ListWithCachedTypeInfo.

Type Aliases

List

List<T> is a bit like &[T], but with some critical differences.

ListWithCachedTypeInfo

A List that additionally stores type information inline to speed up TypeVisitableExt operations.

Foreign Types

OpaqueListContents 🔒

A dummy type used to force List to be unsized while not requiring references to it be wide pointers.