rustc_middle::ty::list

Type Alias List

pub type List<T> = RawList<(), T>;

List<T> is a bit like &[T], but with some critical differences.

IMPORTANT: Every List<T> is required to have unique contents. The type’s correctness relies on this, but it does not enforce it. Therefore, any code that creates a List<T> must ensure uniqueness itself. In practice this is achieved by interning.

The length is stored within the List<T>, so &List<Ty> is a thin pointer.

Because of this, you cannot get a List<T> that is a sub-list of another List<T>. You can get a sub-slice &[T], however.

List<T> can be used with TaggedRef, which is useful within structs whose size must be minimized.

Because of the uniqueness assumption, we can use the address of a List<T> for faster equality comparisons and hashing.

T must be Copy. This lets List<T> be stored in a dropless arena and iterators return a T rather than a &T.

T must not be zero-sized.

Aliased Type

#[repr(C)]
pub struct List<T> {
    skel: ListSkeleton<(), T>,
    opaque: OpaqueListContents,
}

Fields
skel: ListSkeleton<(), T>opaque: OpaqueListContents
Implementations

impl<'tcx> List<PolyExistentialPredicate<'tcx>>

pub fn principal(&self) -> Option<Binder<'tcx, ExistentialTraitRef<'tcx>>>

Returns the “principal DefId” of this set of existential predicates.

A Rust trait object type consists (in addition to a lifetime bound) of a set of trait bounds, which are separated into any number of auto-trait bounds, and at most one non-auto-trait bound. The non-auto-trait bound is called the “principal” of the trait object.

Only the principal can have methods or type parameters (because auto traits can have neither of them). This is important, because it means the auto traits can be treated as an unordered set (methods would force an order for the vtable, while relating traits with type parameters without knowing the order to relate them in is a rather non-trivial task).

For example, in the trait object dyn std::fmt::Debug + Sync, the principal bound is Some(std::fmt::Debug), while the auto-trait bounds are the set {Sync}.

It is also possible to have a “trivial” trait object that consists only of auto traits, with no principal - for example, dyn Send + Sync. In that case, the set of auto-trait bounds is {Send, Sync}, while there is no principal. These trait objects have a “trivial” vtable consisting of just the size, alignment, and destructor.

pub fn principal_def_id(&self) -> Option<DefId>

pub fn projection_bounds( &self, ) -> impl Iterator<Item = Binder<'tcx, ExistentialProjection<'tcx>>>

pub fn auto_traits(&self) -> impl Iterator<Item = DefId>

pub fn without_auto_traits( &self, ) -> impl Iterator<Item = PolyExistentialPredicate<'tcx>>