rustc_middle::ty

Module normalize_erasing_regions

Methods for normalizing when you don’t care about regions (and aren’t doing type inference). If either of those things don’t apply to you, use infcx.normalize(...).

The methods in this file use a TypeFolder to recursively process contents, invoking the underlying normalize_generic_arg_after_erasing_regions query for each type or constant found within. (This underlying query is what is cached.)

Structs

NormalizeAfterErasingRegionsFolder 🔒

TryNormalizeAfterErasingRegionsFolder 🔒

Enums

NormalizationError