Struct rustc_middle::ty::opaque_types::ReverseMapper

pub(super) struct ReverseMapper<'tcx> {
    tcx: TyCtxt<'tcx>,
    map: FxHashMap<GenericArg<'tcx>, GenericArg<'tcx>>,
    do_not_error: bool,
    ignore_errors: bool,
    span: Span,
}

Converts generic params of a TypeFoldable from one item’s generics to another. Usually from a function’s generics list to the opaque type’s own generics.

Fields
tcx: TyCtxt<'tcx>map: FxHashMap<GenericArg<'tcx>, GenericArg<'tcx>>do_not_error: bool

see call sites to fold_kind_no_missing_regions_error for an explanation of this field.

ignore_errors: bool

We do not want to emit any errors in typeck because the spans in typeck are subpar at the moment. Borrowck will do the same work again (this time with lifetime information) and thus report better errors.

span: Span

Span of function being checked.

Implementations

impl<'tcx> ReverseMapper<'tcx>

pub(super) fn new( tcx: TyCtxt<'tcx>, map: FxHashMap<GenericArg<'tcx>, GenericArg<'tcx>>, span: Span, ignore_errors: bool ) -> Self

fn fold_kind_no_missing_regions_error( &mut self, kind: GenericArg<'tcx> ) -> GenericArg<'tcx>

fn fold_kind_normally(&mut self, kind: GenericArg<'tcx>) -> GenericArg<'tcx>

fn fold_closure_args( &mut self, def_id: DefId, args: GenericArgsRef<'tcx> ) -> GenericArgsRef<'tcx>