rustc_middle::ty
Module outlives

The outlives relation T: 'a or 'a: 'b. This code frequently refers to rules defined in RFC 1214 (OutlivesFooBar), so see that RFC for reference.

Enums

Component

Functions

compute_alias_components_recursive

Collect Components for all the args of parent.

push_outlives_components

Push onto out all the things that must outlive 'a for the condition ty0: 'a to hold. Note that ty0 must be a fully resolved type.