rustc_middle::ty

Module predicate

Modules

size_asserts 🔒

Structs

Clause

A subset of predicates which can be assumed by the trait solver. They show up in an item’s where clauses, hence the name Clause, and may either be user-written (such as traits) or may be inserted during lowering.

Predicate

A statement that can be proven by a trait solver. This includes things that may show up in where clauses, such as trait predicates and projection predicates, and also things that are emitted as part of type checking such as DynCompatible predicate which is emitted when a type is coerced to a trait object.

Traits

ExistentialPredicateStableCmpExt

Type Aliases

AliasTerm

ArgOutlivesPredicate

ClauseKind

CoercePredicate

ExistentialPredicate

ExistentialProjection

ExistentialTraitRef

HostEffectPredicate

NormalizesTo

OutlivesPredicate

PolyCoercePredicate

PolyExistentialPredicate

PolyExistentialProjection

PolyExistentialTraitRef

PolyProjectionPredicate

PolyRegionOutlivesPredicate

PolySubtypePredicate

PolyTraitPredicate

PolyTraitRef

PolyTypeOutlivesPredicate

PredicateKind

ProjectionPredicate

RegionOutlivesPredicate

SubtypePredicate

TraitPredicate

TraitRef

TypeOutlivesPredicate