rustc_middle::ty::predicate
Type Alias ClauseKind

pub type ClauseKind<'tcx> = ClauseKind<TyCtxt<'tcx>>;

Aliased Type

enum ClauseKind<'tcx> {
    Trait(TraitPredicate<TyCtxt<'tcx>>),
    RegionOutlives(OutlivesPredicate<TyCtxt<'tcx>, Region<'tcx>>),
    TypeOutlives(OutlivesPredicate<TyCtxt<'tcx>, Ty<'tcx>>),
    Projection(ProjectionPredicate<TyCtxt<'tcx>>),
    ConstArgHasType(Const<'tcx>, Ty<'tcx>),
    WellFormed(GenericArg<'tcx>),
    ConstEvaluatable(Const<'tcx>),
}

Variants

Trait(TraitPredicate<TyCtxt<'tcx>>)

Corresponds to where Foo: Bar<A, B, C>. Foo here would be the Self type of the trait reference and A, B, and C would be the type parameters.

RegionOutlives(OutlivesPredicate<TyCtxt<'tcx>, Region<'tcx>>)

where 'a: 'r

TypeOutlives(OutlivesPredicate<TyCtxt<'tcx>, Ty<'tcx>>)

where T: 'r

Projection(ProjectionPredicate<TyCtxt<'tcx>>)

where <T as TraitRef>::Name == X, approximately. See the ProjectionPredicate struct for details.

ConstArgHasType(Const<'tcx>, Ty<'tcx>)

Ensures that a const generic argument to a parameter const N: u8 is of type u8.

WellFormed(GenericArg<'tcx>)

No syntax: T well-formed.

ConstEvaluatable(Const<'tcx>)

Constant initializer must evaluate successfully.