rustc_middle::ty::predicate

Type Alias TraitPredicate

pub type TraitPredicate<'tcx> = TraitPredicate<TyCtxt<'tcx>>;

Aliased Type

pub struct TraitPredicate<'tcx> {
    pub trait_ref: TraitRef<TyCtxt<'tcx>>,
    pub polarity: PredicatePolarity,
}

Fields
trait_ref: TraitRef<TyCtxt<'tcx>>polarity: PredicatePolarity

If polarity is Positive: we are proving that the trait is implemented.

If polarity is Negative: we are proving that a negative impl of this trait exists. (Note that coherence also checks whether negative impls of supertraits exist via a series of predicates.)