Struct rustc_middle::ty::region::Region

pub struct Region<'tcx>(pub Interned<'tcx, RegionKind<'tcx>>);

Use this rather than RegionKind, whenever possible.

Tuple Fields
0: Interned<'tcx, RegionKind<'tcx>>
Implementations

impl<'tcx> Region<'tcx>

pub fn new_early_param( tcx: TyCtxt<'tcx>, early_bound_region: EarlyParamRegion ) -> Region<'tcx>

pub fn new_bound( tcx: TyCtxt<'tcx>, debruijn: DebruijnIndex, bound_region: BoundRegion ) -> Region<'tcx>

pub fn new_late_param( tcx: TyCtxt<'tcx>, scope: DefId, bound_region: BoundRegionKind ) -> Region<'tcx>

pub fn new_var(tcx: TyCtxt<'tcx>, v: RegionVid) -> Region<'tcx>

pub fn new_placeholder( tcx: TyCtxt<'tcx>, placeholder: PlaceholderRegion ) -> Region<'tcx>

pub fn new_error(tcx: TyCtxt<'tcx>, guar: ErrorGuaranteed) -> Region<'tcx>

Constructs a RegionKind::ReError region.

pub fn new_error_misc(tcx: TyCtxt<'tcx>) -> Region<'tcx>

Constructs a RegionKind::ReError region and registers a delayed bug to ensure it gets used.

pub fn new_error_with_message<S: Into<MultiSpan>>( tcx: TyCtxt<'tcx>, span: S, msg: &'static str ) -> Region<'tcx>

Constructs a RegionKind::ReError region and registers a delayed bug with the given msg to ensure it gets used.

pub fn new_from_kind(tcx: TyCtxt<'tcx>, kind: RegionKind<'tcx>) -> Region<'tcx>

Avoid this in favour of more specific new_* methods, where possible, to avoid the cost of the match.

impl<'tcx> Region<'tcx>

Region utilities

pub fn kind(self) -> RegionKind<'tcx>

pub fn get_name(self) -> Option<Symbol>

pub fn get_name_or_anon(self) -> Symbol

pub fn has_name(self) -> bool

Is this region named by the user?

pub fn is_error(self) -> bool

pub fn is_static(self) -> bool

pub fn is_erased(self) -> bool

pub fn is_bound(self) -> bool

pub fn is_placeholder(self) -> bool

pub fn bound_at_or_above_binder(self, index: DebruijnIndex) -> bool

pub fn type_flags(self) -> TypeFlags

pub fn free_region_binding_scope(self, tcx: TyCtxt<'_>) -> DefId

Given an early-bound or free region, returns the DefId where it was bound. For example, consider the regions in this snippet of code:

impl<'a> Foo {
//   ^^ -- early bound, declared on an impl

    fn bar<'b, 'c>(x: &self, y: &'b u32, z: &'c u64) where 'static: 'c
//         ^^  ^^     ^ anonymous, late-bound
//         |   early-bound, appears in where-clauses
//         late-bound, appears only in fn args
    {..}
}

Here, free_region_binding_scope('a) would return the DefId of the impl, and for all the other highlighted regions, it would return the DefId of the function. In other cases (not shown), this function might return the DefId of a closure.

pub fn is_param(self) -> bool

True for free regions other than 'static.

pub fn is_free(self) -> bool

True for free region in the current context.

This is the case for 'static and param regions.

pub fn is_var(self) -> bool

pub fn as_var(self) -> RegionVid