Module rustc_middle::ty::relate

Generalized type relating mechanism.

A type relation R relates a pair of values (A, B). A and B are usually types or regions but can be other things. Examples of type relations are subtyping, type equality, etc.

Structs

GeneratorWitness 🔒

Enums

Cause

Traits

Relate

TypeRelation

Functions

expected_found

relate_args

relate_args_with_variances

relate_type_and_mut

structurally_relate_consts

Relates a and b structurally, calling the relation for all nested values. Any semantic equality, e.g. of unevaluated consts, and inference variables have to be handled by the caller.

structurally_relate_tys

Relates a and b structurally, calling the relation for all nested values. Any semantic equality, e.g. of projections, and inference variables have to be handled by the caller.

Type Aliases

RelateResult