rustc_middle::ty
Module search_graph

Structs

GlobalCache

SearchGraph

StackDepth

Enums

PathKind

In the initial iteration of a cycle, we do not yet have a provisional result. In the case we return an initial provisional result depending on the kind of cycle.

UsageKind

Traits

Cx

The search graph does not simply use Interner directly to enable its fuzzing without having to stub the rest of the interner. We don’t make this a super trait of Interner as users of the shared type library shouldn’t have to care about Input and Result as they are implementation details of the search graph.

Delegate