rustc_middle::ty

Module significant_drop_order

Functions

extract_component_raw

Returns the list of types with a “potentially significant” that may be dropped by dropping a value of type ty.

extract_component_with_significant_dtor

true_significant_drop_ty 🔒

An additional filter to exclude well-known types from the ecosystem because their drops are trivial. This returns additional types to check if the drops are delegated to those. A typical example is hashbrown::HashMap<K, V>, whose drop is delegated to K and V.

ty_dtor_span

Extract the span of the custom destructor of a type especially the span of the impl Drop header or its entire block when we are working with current local crate.