Struct rustc_middle::ty::Const

pub struct Const<'tcx>(pub(super) Interned<'tcx, WithCachedTypeInfo<ConstData<'tcx>>>);

Use this rather than ConstData, whenever possible.

Tuple Fields
0: Interned<'tcx, WithCachedTypeInfo<ConstData<'tcx>>>
Implementations

impl<'tcx> Const<'tcx>

pub fn walk(self) -> TypeWalker<'tcx>

Iterator that walks self and any types reachable from self, in depth-first order. Note that just walks the types that appear in self, it does not descend into the fields of structs or variants. For example:

isize => { isize }
Foo<Bar<isize>> => { Foo<Bar<isize>>, Bar<isize>, isize }
[isize] => { [isize], isize }

impl<'tcx> Const<'tcx>

pub fn ty(self) -> Ty<'tcx>

pub fn kind(self) -> ConstKind<'tcx>

pub fn flags(self) -> TypeFlags

pub fn outer_exclusive_binder(self) -> DebruijnIndex

pub fn new( tcx: TyCtxt<'tcx>, kind: ConstKind<'tcx>, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_param( tcx: TyCtxt<'tcx>, param: ParamConst, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_var(tcx: TyCtxt<'tcx>, infer: ConstVid, ty: Ty<'tcx>) -> Const<'tcx>

pub fn new_fresh(tcx: TyCtxt<'tcx>, fresh: u32, ty: Ty<'tcx>) -> Const<'tcx>

pub fn new_infer( tcx: TyCtxt<'tcx>, infer: InferConst, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_bound( tcx: TyCtxt<'tcx>, debruijn: DebruijnIndex, var: BoundVar, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_placeholder( tcx: TyCtxt<'tcx>, placeholder: PlaceholderConst, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_unevaluated( tcx: TyCtxt<'tcx>, uv: UnevaluatedConst<'tcx>, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_value( tcx: TyCtxt<'tcx>, val: ValTree<'tcx>, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_expr( tcx: TyCtxt<'tcx>, expr: Expr<'tcx>, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_error( tcx: TyCtxt<'tcx>, e: ErrorGuaranteed, ty: Ty<'tcx> ) -> Const<'tcx>

pub fn new_misc_error(tcx: TyCtxt<'tcx>, ty: Ty<'tcx>) -> Const<'tcx>

Like Ty::new_error but for constants.

pub fn new_error_with_message<S: Into<MultiSpan>>( tcx: TyCtxt<'tcx>, ty: Ty<'tcx>, span: S, msg: &'static str ) -> Const<'tcx>

Like Ty::new_error_with_message but for constants.

impl<'tcx> Const<'tcx>

pub fn from_anon_const(tcx: TyCtxt<'tcx>, def: LocalDefId) -> Self

Literals and const generic parameters are eagerly converted to a constant, everything else becomes Unevaluated.

fn try_from_lit_or_param( tcx: TyCtxt<'tcx>, ty: Ty<'tcx>, expr: &'tcx Expr<'tcx> ) -> Option<Self>

pub fn from_bits( tcx: TyCtxt<'tcx>, bits: u128, ty: ParamEnvAnd<'tcx, Ty<'tcx>> ) -> Self

Creates a constant with the given integer value and interns it.

pub fn zero_sized(tcx: TyCtxt<'tcx>, ty: Ty<'tcx>) -> Self

Creates an interned zst constant.

pub fn from_bool(tcx: TyCtxt<'tcx>, v: bool) -> Self

Creates an interned bool constant.

pub fn from_target_usize(tcx: TyCtxt<'tcx>, n: u64) -> Self

Creates an interned usize constant.

pub fn eval( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, span: Option<Span> ) -> Result<ValTree<'tcx>, ErrorHandled>

Returns the evaluated constant

pub fn normalize(self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>) -> Self

Normalizes the constant to a value or an error if possible.

pub fn try_eval_scalar( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> Option<Scalar>

pub fn try_eval_scalar_int( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> Option<ScalarInt>

Attempts to evaluate the given constant to bits. Can fail to evaluate in the presence of generics (or erroneous code) or if the value can’t be represented as bits (e.g. because it contains const generic parameters or pointers).

pub fn try_eval_bits( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> Option<u128>

Attempts to evaluate the given constant to bits. Can fail to evaluate in the presence of generics (or erroneous code) or if the value can’t be represented as bits (e.g. because it contains const generic parameters or pointers).

pub fn eval_bits(self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>) -> u128

Panics if the value cannot be evaluated or doesn’t contain a valid integer of the given type.

pub fn try_eval_target_usize( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> Option<u64>

pub fn try_eval_bool( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> Option<bool>

pub fn eval_target_usize( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx> ) -> u64

Panics if the value cannot be evaluated or doesn’t contain a valid usize.

pub fn to_valtree(self) -> ValTree<'tcx>

Panics if self.kind != ty::ConstKind::Value

pub fn try_to_valtree(self) -> Option<ValTree<'tcx>>

Attempts to convert to a ValTree

pub fn try_to_scalar(self) -> Option<Scalar>

pub fn try_to_target_usize(self, tcx: TyCtxt<'tcx>) -> Option<u64>

pub fn is_ct_infer(self) -> bool