Struct rustc_middle::ty::GeneratorArgs

pub struct GeneratorArgs<'tcx> {
    pub args: GenericArgsRef<'tcx>,
}

Similar to ClosureArgs; see the above documentation for more.

Fields
args: GenericArgsRef<'tcx>
Implementations

impl<'tcx> GeneratorArgs<'tcx>

pub fn new( tcx: TyCtxt<'tcx>, parts: GeneratorArgsParts<'tcx, Ty<'tcx>> ) -> GeneratorArgs<'tcx>

Construct GeneratorArgs from GeneratorArgsParts, containing Args for the generator parent, alongside additional generator-specific components.

fn split(self) -> GeneratorArgsParts<'tcx, GenericArg<'tcx>>

Divides the generator args into their respective components. The ordering assumed here must match that used by GeneratorArgs::new above.

pub fn is_valid(self) -> bool

Returns true only if enough of the synthetic types are known to allow using all of the methods on GeneratorArgs without panicking.

Used primarily by ty::print::pretty to be able to handle generator types that haven’t had their synthetic types substituted in.

pub fn parent_args(self) -> &'tcx [GenericArg<'tcx>]

Returns the substitutions of the generator’s parent.

pub fn witness(self) -> Ty<'tcx>

This describes the types that can be contained in a generator. It will be a type variable initially and unified in the last stages of typeck of a body. It contains a tuple of all the types that could end up on a generator frame. The state transformation MIR pass may only produce layouts which mention types in this tuple. Upvars are not counted here.

pub fn upvar_tys(self) -> &'tcx List<Ty<'tcx>>

Returns an iterator over the list of types of captured paths by the generator. In case there was a type error in figuring out the types of the captured path, an empty iterator is returned.

pub fn tupled_upvars_ty(self) -> Ty<'tcx>

Returns the tuple type representing the upvars for this generator.

pub fn resume_ty(self) -> Ty<'tcx>

Returns the type representing the resume type of the generator.

pub fn yield_ty(self) -> Ty<'tcx>

Returns the type representing the yield type of the generator.

pub fn return_ty(self) -> Ty<'tcx>

Returns the type representing the return type of the generator.

pub fn poly_sig(self) -> PolyGenSig<'tcx>

Returns the “generator signature”, which consists of its yield and return types.

N.B., some bits of the code prefers to see this wrapped in a binder, but it never contains bound regions. Probably this function should be removed.

pub fn sig(self) -> GenSig<'tcx>

Returns the “generator signature”, which consists of its resume, yield and return types.

impl<'tcx> GeneratorArgs<'tcx>

pub const UNRESUMED: usize = 0usize

Generator has not been resumed yet.

pub const RETURNED: usize = 1usize

Generator has returned or is completed.

pub const POISONED: usize = 2usize

Generator has been poisoned.

const UNRESUMED_NAME: &'static str = "Unresumed"

const RETURNED_NAME: &'static str = "Returned"

const POISONED_NAME: &'static str = "Panicked"

pub fn variant_range( &self, def_id: DefId, tcx: TyCtxt<'tcx> ) -> Range<VariantIdx>

The valid variant indices of this generator.

pub fn discriminant_for_variant( &self, def_id: DefId, tcx: TyCtxt<'tcx>, variant_index: VariantIdx ) -> Discr<'tcx>

The discriminant for the given variant. Panics if the variant_index is out of range.

pub fn discriminants( self, def_id: DefId, tcx: TyCtxt<'tcx> ) -> impl Iterator<Item = (VariantIdx, Discr<'tcx>)> + Captures<'tcx>

The set of all discriminants for the generator, enumerated with their variant indices.

pub fn variant_name(v: VariantIdx) -> Cow<'static, str>

Calls f with a reference to the name of the enumerator for the given variant v.

pub fn discr_ty(&self, tcx: TyCtxt<'tcx>) -> Ty<'tcx>

The type of the state discriminant used in the generator type.

pub fn state_tys( self, def_id: DefId, tcx: TyCtxt<'tcx> ) -> impl Iterator<Item = impl Iterator<Item = Ty<'tcx>> + Captures<'tcx>>

This returns the types of the MIR locals which had to be stored across suspension points. It is calculated in rustc_mir_transform::generator::StateTransform. All the types here must be in the tuple in GeneratorInterior.

The locals are grouped by their variant number. Note that some locals may be repeated in multiple variants.

pub fn prefix_tys(self) -> &'tcx List<Ty<'tcx>>

This is the types of the fields of a generator which are not stored in a variant.