Struct rustc_middle::ty::List

#[repr(C)]
pub struct List<T> {
    len: usize,
    data: [T; 0],
    opaque: OpaqueListContents,
}

List<T> is a bit like &[T], but with some critical differences.

IMPORTANT: Every List<T> is required to have unique contents. The type’s correctness relies on this, but it does not enforce it. Therefore, any code that creates a List<T> must ensure uniqueness itself. In practice this is achieved by interning.

The length is stored within the List<T>, so &List<Ty> is a thin pointer.

Because of this, you cannot get a List<T> that is a sub-list of another List<T>. You can get a sub-slice &[T], however.

List<T> can be used with CopyTaggedPtr, which is useful within structs whose size must be minimized.

Because of the uniqueness assumption, we can use the address of a List<T> for faster equality comparisons and hashing.

T must be Copy. This lets List<T> be stored in a dropless arena and iterators return a T rather than a &T.

T must not be zero-sized.

Fields
len: usizedata: [T; 0]

Although this claims to be a zero-length array, in practice len elements are actually present.

opaque: OpaqueListContents
Implementations

impl<'tcx> List<Ty<'tcx>>

pub fn as_substs(&'tcx self) -> SubstsRef<'tcx>

Allows to freely switch between List<Ty<'tcx>> and List<GenericArg<'tcx>>.

As lists are interned, List<Ty<'tcx>> and List<GenericArg<'tcx>> have be interned together, see intern_type_list for more details.

impl<'tcx> List<GenericArg<'tcx>>

pub fn try_as_type_list(&'tcx self) -> Option<&'tcx List<Ty<'tcx>>>

Checks whether all elements of this list are types, if so, transmute.

pub fn as_closure(&'tcx self) -> ClosureSubsts<'tcx>

Interpret these substitutions as the substitutions of a closure type. Closure substitutions have a particular structure controlled by the compiler that encodes information like the signature and closure kind; see ty::ClosureSubsts struct for more comments.

pub fn as_generator(&'tcx self) -> GeneratorSubsts<'tcx>

Interpret these substitutions as the substitutions of a generator type. Generator substitutions have a particular structure controlled by the compiler that encodes information like the signature and generator kind; see ty::GeneratorSubsts struct for more comments.

pub fn as_inline_const(&'tcx self) -> InlineConstSubsts<'tcx>

Interpret these substitutions as the substitutions of an inline const. Inline const substitutions have a particular structure controlled by the compiler that encodes information like the inferred type; see ty::InlineConstSubsts struct for more comments.

pub fn identity_for_item(tcx: TyCtxt<'tcx>, def_id: DefId) -> SubstsRef<'tcx>

Creates an InternalSubsts that maps each generic parameter to itself.

pub fn for_item<F>(
tcx: TyCtxt<'tcx>,
def_id: DefId,
mk_kind: F
) -> SubstsRef<'tcx>where
F: FnMut(&GenericParamDef, &[GenericArg<'tcx>]) -> GenericArg<'tcx>,

Creates an InternalSubsts for generic parameter definitions, by calling closures to obtain each kind. The closures get to observe the InternalSubsts as they’re being built, which can be used to correctly substitute defaults of generic parameters.

pub fn extend_to<F>(
&self,
tcx: TyCtxt<'tcx>,
def_id: DefId,
mk_kind: F
) -> SubstsRef<'tcx>where
F: FnMut(&GenericParamDef, &[GenericArg<'tcx>]) -> GenericArg<'tcx>,

pub fn fill_item<F>(
substs: &mut SmallVec<[GenericArg<'tcx>; 8]>,
tcx: TyCtxt<'tcx>,
defs: &Generics,
mk_kind: &mut F
)where
F: FnMut(&GenericParamDef, &[GenericArg<'tcx>]) -> GenericArg<'tcx>,

pub fn fill_single<F>(
substs: &mut SmallVec<[GenericArg<'tcx>; 8]>,
defs: &Generics,
mk_kind: &mut F
)where
F: FnMut(&GenericParamDef, &[GenericArg<'tcx>]) -> GenericArg<'tcx>,

pub fn types(&'tcx self) -> impl DoubleEndedIterator<Item = Ty<'tcx>> + 'tcx

pub fn regions(
&'tcx self
) -> impl DoubleEndedIterator<Item = Region<'tcx>> + 'tcx

pub fn consts(&'tcx self) -> impl DoubleEndedIterator<Item = Const<'tcx>> + 'tcx

pub fn non_erasable_generics(
&'tcx self
) -> impl DoubleEndedIterator<Item = GenericArgKind<'tcx>> + 'tcx

pub fn type_at(&self, i: usize) -> Ty<'tcx>

pub fn region_at(&self, i: usize) -> Region<'tcx>

pub fn const_at(&self, i: usize) -> Const<'tcx>

pub fn type_for_def(&self, def: &GenericParamDef) -> GenericArg<'tcx>

pub fn rebase_onto(
&self,
tcx: TyCtxt<'tcx>,
source_ancestor: DefId,
target_substs: SubstsRef<'tcx>
) -> SubstsRef<'tcx>

Transform from substitutions for a child of source_ancestor (e.g., a trait or impl) to substitutions for the same child in a different item, with target_substs as the base for the target impl/trait, with the source child-specific parameters (e.g., method parameters) on top of that base.

For example given:

trait X<S> { fn f<T>(); }
impl<U> X<U> for U { fn f<V>() {} }

If self is [Self, S, T]: the identity substs of f in the trait.

If source_ancestor is the def_id of the trait.

If target_substs is [U], the substs for the impl.

Then we will return [U, T], the subst for f in the impl that are needed for it to match the trait.

pub fn truncate_to(
&self,
tcx: TyCtxt<'tcx>,
generics: &Generics
) -> SubstsRef<'tcx>

impl<T> List<T>

pub fn empty<'a>() -> &'a List<T>

Returns a reference to the (unique, static) empty list.

pub fn len(&self) -> usize

pub fn as_slice(&self) -> &[T]

impl<T: Copy> List<T>

pub(super) fn from_arena<'tcx>(
arena: &'tcx Arena<'tcx>,
slice: &[T]
) -> &'tcx List<T>

Allocates a list from arena and copies the contents of slice into it.

WARNING: the contents must be unique, such that no list with these contents has been previously created. If not, operations such as eq and hash might give incorrect results.

Panics if T is Drop, or T is zero-sized, or the slice is empty (because the empty list exists statically, and is available via empty()).

pub fn iter(&self) -> Copied<Iter<'_, T>>

impl<'tcx> List<Binder<'tcx, ExistentialPredicate<'tcx>>>

pub fn principal(&self) -> Option<Binder<'tcx, ExistentialTraitRef<'tcx>>>

Returns the “principal DefId” of this set of existential predicates.

A Rust trait object type consists (in addition to a lifetime bound) of a set of trait bounds, which are separated into any number of auto-trait bounds, and at most one non-auto-trait bound. The non-auto-trait bound is called the “principal” of the trait object.

Only the principal can have methods or type parameters (because auto traits can have neither of them). This is important, because it means the auto traits can be treated as an unordered set (methods would force an order for the vtable, while relating traits with type parameters without knowing the order to relate them in is a rather non-trivial task).

For example, in the trait object dyn fmt::Debug + Sync, the principal bound is Some(fmt::Debug), while the auto-trait bounds are the set {Sync}.

It is also possible to have a “trivial” trait object that consists only of auto traits, with no principal - for example, dyn Send + Sync. In that case, the set of auto-trait bounds is {Send, Sync}, while there is no principal. These trait objects have a “trivial” vtable consisting of just the size, alignment, and destructor.

pub fn principal_def_id(&self) -> Option<DefId>

pub fn projection_bounds<'a>(
&'a self
) -> impl Iterator<Item = Binder<'tcx, ExistentialProjection<'tcx>>> + 'a

pub fn auto_traits<'a>(
&'a self
) -> impl Iterator<Item = DefId> + Captures<'tcx> + 'a