Struct rustc_middle::ty::Region

pub struct Region<'tcx>(pub Interned<'tcx, RegionKind<'tcx>>);

Use this rather than RegionKind, whenever possible.

Tuple Fields
0: Interned<'tcx, RegionKind<'tcx>>
Implementations

impl<'tcx> Region<'tcx>

Region utilities

pub fn kind(self) -> RegionKind<'tcx>

pub fn get_name(self) -> Option<Symbol>

pub fn has_name(self) -> bool

Is this region named by the user?

pub fn is_static(self) -> bool

pub fn is_erased(self) -> bool

pub fn is_late_bound(self) -> bool

pub fn is_placeholder(self) -> bool

pub fn bound_at_or_above_binder(self, index: DebruijnIndex) -> bool

pub fn type_flags(self) -> TypeFlags

pub fn free_region_binding_scope(self, tcx: TyCtxt<'_>) -> DefId

Given an early-bound or free region, returns the DefId where it was bound. For example, consider the regions in this snippet of code:

impl<'a> Foo {
//   ^^ -- early bound, declared on an impl

    fn bar<'b, 'c>(x: &self, y: &'b u32, z: &'c u64) where 'static: 'c
//         ^^  ^^     ^ anonymous, late-bound
//         |   early-bound, appears in where-clauses
//         late-bound, appears only in fn args
    {..}
}

Here, free_region_binding_scope('a) would return the DefId of the impl, and for all the other highlighted regions, it would return the DefId of the function. In other cases (not shown), this function might return the DefId of a closure.

pub fn is_free(self) -> bool

True for free regions other than 'static.

pub fn is_free_or_static(self) -> bool

True if self is a free region or static.

pub fn is_var(self) -> bool