Struct rustc_middle::ty::ResolverGlobalCtxt

pub struct ResolverGlobalCtxt {

    pub visibilities: FxHashMap<LocalDefId, Visibility>,
    pub has_pub_restricted: bool,
    pub expn_that_defined: FxHashMap<LocalDefId, ExpnId>,
    pub effective_visibilities: EffectiveVisibilities,
    pub extern_crate_map: FxHashMap<LocalDefId, CrateNum>,
    pub maybe_unused_trait_imports: FxIndexSet<LocalDefId>,
    pub maybe_unused_extern_crates: Vec<(LocalDefId, Span)>,
    pub reexport_map: FxHashMap<LocalDefId, Vec<ModChild>>,
    pub glob_map: FxHashMap<LocalDefId, FxHashSet<Symbol>>,
    pub extern_prelude: FxHashMap<Symbol, bool>,
    pub main_def: Option<MainDefinition>,
    pub trait_impls: FxIndexMap<DefId, Vec<LocalDefId>>,
    pub proc_macros: Vec<LocalDefId>,
    pub confused_type_with_std_module: FxHashMap<Span, Span>,
    pub registered_tools: RegisteredTools,

}

Fields
visibilities: FxHashMap<LocalDefId, Visibility>has_pub_restricted: bool

This field is used to decide whether we should make PRIVATE_IN_PUBLIC a hard error.

expn_that_defined: FxHashMap<LocalDefId, ExpnId>

Item with a given LocalDefId was defined during macro expansion with ID ExpnId.

effective_visibilities: EffectiveVisibilitiesextern_crate_map: FxHashMap<LocalDefId, CrateNum>maybe_unused_trait_imports: FxIndexSet<LocalDefId>maybe_unused_extern_crates: Vec<(LocalDefId, Span)>reexport_map: FxHashMap<LocalDefId, Vec<ModChild>>glob_map: FxHashMap<LocalDefId, FxHashSet<Symbol>>extern_prelude: FxHashMap<Symbol, bool>

Extern prelude entries. The value is true if the entry was introduced via extern crate item and not --extern option or compiler built-in.

main_def: Option<MainDefinition>trait_impls: FxIndexMap<DefId, Vec<LocalDefId>>proc_macros: Vec<LocalDefId>

A list of proc macro LocalDefIds, written out in the order in which they are declared in the static array generated by proc_macro_harness.

confused_type_with_std_module: FxHashMap<Span, Span>

Mapping from ident span to path span for paths that don’t exist as written, but that exist under std. For example, wrote str::from_utf8 instead of std::str::from_utf8.

registered_tools: RegisteredTools