Struct rustc_middle::ty::Ty

pub struct Ty<'tcx>(Interned<'tcx, WithStableHash<TyS<'tcx>>>);

Use this rather than TyS, whenever possible.

Tuple Fields
0: Interned<'tcx, WithStableHash<TyS<'tcx>>>
Implementations

impl<'tcx> Ty<'tcx>

pub fn sort_string(self, tcx: TyCtxt<'_>) -> Cow<'static, str>

pub fn prefix_string(self, tcx: TyCtxt<'_>) -> Cow<'static, str>

impl<'tcx> Ty<'tcx>

fn uninhabited_from(
self,
tcx: TyCtxt<'tcx>,
param_env: ParamEnv<'tcx>
) -> DefIdForest<'tcx>

Calculates the forest of DefIds from which this type is visibly uninhabited.

impl<'tcx> Ty<'tcx>

pub fn numeric_max_val(self, tcx: TyCtxt<'tcx>) -> Option<Const<'tcx>>

Returns the maximum value for the given numeric type (including chars) or returns None if the type is not numeric.

pub fn numeric_min_val(self, tcx: TyCtxt<'tcx>) -> Option<Const<'tcx>>

Returns the minimum value for the given numeric type (including chars) or returns None if the type is not numeric.

pub fn is_copy_modulo_regions(
self,
tcx_at: TyCtxtAt<'tcx>,
param_env: ParamEnv<'tcx>
) -> bool

Checks whether values of this type T are moved or copied when referenced – this amounts to a check for whether T: Copy, but note that we don’t consider lifetimes when doing this check. This means that we may generate MIR which does copies even when the type actually doesn’t satisfy the full requirements for the Copy trait (cc #29149) – this winds up being reported as an error during NLL borrow check.

pub fn is_sized(self, tcx_at: TyCtxtAt<'tcx>, param_env: ParamEnv<'tcx>) -> bool

Checks whether values of this type T have a size known at compile time (i.e., whether T: Sized). Lifetimes are ignored for the purposes of this check, so it can be an over-approximation in generic contexts, where one can have strange rules like <T as Foo<'static>>::Bar: Sized that actually carry lifetime requirements.

pub fn is_freeze(self, tcx_at: TyCtxtAt<'tcx>, param_env: ParamEnv<'tcx>) -> bool

Checks whether values of this type T implement the Freeze trait – frozen types are those that do not contain an UnsafeCell anywhere. This is a language concept used to distinguish “true immutability”, which is relevant to optimization as well as the rules around static values. Note that the Freeze trait is not exposed to end users and is effectively an implementation detail.

fn is_trivially_freeze(self) -> bool

Fast path helper for testing if a type is Freeze.

Returning true means the type is known to be Freeze. Returning false means nothing – could be Freeze, might not be.

pub fn is_unpin(self, tcx_at: TyCtxtAt<'tcx>, param_env: ParamEnv<'tcx>) -> bool

Checks whether values of this type T implement the Unpin trait.

fn is_trivially_unpin(self) -> bool

Fast path helper for testing if a type is Unpin.

Returning true means the type is known to be Unpin. Returning false means nothing – could be Unpin, might not be.

pub fn needs_drop(self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>) -> bool

If ty.needs_drop(...) returns true, then ty is definitely non-copy and might have a destructor attached; if it returns false, then ty definitely has no destructor (i.e., no drop glue).

(Note that this implies that if ty has a destructor attached, then needs_drop will definitely return true for ty.)

Note that this method is used to check eligible types in unions.

pub fn has_significant_drop(
self,
tcx: TyCtxt<'tcx>,
param_env: ParamEnv<'tcx>
) -> bool

Checks if ty has has a significant drop.

Note that this method can return false even if ty has a destructor attached; even if that is the case then the adt has been marked with the attribute rustc_insignificant_dtor.

Note that this method is used to check for change in drop order for 2229 drop reorder migration analysis.

pub fn is_structural_eq_shallow(self, tcx: TyCtxt<'tcx>) -> bool

Returns true if equality for this type is both reflexive and structural.

Reflexive equality for a type is indicated by an Eq impl for that type.

Primitive types (u32, str) have structural equality by definition. For composite data types, equality for the type as a whole is structural when it is the same as equality between all components (fields, array elements, etc.) of that type. For ADTs, structural equality is indicated by an implementation of PartialStructuralEq and StructuralEq for that type.

This function is “shallow” because it may return true for a composite type whose fields are not StructuralEq. For example, [T; 4] has structural equality regardless of T because equality for arrays is determined by the equality of each array element. If you want to know whether a given call to PartialEq::eq will proceed structurally all the way down, you will need to use a type visitor.

pub fn peel_refs(self) -> Ty<'tcx>

Peel off all reference types in this type until there are none left.

This method is idempotent, i.e. ty.peel_refs().peel_refs() == ty.peel_refs().

Examples

u8 -> u8

&'a mut u8 -> u8

&'a &'b u8 -> u8

&'a *const &'b u8 -> *const &'b u8

pub fn outer_exclusive_binder(self) -> DebruijnIndex

impl<'tcx> Ty<'tcx>

pub fn walk(self) -> TypeWalker<'tcx>

Iterator that walks self and any types reachable from self, in depth-first order. Note that just walks the types that appear in self, it does not descend into the fields of structs or variants. For example:

isize => { isize }
Foo<Bar<isize>> => { Foo<Bar<isize>>, Bar<isize>, isize }
[isize] => { [isize], isize }

impl<'tcx> Ty<'tcx>

pub fn is_primitive_ty(self) -> bool

Similar to Ty::is_primitive, but also considers inferred numeric values to be primitive.

pub fn is_simple_ty(self) -> bool

Whether the type is succinctly representable as a type instead of just referred to with a description in error messages. This is used in the main error message.

pub fn is_simple_text(self) -> bool

Whether the type is succinctly representable as a type instead of just referred to with a description in error messages. This is used in the primary span label. Beyond what is_simple_ty includes, it also accepts ADTs with no type arguments and references to ADTs with no type arguments.

impl<'tcx> Ty<'tcx>

Type utilities

pub fn kind(self) -> &'tcx TyKind<'tcx>

pub fn flags(self) -> TypeFlags

pub fn is_unit(self) -> bool

pub fn is_never(self) -> bool

pub fn is_primitive(self) -> bool

pub fn is_adt(self) -> bool

pub fn is_ref(self) -> bool

pub fn is_ty_var(self) -> bool

pub fn ty_vid(self) -> Option<TyVid>

pub fn is_ty_infer(self) -> bool

pub fn is_phantom_data(self) -> bool

pub fn is_bool(self) -> bool

pub fn is_str(self) -> bool

Returns true if this type is a str.

pub fn is_param(self, index: u32) -> bool

pub fn is_slice(self) -> bool

pub fn is_array_slice(self) -> bool

pub fn is_array(self) -> bool

pub fn is_simd(self) -> bool

pub fn sequence_element_type(self, tcx: TyCtxt<'tcx>) -> Ty<'tcx>

pub fn simd_size_and_type(self, tcx: TyCtxt<'tcx>) -> (u64, Ty<'tcx>)

pub fn is_region_ptr(self) -> bool

pub fn is_mutable_ptr(self) -> bool

pub fn ref_mutability(self) -> Option<Mutability>

Get the mutability of the reference or None when not a reference

pub fn is_unsafe_ptr(self) -> bool

pub fn is_any_ptr(self) -> bool

Tests if this is any kind of primitive pointer type (reference, raw pointer, fn pointer).

pub fn is_box(self) -> bool

pub fn boxed_ty(self) -> Ty<'tcx>

Panics if called on any type other than Box<T>.

pub fn is_scalar(self) -> bool

A scalar type is one that denotes an atomic datum, with no sub-components. (A RawPtr is scalar because it represents a non-managed pointer, so its contents are abstract to rustc.)

pub fn is_floating_point(self) -> bool

Returns true if this type is a floating point type.

pub fn is_trait(self) -> bool

pub fn is_dyn_star(self) -> bool

pub fn is_enum(self) -> bool

pub fn is_union(self) -> bool

pub fn is_closure(self) -> bool

pub fn is_generator(self) -> bool

pub fn is_integral(self) -> bool

pub fn is_fresh_ty(self) -> bool

pub fn is_fresh(self) -> bool

pub fn is_char(self) -> bool

pub fn is_numeric(self) -> bool

pub fn is_signed(self) -> bool

pub fn is_ptr_sized_integral(self) -> bool

pub fn has_concrete_skeleton(self) -> bool

pub fn contains(self, other: Ty<'tcx>) -> bool

Checks whether a type recursively contains another type

Example: Option<()> contains ()

pub fn builtin_deref(self, explicit: bool) -> Option<TypeAndMut<'tcx>>

Returns the type and mutability of *ty.

The parameter explicit indicates if this is an explicit dereference. Some types – notably unsafe ptrs – can only be dereferenced explicitly.

pub fn builtin_index(self) -> Option<Ty<'tcx>>

Returns the type of ty[i].

pub fn fn_sig(self, tcx: TyCtxt<'tcx>) -> PolyFnSig<'tcx>

pub fn is_fn(self) -> bool

pub fn is_fn_ptr(self) -> bool

pub fn is_impl_trait(self) -> bool

pub fn ty_adt_def(self) -> Option<AdtDef<'tcx>>

pub fn tuple_fields(self) -> &'tcx List<Ty<'tcx>>

Iterates over tuple fields. Panics when called on anything but a tuple.

pub fn variant_range(self, tcx: TyCtxt<'tcx>) -> Option<Range<VariantIdx>>

If the type contains variants, returns the valid range of variant indices.

pub fn discriminant_for_variant(
self,
tcx: TyCtxt<'tcx>,
variant_index: VariantIdx
) -> Option<Discr<'tcx>>

If the type contains variants, returns the variant for variant_index. Panics if variant_index is out of range.

pub fn discriminant_ty(self, tcx: TyCtxt<'tcx>) -> Ty<'tcx>

Returns the type of the discriminant of this type.

pub fn ptr_metadata_ty(
self,
tcx: TyCtxt<'tcx>,
normalize: impl FnMut(Ty<'tcx>) -> Ty<'tcx>
) -> (Ty<'tcx>, bool)

Returns the type of metadata for (potentially fat) pointers to this type, and a boolean signifying if this is conditional on this type being Sized.

pub fn to_opt_closure_kind(self) -> Option<ClosureKind>

When we create a closure, we record its kind (i.e., what trait it implements) into its ClosureSubsts using a type parameter. This is kind of a phantom type, except that the most convenient thing for us to are the integral types. This function converts such a special type into the closure kind. To go the other way, use tcx.closure_kind_ty(closure_kind).

Note that during type checking, we use an inference variable to represent the closure kind, because it has not yet been inferred. Once upvar inference (in rustc_typeck/src/check/upvar.rs) is complete, that type variable will be unified.

pub fn is_trivially_sized(self, tcx: TyCtxt<'tcx>) -> bool

Fast path helper for testing if a type is Sized.

Returning true means the type is known to be sized. Returning false means nothing – could be sized, might not be.

Note that we could never rely on the fact that a type such as [_] is trivially !Sized because we could be in a type environment with a bound such as [_]: Copy. A function with such a bound obviously never can be called, but that doesn’t mean it shouldn’t typecheck. This is why this method doesn’t return Option<bool>.

pub fn is_trivially_pure_clone_copy(self) -> bool

Fast path helper for primitives which are always Copy and which have a side-effect-free Clone impl.

Returning true means the type is known to be pure and Copy+Clone. Returning false means nothing – could be Copy, might not be.

This is mostly useful for optimizations, as there are the types on which we can replace cloning with dereferencing.