Struct rustc_middle::ty::TyCtxtFeed

pub struct TyCtxtFeed<'tcx, KEY: Copy> {
    pub tcx: TyCtxt<'tcx>,
    key: KEY,
}

This struct should only be created by create_def.

Fields
tcx: TyCtxt<'tcx>key: KEY
Implementations

impl<'tcx, KEY: Copy> TyCtxtFeed<'tcx, KEY>

pub fn key(&self) -> KEY

impl<'tcx> TyCtxtFeed<'tcx, LocalDefId>

pub fn def_id(&self) -> LocalDefId

impl<'tcx, K: IntoQueryParam<()> + Copy> TyCtxtFeed<'tcx, K>

pub fn resolutions(self, value: ProvidedValue<'tcx>)

[query description - consider adding a doc-comment!] getting the resolver outputs

impl<'tcx, K: IntoQueryParam<LocalDefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn opt_local_def_id_to_hir_id(self, value: ProvidedValue<'tcx>)

Gives access to the HIR ID for the given LocalDefId owner key if any.

Definitions that were generated with no HIR, would be fed to return None.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn type_of(self, value: ProvidedValue<'tcx>)

Returns the Ty of the given DefId. If the DefId points to an alias, it will “skip” this alias to return the aliased type.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn is_type_alias_impl_trait(self, value: ProvidedValue<'tcx>)

[query description - consider adding a doc-comment!] determine whether the opaque is a type-alias impl trait

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn generics_of(self, value: ProvidedValue<'tcx>)

Maps from the DefId of an item (trait/struct/enum/fn) to its associated generics.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn explicit_item_bounds(self, value: ProvidedValue<'tcx>)

Returns the list of bounds that can be used for SelectionCandidate::ProjectionCandidate(_) and ProjectionTyCandidate::TraitDef. Specifically this is the bounds written on the trait’s type definition, or those after the impl keyword

type X: Bound + 'lt
//      ^^^^^^^^^^^
impl Debug + Display
//   ^^^^^^^^^^^^^^^

key is the DefId of the associated type or opaque type.

Bounds from the parent (e.g. with nested impl trait) are not included.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn explicit_predicates_of(self, value: ProvidedValue<'tcx>)

Returns the predicates written explicitly by the user.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn inferred_outlives_of(self, value: ProvidedValue<'tcx>)

Returns the inferred outlives predicates (e.g., for struct Foo<'a, T> { x: &'a T }, this would return T: 'a).

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn associated_item(self, value: ProvidedValue<'tcx>)

Maps from a trait item to the trait item “descriptor”.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn opt_def_kind(self, value: ProvidedValue<'tcx>)

[query description - consider adding a doc-comment!] looking up definition kind of tcx.def_path_str(def_id)

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn def_span(self, value: ProvidedValue<'tcx>)

Gets the span for the definition.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn def_ident_span(self, value: ProvidedValue<'tcx>)

Gets the span for the identifier of the definition.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn param_env(self, value: ProvidedValue<'tcx>)

Gets the ParameterEnvironment for a given item; this environment will be in “user-facing” mode, meaning that it is suitable for type-checking etc, and it does not normalize specializable associated types. This is almost always what you want, unless you are doing MIR optimizations, in which case you might want to use reveal_all() method to change modes.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn defaultness(self, value: ProvidedValue<'tcx>)

Returns whether the impl or associated function has the default keyword.

impl<'tcx, K: IntoQueryParam<DefId> + Copy> TyCtxtFeed<'tcx, K>

pub fn visibility(self, value: ProvidedValue<'tcx>)

Computes the visibility of the provided def_id.

If the item from the def_id doesn’t have a visibility, it will panic. For example a generic type parameter will panic if you call this method on it:

use std::fmt::Debug;

pub trait Foo<T: Debug> {}

In here, if you call visibility on T, it’ll panic.

impl<'tcx, K: IntoQueryParam<CrateNum> + Copy> TyCtxtFeed<'tcx, K>

pub fn crate_name(self, value: ProvidedValue<'tcx>)

Gets the name of the crate.

impl<'tcx, K: IntoQueryParam<()> + Copy> TyCtxtFeed<'tcx, K>

pub fn output_filenames(self, value: ProvidedValue<'tcx>)

Return the filenames where output artefacts shall be stored.

This query returns an &Arc because codegen backends need the value even after the TyCtxt has been destroyed.

impl<'tcx, K: IntoQueryParam<()> + Copy> TyCtxtFeed<'tcx, K>

pub fn features_query(self, value: ProvidedValue<'tcx>)

[query description - consider adding a doc-comment!] looking up enabled feature gates

impl<'tcx, K: IntoQueryParam<()> + Copy> TyCtxtFeed<'tcx, K>

pub fn crate_for_resolver(self, value: ProvidedValue<'tcx>)

[query description - consider adding a doc-comment!] the ast before macro expansion and name resolution

impl<'tcx, K: IntoQueryParam<CrateNum> + Copy> TyCtxtFeed<'tcx, K>

pub fn stripped_cfg_items(self, value: ProvidedValue<'tcx>)

Get all item paths that were stripped by a #[cfg] in a particular crate. Should not be called for the local crate before the resolver outputs are created, as it is only fed there.