Struct rustc_middle::ty::WithOptConstParam

pub struct WithOptConstParam<T> {
    pub did: T,
    pub const_param_did: Option<DefId>,
}

A DefId which, in case it is a const argument, is potentially bundled with the DefId of the generic parameter it instantiates.

This is used to avoid calls to type_of for const arguments during typeck which cause cycle errors.

struct A;
impl A {
    fn foo<const N: usize>(&self) -> [u8; N] { [0; N] }
    //           ^ const parameter
}
struct B;
impl B {
    fn foo<const M: u8>(&self) -> usize { 42 }
    //           ^ const parameter
}

fn main() {
    let a = A;
    let _b = a.foo::<{ 3 + 7 }>();
    //               ^^^^^^^^^ const argument
}

Let’s look at the call a.foo::<{ 3 + 7 }>() here. We do not know which foo is used until we know the type of a.

We only know the type of a once we are inside of typeck(main). We also end up normalizing the type of _b during typeck(main) which requires us to evaluate the const argument.

To evaluate that const argument we need to know its type, which we would get using type_of(const_arg). This requires us to resolve foo as it can be either usize or u8 in this example. However, resolving foo once again requires typeck(main) to get the type of a, which results in a cycle.

In short we must not call type_of(const_arg) during typeck(main).

When first creating the ty::Const of the const argument inside of typeck we have already resolved foo so we know which const parameter this argument instantiates. This means that we also know the expected result of type_of(const_arg) even if we aren’t allowed to call that query: it is equal to type_of(const_param) which is trivial to compute.

If we now want to use that constant in a place which potentially needs its type we also pass the type of its const_param. This is the point of WithOptConstParam, except that instead of a Ty we bundle the DefId of the const parameter. Meaning that we need to use type_of(const_param_did) if const_param_did is Some to get the type of did.

Fields
did: Tconst_param_did: Option<DefId>

The DefId of the corresponding generic parameter in case did is a const argument.

Note that even if did is a const argument, this may still be None. All queries taking WithOptConstParam start by calling tcx.opt_const_param_of(def.did) to potentially update param_did in the case it is None.

Implementations

impl<T> WithOptConstParam<T>

pub fn unknown(did: T) -> WithOptConstParam<T>

Creates a new WithOptConstParam setting const_param_did to None.

impl WithOptConstParam<LocalDefId>

pub fn try_lookup(
did: LocalDefId,
tcx: TyCtxt<'_>
) -> Option<(LocalDefId, DefId)>

Returns Some((did, param_did)) if def_id is a const argument, None otherwise.

pub fn try_upgrade(
self,
tcx: TyCtxt<'_>
) -> Option<WithOptConstParam<LocalDefId>>

In case self is unknown but self.did is a const argument, this returns a WithOptConstParam with the correct const_param_did.

pub fn to_global(self) -> WithOptConstParam<DefId>

pub fn def_id_for_type_of(self) -> DefId

impl WithOptConstParam<DefId>

pub fn as_local(self) -> Option<WithOptConstParam<LocalDefId>>

pub fn as_const_arg(self) -> Option<(LocalDefId, DefId)>

pub fn is_local(self) -> bool

pub fn def_id_for_type_of(self) -> DefId