rustc_middle::ty
Trait ExistentialPredicateStableCmpExt

pub trait ExistentialPredicateStableCmpExt<'tcx> {
    // Required method
    fn stable_cmp(&self, tcx: TyCtxt<'tcx>, other: &Self) -> Ordering;
}

Required Methods

fn stable_cmp(&self, tcx: TyCtxt<'tcx>, other: &Self) -> Ordering

Compares via an ordering that will not change if modules are reordered or other changes are made to the tree. In particular, this ordering is preserved across incremental compilations.

Object Safety

This trait is not object safe.