rustc_middle::ty::trait_def
Struct TraitDef

pub struct TraitDef {

    pub def_id: DefId,
    pub safety: Safety,
    pub constness: Constness,
    pub paren_sugar: bool,
    pub has_auto_impl: bool,
    pub is_marker: bool,
    pub is_coinductive: bool,
    pub is_fundamental: bool,
    pub skip_array_during_method_dispatch: bool,
    pub skip_boxed_slice_during_method_dispatch: bool,
    pub specialization_kind: TraitSpecializationKind,
    pub must_implement_one_of: Option<Box<[Ident]>>,
    pub implement_via_object: bool,
    pub deny_explicit_impl: bool,

}

A trait’s definition with type information.

Fields
def_id: DefIdsafety: Safetyconstness: Constness

Whether this trait has been annotated with #[const_trait].

paren_sugar: bool

If true, then this trait had the #[rustc_paren_sugar] attribute, indicating that it should be used with Foo() sugar. This is a temporary thing – eventually any trait will be usable with the sugar (or without it).

has_auto_impl: boolis_marker: bool

If true, then this trait has the #[marker] attribute, indicating that all its associated items have defaults that cannot be overridden, and thus impls of it are allowed to overlap.

is_coinductive: bool

If true, then this trait has the #[rustc_coinductive] attribute or is an auto trait. This indicates that trait solver cycles involving an X: ThisTrait goal are accepted.

In the future all traits should be coinductive, but we need a better formal understanding of what exactly that means and should probably also have already switched to the new trait solver.

is_fundamental: bool

If true, then this trait has the #[fundamental] attribute. This affects how conherence computes whether a trait may have trait implementations added in the future.

skip_array_during_method_dispatch: bool

If true, then this trait has the #[rustc_skip_during_method_dispatch(array)] attribute, indicating that editions before 2021 should not consider this trait during method dispatch if the receiver is an array.

skip_boxed_slice_during_method_dispatch: bool

If true, then this trait has the #[rustc_skip_during_method_dispatch(boxed_slice)] attribute, indicating that editions before 2024 should not consider this trait during method dispatch if the receiver is a boxed slice.

specialization_kind: TraitSpecializationKind

Used to determine whether the standard library is allowed to specialize on this trait.

must_implement_one_of: Option<Box<[Ident]>>

List of functions from #[rustc_must_implement_one_of] attribute one of which must be implemented.

implement_via_object: bool

Whether to add a builtin dyn Trait: Trait implementation. This is enabled for all traits except ones marked with #[rustc_deny_explicit_impl(implement_via_object = false)].

deny_explicit_impl: bool

Whether a trait is fully built-in, and any implementation is disallowed. This only applies to built-in traits, and is marked via #[rustc_deny_explicit_impl(implement_via_object = ...)].

Implementations

impl<'tcx> TraitDef

pub fn ancestors( &self, tcx: TyCtxt<'tcx>, of_impl: DefId, ) -> Result<Ancestors<'tcx>, ErrorGuaranteed>