rustc_middle::ty
Type Alias CanonicalUserTypeAnnotations

pub type CanonicalUserTypeAnnotations<'tcx> = IndexVec<UserTypeAnnotationIndex, CanonicalUserTypeAnnotation<'tcx>>;

Mapping of type annotation indices to canonical user type annotations.

Aliased Type

struct CanonicalUserTypeAnnotations<'tcx> {
    pub raw: Vec<CanonicalUserTypeAnnotation<'tcx>>,
    /* private fields */
}

Fields
raw: Vec<CanonicalUserTypeAnnotation<'tcx>>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 24 bytes