Type Definition rustc_middle::ty::PolyProjectionPredicate

pub type PolyProjectionPredicate<'tcx> = Binder<'tcx, ProjectionPredicate<'tcx>>;

Implementations

impl<'tcx> PolyProjectionPredicate<'tcx>

pub fn trait_def_id(&self, tcx: TyCtxt<'tcx>) -> DefId

Returns the DefId of the trait of the associated item being projected.

pub fn required_poly_trait_ref(&self, tcx: TyCtxt<'tcx>) -> PolyTraitRef<'tcx>

Get the PolyTraitRef required for this projection to be well formed. Note that for generic associated types the predicates of the associated type also need to be checked.

pub fn term(&self) -> Binder<'tcx, Term<'tcx>>

pub fn projection_def_id(&self) -> DefId

The DefId of the TraitItem for the associated type.

Note that this is not the DefId of the TraitRef containing this associated type, which is in tcx.associated_item(projection_def_id()).container.