rustc_middle::ty::typeck_results
Function validate_hir_id_for_typeck_results

fn validate_hir_id_for_typeck_results(hir_owner: OwnerId, hir_id: HirId)

Validate that the given HirId (respectively its local_id part) can be safely used as a key in the maps of a TypeckResults. For that to be the case, the HirId must have the same owner as all the other IDs in this table (signified by hir_owner). Otherwise the HirId would be in a different frame of reference and using its local_id would result in lookup errors, or worse, in silently wrong data being stored/returned.