rustc_middle::ty::util
Function intrinsic_raw

pub fn intrinsic_raw(
    tcx: TyCtxt<'_>,
    def_id: LocalDefId,
) -> Option<IntrinsicDef>

Determines whether an item is an intrinsic (which may be via Abi or via the rustc_intrinsic attribute).

We double check the feature gate here because whether a function may be defined as an intrinsic causes the compiler to make some assumptions about its shape; if the user doesn’t use a feature gate, they may cause an ICE that we otherwise may want to prevent.