rustc_middle::ty::util

Function needs_drop_components_with_async

pub fn needs_drop_components_with_async<'tcx>(
    tcx: TyCtxt<'tcx>,
    ty: Ty<'tcx>,
    asyncness: Asyncness,
) -> Result<SmallVec<[Ty<'tcx>; 2]>, AlwaysRequiresDrop>

Returns a list of types such that the given type needs drop if and only if any of the returned types need drop. Returns Err(AlwaysRequiresDrop) if this type always needs drop.