rustc_middle::ty
Module visit

Structs

LateBoundRegionsCollector 🔒

Collects all the late-bound regions at the innermost binding level into a hash set.

MaxUniverse

Finds the max universe present

Traits

TypeSuperVisitable

TypeVisitable

This trait is implemented for every type that can be visited, providing the skeleton of the traversal.

TypeVisitableExt

TypeVisitor

This trait is implemented for every visiting traversal. There is a visit method defined for every type of interest. Each such method has a default that recurses into the type’s fields in a non-custom fashion.