Struct rustc_middle::ty::visit::LateBoundRegionsCollector

struct LateBoundRegionsCollector {
    current_index: DebruijnIndex,
    regions: FxHashSet<BoundRegionKind>,
    just_constrained: bool,
}

Collects all the late-bound regions at the innermost binding level into a hash set.

Fields
current_index: DebruijnIndexregions: FxHashSet<BoundRegionKind>just_constrained: bool

true if we only want regions that are known to be “constrained” when you equate this type with another type. In particular, if you have e.g., &'a u32 and &'b u32, equating them constraints 'a == 'b. But if you have <&'a u32 as Trait>::Foo and <&'b u32 as Trait>::Foo, normalizing those types may mean that 'a and 'b don’t appear in the results, so they are not considered constrained.

Implementations

impl LateBoundRegionsCollector

fn new(just_constrained: bool) -> Self