Trait rustc_middle::ty::visit::TypeVisitable

pub trait TypeVisitable<'tcx>: Debug + Clone {

    fn visit_with<V: TypeVisitor<'tcx>>(
        &self,
        visitor: &mut V
    ) -> ControlFlow<V::BreakTy>;

    fn has_vars_bound_at_or_above(&self, binder: DebruijnIndex) -> bool { ... }
    fn has_vars_bound_above(&self, binder: DebruijnIndex) -> bool { ... }
    fn has_escaping_bound_vars(&self) -> bool { ... }
    fn has_type_flags(&self, flags: TypeFlags) -> bool { ... }
    fn has_projections(&self) -> bool { ... }
    fn has_opaque_types(&self) -> bool { ... }
    fn references_error(&self) -> bool { ... }
    fn error_reported(&self) -> Option<ErrorGuaranteed> { ... }
    fn has_param_types_or_consts(&self) -> bool { ... }
    fn has_infer_regions(&self) -> bool { ... }
    fn has_infer_types(&self) -> bool { ... }
    fn has_infer_types_or_consts(&self) -> bool { ... }
    fn needs_infer(&self) -> bool { ... }
    fn has_placeholders(&self) -> bool { ... }
    fn needs_subst(&self) -> bool { ... }
    fn has_free_regions(&self) -> bool { ... }
    fn has_erased_regions(&self) -> bool { ... }
    fn has_erasable_regions(&self) -> bool { ... }
    fn is_global(&self) -> bool { ... }
    fn has_late_bound_regions(&self) -> bool { ... }
    fn still_further_specializable(&self) -> bool { ... }

}

This trait is implemented for every type that can be visited, providing the skeleton of the traversal.

To implement this conveniently, use the derive macro located in rustc_macros.

Required Methods

fn visit_with<V: TypeVisitor<'tcx>>(
&self,
visitor: &mut V
) -> ControlFlow<V::BreakTy>

The entry point for visiting. To visit a value t with a visitor v call: t.visit_with(v).

For most types, this just traverses the value, calling visit_with on each field/element.

For types of interest (such as Ty), the implementation of this method that calls a visitor method specifically for that type (such as V::visit_ty). This is where control transfers from TypeFoldable to TypeVisitor.

Provided Methods

fn has_vars_bound_at_or_above(&self, binder: DebruijnIndex) -> bool

Returns true if self has any late-bound regions that are either bound by binder or bound by some binder outside of binder. If binder is ty::INNERMOST, this indicates whether there are any late-bound regions that appear free.

fn has_vars_bound_above(&self, binder: DebruijnIndex) -> bool

Returns true if this self has any regions that escape binder (and hence are not bound by it).

fn has_escaping_bound_vars(&self) -> bool

fn has_type_flags(&self, flags: TypeFlags) -> bool

fn has_projections(&self) -> bool

fn has_opaque_types(&self) -> bool

fn references_error(&self) -> bool

fn error_reported(&self) -> Option<ErrorGuaranteed>

fn has_param_types_or_consts(&self) -> bool

fn has_infer_regions(&self) -> bool

fn has_infer_types(&self) -> bool

fn has_infer_types_or_consts(&self) -> bool

fn needs_infer(&self) -> bool

fn has_placeholders(&self) -> bool

fn needs_subst(&self) -> bool

fn has_free_regions(&self) -> bool

“Free” regions in this context means that it has any region that is not (a) erased or (b) late-bound.

fn has_erased_regions(&self) -> bool

fn has_erasable_regions(&self) -> bool

True if there are any un-erased free regions.

fn is_global(&self) -> bool

Indicates whether this value references only ‘global’ generic parameters that are the same regardless of what fn we are in. This is used for caching.

fn has_late_bound_regions(&self) -> bool

True if there are any late-bound regions

fn still_further_specializable(&self) -> bool

Indicates whether this value still has parameters/placeholders/inference variables which could be replaced later, in a way that would change the results of impl specialization.