Trait rustc_middle::ty::visit::TypeVisitableExt

pub trait TypeVisitableExt<'tcx>: TypeVisitable<TyCtxt<'tcx>> {

    // Provided methods
    fn has_vars_bound_at_or_above(&self, binder: DebruijnIndex) -> bool { ... }
    fn has_vars_bound_above(&self, binder: DebruijnIndex) -> bool { ... }
    fn has_escaping_bound_vars(&self) -> bool { ... }
    fn has_type_flags(&self, flags: TypeFlags) -> bool { ... }
    fn has_projections(&self) -> bool { ... }
    fn has_inherent_projections(&self) -> bool { ... }
    fn has_opaque_types(&self) -> bool { ... }
    fn has_generators(&self) -> bool { ... }
    fn references_error(&self) -> bool { ... }
    fn error_reported(&self) -> Result<(), ErrorGuaranteed> { ... }
    fn has_non_region_param(&self) -> bool { ... }
    fn has_infer_regions(&self) -> bool { ... }
    fn has_infer_types(&self) -> bool { ... }
    fn has_non_region_infer(&self) -> bool { ... }
    fn has_infer(&self) -> bool { ... }
    fn has_placeholders(&self) -> bool { ... }
    fn has_non_region_placeholders(&self) -> bool { ... }
    fn has_param(&self) -> bool { ... }
    fn has_free_regions(&self) -> bool { ... }
    fn has_erased_regions(&self) -> bool { ... }
    fn has_erasable_regions(&self) -> bool { ... }
    fn is_global(&self) -> bool { ... }
    fn has_late_bound_regions(&self) -> bool { ... }
    fn has_non_region_late_bound(&self) -> bool { ... }
    fn has_late_bound_vars(&self) -> bool { ... }
    fn still_further_specializable(&self) -> bool { ... }

}

Provided Methods

fn has_vars_bound_at_or_above(&self, binder: DebruijnIndex) -> bool

Returns true if self has any late-bound regions that are either bound by binder or bound by some binder outside of binder. If binder is ty::INNERMOST, this indicates whether there are any late-bound regions that appear free.

fn has_vars_bound_above(&self, binder: DebruijnIndex) -> bool

Returns true if this type has any regions that escape binder (and hence are not bound by it).

fn has_escaping_bound_vars(&self) -> bool

Return true if this type has regions that are not a part of the type. For example, for<'a> fn(&'a i32) return false, while fn(&'a i32) would return true. The latter can occur when traversing through the former.

See HasEscapingVarsVisitor for more information.

fn has_type_flags(&self, flags: TypeFlags) -> bool

fn has_projections(&self) -> bool

fn has_inherent_projections(&self) -> bool

fn has_opaque_types(&self) -> bool

fn has_generators(&self) -> bool

fn references_error(&self) -> bool

fn error_reported(&self) -> Result<(), ErrorGuaranteed>

fn has_non_region_param(&self) -> bool

fn has_infer_regions(&self) -> bool

fn has_infer_types(&self) -> bool

fn has_non_region_infer(&self) -> bool

fn has_infer(&self) -> bool

fn has_placeholders(&self) -> bool

fn has_non_region_placeholders(&self) -> bool

fn has_param(&self) -> bool

fn has_free_regions(&self) -> bool

“Free” regions in this context means that it has any region that is not (a) erased or (b) late-bound.

fn has_erased_regions(&self) -> bool

fn has_erasable_regions(&self) -> bool

True if there are any un-erased free regions.

fn is_global(&self) -> bool

Indicates whether this value references only ‘global’ generic parameters that are the same regardless of what fn we are in. This is used for caching.

fn has_late_bound_regions(&self) -> bool

True if there are any late-bound regions

fn has_non_region_late_bound(&self) -> bool

True if there are any late-bound non-region variables

fn has_late_bound_vars(&self) -> bool

True if there are any late-bound variables

fn still_further_specializable(&self) -> bool

Indicates whether this value still has parameters/placeholders/inference variables which could be replaced later, in a way that would change the results of impl specialization.