Function rustc_middle::ty::vtable::vtable_allocation_provider

pub(super) fn vtable_allocation_provider<'tcx>(
    tcx: TyCtxt<'tcx>,
    key: (Ty<'tcx>, Option<PolyExistentialTraitRef<'tcx>>)
) -> AllocId

Retrieves an allocation that represents the contents of a vtable. Since this is a query, allocations are cached and not duplicated.