rustc_middle::util
Module bug

Functions

bug_fmt

opt_span_bug_fmt 🔒

provide

span_bug_fmt

trigger_delayed_bug

A query to trigger a delayed bug. Clearly, if one has a tcx one can already trigger a delayed bug, so what is the point of this? It exists to help us test the interaction of delayed bugs with the query system and incremental.