rustc_middle::dep_graph

Module dep_node

Modules

dep_kinds

label_strs

Contains variant => str representations for constructing DepNode groups for tests.

Macros

define_dep_nodes ğŸ”’

Structs

DepKind

This serves as an index into arrays built by make_dep_kind_array.

DepNode

Enums

DepKindDefs ğŸ”’

This enum serves as an index into arrays built by make_dep_kind_array.

Constants

DEP_KIND_VARIANTS ğŸ”’

Traits

DepContext ğŸ”’

DepNodeExt

DepNodeParams ğŸ”’

Functions

dep_kind_from_label_string ğŸ”’

make_compile_codegen_unit ğŸ”’

make_compile_mono_item ğŸ”’

make_metadata ğŸ”’