Enum rustc_middle::dep_graph::dep_node::DepKindDefs

#[repr(u16)]
enum DepKindDefs {

    Null = 0,
    Red = 1,
    TraitSelect = 2,
    CompileCodegenUnit = 3,
    CompileMonoItem = 4,
    trigger_delayed_bug = 5,
    registered_tools = 6,
    early_lint_checks = 7,
    resolutions = 8,
    resolver_for_lowering_raw = 9,
    source_span = 10,
    hir_crate = 11,
    hir_crate_items = 12,
    hir_module_items = 13,
    local_def_id_to_hir_id = 14,
    hir_owner_parent = 15,
    opt_hir_owner_nodes = 16,
    hir_attrs = 17,
    const_param_default = 18,
    type_of = 19,
    type_of_opaque = 20,
    type_alias_is_lazy = 21,
    collect_return_position_impl_trait_in_trait_tys = 22,
    is_type_alias_impl_trait = 23,
    unsizing_params_for_adt = 24,
    analysis = 25,
    check_expectations = 26,
    generics_of = 27,
    predicates_of = 28,
    opaque_types_defined_by = 29,
    explicit_item_bounds = 30,
    item_bounds = 31,
    native_libraries = 32,
    shallow_lint_levels_on = 33,
    lint_expectations = 34,
    expn_that_defined = 35,
    is_panic_runtime = 36,
    representability = 37,
    representability_adt_ty = 38,
    params_in_repr = 39,
    thir_body = 40,
    thir_tree = 41,
    thir_flat = 42,
    mir_keys = 43,
    mir_const_qualif = 44,
    mir_built = 45,
    mir_const = 46,
    thir_abstract_const = 47,
    mir_drops_elaborated_and_const_checked = 48,
    mir_for_ctfe = 49,
    mir_promoted = 50,
    closure_typeinfo = 51,
    closure_saved_names_of_captured_variables = 52,
    mir_coroutine_witnesses = 53,
    check_coroutine_obligations = 54,
    optimized_mir = 55,
    coverage_ids_info = 56,
    promoted_mir = 57,
    erase_regions_ty = 58,
    wasm_import_module_map = 59,
    predicates_defined_on = 60,
    trait_explicit_predicates_and_bounds = 61,
    explicit_predicates_of = 62,
    inferred_outlives_of = 63,
    super_predicates_of = 64,
    implied_predicates_of = 65,
    super_predicates_that_define_assoc_item = 66,
    type_param_predicates = 67,
    trait_def = 68,
    adt_def = 69,
    adt_destructor = 70,
    adt_sized_constraint = 71,
    adt_dtorck_constraint = 72,
    constness = 73,
    asyncness = 74,
    is_promotable_const_fn = 75,
    coroutine_kind = 76,
    coroutine_for_closure = 77,
    crate_variances = 78,
    variances_of = 79,
    inferred_outlives_crate = 80,
    associated_item_def_ids = 81,
    associated_item = 82,
    associated_items = 83,
    impl_item_implementor_ids = 84,
    associated_types_for_impl_traits_in_associated_fn = 85,
    associated_type_for_impl_trait_in_trait = 86,
    impl_trait_header = 87,
    issue33140_self_ty = 88,
    inherent_impls = 89,
    incoherent_impls = 90,
    mir_unsafety_check_result = 91,
    check_unsafety = 92,
    assumed_wf_types = 93,
    assumed_wf_types_for_rpitit = 94,
    fn_sig = 95,
    lint_mod = 96,
    check_unused_traits = 97,
    check_mod_attrs = 98,
    check_mod_unstable_api_usage = 99,
    check_mod_const_bodies = 100,
    check_mod_loops = 101,
    check_mod_naked_functions = 102,
    check_mod_privacy = 103,
    check_liveness = 104,
    live_symbols_and_ignored_derived_traits = 105,
    check_mod_deathness = 106,
    check_mod_type_wf = 107,
    coerce_unsized_info = 108,
    typeck = 109,
    diagnostic_only_typeck = 110,
    used_trait_imports = 111,
    has_typeck_results = 112,
    coherent_trait = 113,
    mir_borrowck = 114,
    crate_inherent_impls = 115,
    crate_inherent_impls_overlap_check = 116,
    orphan_check_impl = 117,
    mir_callgraph_reachable = 118,
    mir_inliner_callees = 119,
    eval_to_allocation_raw = 120,
    eval_static_initializer = 121,
    eval_to_const_value_raw = 122,
    eval_to_valtree = 123,
    valtree_to_const_val = 124,
    destructure_const = 125,
    lit_to_const = 126,
    check_match = 127,
    effective_visibilities = 128,
    check_private_in_public = 129,
    reachable_set = 130,
    region_scope_tree = 131,
    mir_shims = 132,
    symbol_name = 133,
    def_kind = 134,
    def_span = 135,
    def_ident_span = 136,
    lookup_stability = 137,
    lookup_const_stability = 138,
    lookup_default_body_stability = 139,
    should_inherit_track_caller = 140,
    lookup_deprecation_entry = 141,
    is_doc_hidden = 142,
    is_doc_notable_trait = 143,
    item_attrs = 144,
    codegen_fn_attrs = 145,
    asm_target_features = 146,
    fn_arg_names = 147,
    rendered_const = 148,
    impl_parent = 149,
    is_ctfe_mir_available = 150,
    is_mir_available = 151,
    own_existential_vtable_entries = 152,
    vtable_entries = 153,
    vtable_trait_upcasting_coercion_new_vptr_slot = 154,
    vtable_allocation = 155,
    codegen_select_candidate = 156,
    all_local_trait_impls = 157,
    trait_impls_of = 158,
    specialization_graph_of = 159,
    object_safety_violations = 160,
    check_is_object_safe = 161,
    param_env = 162,
    param_env_reveal_all_normalized = 163,
    is_copy_raw = 164,
    is_sized_raw = 165,
    is_freeze_raw = 166,
    is_unpin_raw = 167,
    needs_drop_raw = 168,
    has_significant_drop_raw = 169,
    has_structural_eq_impl = 170,
    adt_drop_tys = 171,
    adt_significant_drop_tys = 172,
    layout_of = 173,
    fn_abi_of_fn_ptr = 174,
    fn_abi_of_instance = 175,
    dylib_dependency_formats = 176,
    dependency_formats = 177,
    is_compiler_builtins = 178,
    has_global_allocator = 179,
    has_alloc_error_handler = 180,
    has_panic_handler = 181,
    is_profiler_runtime = 182,
    has_ffi_unwind_calls = 183,
    required_panic_strategy = 184,
    panic_in_drop_strategy = 185,
    is_no_builtins = 186,
    symbol_mangling_version = 187,
    extern_crate = 188,
    specializes = 189,
    in_scope_traits_map = 190,
    defaultness = 191,
    check_well_formed = 192,
    reachable_non_generics = 193,
    is_reachable_non_generic = 194,
    is_unreachable_local_definition = 195,
    upstream_monomorphizations = 196,
    upstream_monomorphizations_for = 197,
    upstream_drop_glue_for = 198,
    foreign_modules = 199,
    clashing_extern_declarations = 200,
    entry_fn = 201,
    proc_macro_decls_static = 202,
    crate_hash = 203,
    crate_host_hash = 204,
    extra_filename = 205,
    crate_extern_paths = 206,
    implementations_of_trait = 207,
    crate_incoherent_impls = 208,
    native_library = 209,
    resolve_bound_vars = 210,
    named_variable_map = 211,
    is_late_bound_map = 212,
    object_lifetime_default = 213,
    late_bound_vars_map = 214,
    visibility = 215,
    inhabited_predicate_adt = 216,
    inhabited_predicate_type = 217,
    dep_kind = 218,
    crate_name = 219,
    module_children = 220,
    extern_mod_stmt_cnum = 221,
    lib_features = 222,
    stability_implications = 223,
    intrinsic_raw = 224,
    get_lang_items = 225,
    all_diagnostic_items = 226,
    defined_lang_items = 227,
    diagnostic_items = 228,
    missing_lang_items = 229,
    visible_parent_map = 230,
    trimmed_def_paths = 231,
    missing_extern_crate_item = 232,
    used_crate_source = 233,
    debugger_visualizers = 234,
    postorder_cnums = 235,
    is_private_dep = 236,
    allocator_kind = 237,
    alloc_error_handler_kind = 238,
    upvars_mentioned = 239,
    maybe_unused_trait_imports = 240,
    names_imported_by_glob_use = 241,
    stability_index = 242,
    crates = 243,
    used_crates = 244,
    traits = 245,
    trait_impls_in_crate = 246,
    exported_symbols = 247,
    collect_and_partition_mono_items = 248,
    is_codegened_item = 249,
    codegen_unit = 250,
    unused_generic_params = 251,
    backend_optimization_level = 252,
    output_filenames = 253,
    normalize_canonicalized_projection_ty = 254,
    normalize_canonicalized_weak_ty = 255,
    normalize_canonicalized_inherent_projection_ty = 256,
    try_normalize_generic_arg_after_erasing_regions = 257,
    implied_outlives_bounds_compat = 258,
    implied_outlives_bounds = 259,
    dropck_outlives = 260,
    evaluate_obligation = 261,
    type_op_ascribe_user_type = 262,
    type_op_eq = 263,
    type_op_subtype = 264,
    type_op_prove_predicate = 265,
    type_op_normalize_ty = 266,
    type_op_normalize_clause = 267,
    type_op_normalize_poly_fn_sig = 268,
    type_op_normalize_fn_sig = 269,
    instantiate_and_check_impossible_predicates = 270,
    is_impossible_associated_item = 271,
    method_autoderef_steps = 272,
    supported_target_features = 273,
    features_query = 274,
    crate_for_resolver = 275,
    resolve_instance = 276,
    reveal_opaque_types_in_bounds = 277,
    limits = 278,
    diagnostic_hir_wf_check = 279,
    global_backend_features = 280,
    check_validity_requirement = 281,
    compare_impl_const = 282,
    deduced_param_attrs = 283,
    doc_link_resolutions = 284,
    doc_link_traits_in_scope = 285,
    check_tys_might_be_eq = 286,
    stripped_cfg_items = 287,
    generics_require_sized_self = 288,
    cross_crate_inlinable = 289,
    find_field = 290,

}

This enum serves as an index into arrays built by make_dep_kind_array.

Variants

Null = 0

We use this for most things when incr. comp. is turned off.

Red = 1

We use this to create a forever-red node.

TraitSelect = 2

CompileCodegenUnit = 3

CompileMonoItem = 4

trigger_delayed_bug = 5

This exists purely for testing the interactions between delayed bugs and incremental.

registered_tools = 6

Collects the list of all tools registered using #![register_tool].

early_lint_checks = 7

[query description - consider adding a doc-comment!] perform lints prior to macro expansion

resolutions = 8

[query description - consider adding a doc-comment!] getting the resolver outputs

resolver_for_lowering_raw = 9

[query description - consider adding a doc-comment!] getting the resolver for lowering

source_span = 10

Return the span for a definition. Contrary to def_span below, this query returns the full absolute span of the definition. This span is meant for dep-tracking rather than diagnostics. It should not be used outside of rustc_middle::hir::source_map.

hir_crate = 11

Represents crate as a whole (as distinct from the top-level crate module). If you call hir_crate (e.g., indirectly by calling tcx.hir().krate()), we will have to assume that any change means that you need to be recompiled. This is because the hir_crate query gives you access to all other items. To avoid this fate, do not call tcx.hir().krate(); instead, prefer wrappers like tcx.visit_all_items_in_krate().

hir_crate_items = 12

All items in the crate.

hir_module_items = 13

The items in a module.

This can be conveniently accessed by tcx.hir().visit_item_likes_in_module. Avoid calling this query directly.

local_def_id_to_hir_id = 14

Returns HIR ID for the given LocalDefId.

hir_owner_parent = 15

Gives access to the HIR node’s parent for the HIR owner key.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

opt_hir_owner_nodes = 16

Gives access to the HIR nodes and bodies inside key if it’s a HIR owner.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

hir_attrs = 17

Gives access to the HIR attributes inside the HIR owner key.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

const_param_default = 18

Given the def_id of a const-generic parameter, computes the associated default const parameter. e.g. fn example<const N: usize=3> called on N would return 3.

type_of = 19

Returns the Ty of the given DefId. If the DefId points to an alias, it will “skip” this alias to return the aliased type.

type_of_opaque = 20

Specialized instance of type_of that detects cycles that are due to revealing opaque because of an auto trait bound. Unless CyclePlaceholder needs to be handled separately, call type_of instead.

type_alias_is_lazy = 21

[query description - consider adding a doc-comment!] computing whether {path} is a lazy type alias

collect_return_position_impl_trait_in_trait_tys = 22

[query description - consider adding a doc-comment!] comparing an impl and trait method signature, inferring any hidden impl Trait types in the process

is_type_alias_impl_trait = 23

[query description - consider adding a doc-comment!] determine whether the opaque is a type-alias impl trait

unsizing_params_for_adt = 24

[query description - consider adding a doc-comment!] determining what parameters of tcx.def_path_str(key) can participate in unsizing

analysis = 25

The root query triggering all analysis passes like typeck or borrowck.

check_expectations = 26

This query checks the fulfillment of collected lint expectations. All lint emitting queries have to be done before this is executed to ensure that all expectations can be fulfilled.

This is an extra query to enable other drivers (like rustdoc) to only execute a small subset of the analysis query, while allowing lints to be expected. In rustc, this query will be executed as part of the analysis query and doesn’t have to be called a second time.

Tools can additionally pass in a tool filter. That will restrict the expectations to only trigger for lints starting with the listed tool name. This is useful for cases were not all linting code from rustc was called. With the default None all registered lints will also be checked for expectation fulfillment.

generics_of = 27

Maps from the DefId of an item (trait/struct/enum/fn) to its associated generics.

predicates_of = 28

Maps from the DefId of an item (trait/struct/enum/fn) to the predicates (where-clauses) that must be proven true in order to reference it. This is almost always the “predicates query” that you want.

predicates_of builds on predicates_defined_on – in fact, it is almost always the same as that query, except for the case of traits. For traits, predicates_of contains an additional Self: Trait<...> predicate that users don’t actually write. This reflects the fact that to invoke the trait (e.g., via Default::default) you must supply types that actually implement the trait. (However, this extra predicate gets in the way of some checks, which are intended to operate over only the actual where-clauses written by the user.)

opaque_types_defined_by = 29

[query description - consider adding a doc-comment!] computing the opaque types defined by tcx.def_path_str(key.to_def_id())

explicit_item_bounds = 30

Returns the list of bounds that can be used for SelectionCandidate::ProjectionCandidate(_) and ProjectionTyCandidate::TraitDef. Specifically this is the bounds written on the trait’s type definition, or those after the impl keyword

type X: Bound + 'lt
//      ^^^^^^^^^^^
impl Debug + Display
//   ^^^^^^^^^^^^^^^

key is the DefId of the associated type or opaque type.

Bounds from the parent (e.g. with nested impl trait) are not included.

item_bounds = 31

Elaborated version of the predicates from explicit_item_bounds.

For example:

trait MyTrait {
    type MyAType: Eq + ?Sized;
}

explicit_item_bounds returns [<Self as MyTrait>::MyAType: Eq], and item_bounds returns

[
    <Self as Trait>::MyAType: Eq,
    <Self as Trait>::MyAType: PartialEq<<Self as Trait>::MyAType>
]

Bounds from the parent (e.g. with nested impl trait) are not included.

native_libraries = 32

Look up all native libraries this crate depends on. These are assembled from the following places:

extern blocks (depending on their link attributes)

the libs (-l) option

shallow_lint_levels_on = 33

[query description - consider adding a doc-comment!] looking up lint levels for tcx.def_path_str(key)

lint_expectations = 34

[query description - consider adding a doc-comment!] computing #[expect]ed lints in this crate

expn_that_defined = 35

[query description - consider adding a doc-comment!] getting the expansion that defined tcx.def_path_str(key)

is_panic_runtime = 36

[query description - consider adding a doc-comment!] checking if the crate is_panic_runtime

representability = 37

Checks whether a type is representable or infinitely sized

representability_adt_ty = 38

An implementation detail for the representability query

params_in_repr = 39

Set of param indexes for type params that are in the type’s representation

thir_body = 40

Fetch the THIR for a given body. If typeck for that body failed, returns an empty Thir.

thir_tree = 41

Create a THIR tree for debugging.

thir_flat = 42

Create a list-like THIR representation for debugging.

mir_keys = 43

Set of all the DefIds in this crate that have MIR associated with them. This includes all the body owners, but also things like struct constructors.

mir_const_qualif = 44

Maps DefId’s that have an associated mir::Body to the result of the MIR const-checking pass. This is the set of qualifs in the final value of a const.

mir_built = 45

Fetch the MIR for a given DefId right after it’s built - this includes unreachable code.

mir_const = 46

Fetch the MIR for a given DefId up till the point where it is ready for const qualification.

See the README for the mir module for details.

thir_abstract_const = 47

Try to build an abstract representation of the given constant.

mir_drops_elaborated_and_const_checked = 48

[query description - consider adding a doc-comment!] elaborating drops for tcx.def_path_str(key)

mir_for_ctfe = 49

[query description - consider adding a doc-comment!] caching mir of tcx.def_path_str(key) for CTFE

mir_promoted = 50

[query description - consider adding a doc-comment!] promoting constants in MIR for tcx.def_path_str(key)

closure_typeinfo = 51

[query description - consider adding a doc-comment!] finding symbols for captures of closure tcx.def_path_str(key)

closure_saved_names_of_captured_variables = 52

Returns names of captured upvars for closures and coroutines.

Here are some examples:

name__field1__field2 when the upvar is captured by value.

_ref__name__field when the upvar is captured by reference.

For coroutines this only contains upvars that are shared by all states.

mir_coroutine_witnesses = 53

[query description - consider adding a doc-comment!] coroutine witness types for tcx.def_path_str(key)

check_coroutine_obligations = 54

[query description - consider adding a doc-comment!] verify auto trait bounds for coroutine interior type tcx.def_path_str(key)

optimized_mir = 55

MIR after our optimization passes have run. This is MIR that is ready for codegen. This is also the only query that can fetch non-local MIR, at present.

coverage_ids_info = 56

Summarizes coverage IDs inserted by the InstrumentCoverage MIR pass (for compiler option -Cinstrument-coverage), after MIR optimizations have had a chance to potentially remove some of them.

promoted_mir = 57

The DefId is the DefId of the containing MIR body. Promoteds do not have their own DefId. This function returns all promoteds in the specified body. The body references promoteds by the DefId and the mir::Promoted index. This is necessary, because after inlining a body may refer to promoteds from other bodies. In that case you still need to use the DefId of the original body.

erase_regions_ty = 58

Erases regions from ty to yield a new type. Normally you would just use tcx.erase_regions(value), however, which uses this query as a kind of cache.

wasm_import_module_map = 59

[query description - consider adding a doc-comment!] getting wasm import module map

predicates_defined_on = 60

Maps from the DefId of an item (trait/struct/enum/fn) to the predicates (where-clauses) directly defined on it. This is equal to the explicit_predicates_of predicates plus the inferred_outlives_of predicates.

trait_explicit_predicates_and_bounds = 61

Returns everything that looks like a predicate written explicitly by the user on a trait item.

Traits are unusual, because predicates on associated types are converted into bounds on that type for backwards compatibility:

trait X where Self::U: Copy { type U; }

becomes

trait X { type U: Copy; }

explicit_predicates_of and explicit_item_bounds will then take the appropriate subsets of the predicates here.

explicit_predicates_of = 62

Returns the predicates written explicitly by the user.

inferred_outlives_of = 63

Returns the inferred outlives predicates (e.g., for struct Foo<'a, T> { x: &'a T }, this would return T: 'a).

super_predicates_of = 64

Maps from the DefId of a trait to the list of super-predicates. This is a subset of the full list of predicates. We store these in a separate map because we must evaluate them even during type conversion, often before the full predicates are available (note that supertraits have additional acyclicity requirements).

implied_predicates_of = 65

[query description - consider adding a doc-comment!] computing the implied predicates of tcx.def_path_str(key)

super_predicates_that_define_assoc_item = 66

The Option<Ident> is the name of an associated type. If it is None, then this query returns the full set of predicates. If Some<Ident>, then the query returns only the subset of super-predicates that reference traits that define the given associated type. This is used to avoid cycles in resolving types like T::Item.

type_param_predicates = 67

To avoid cycles within the predicates of a single item we compute per-type-parameter predicates for resolving T::AssocTy.

trait_def = 68

[query description - consider adding a doc-comment!] computing trait definition for tcx.def_path_str(key)

adt_def = 69

[query description - consider adding a doc-comment!] computing ADT definition for tcx.def_path_str(key)

adt_destructor = 70

[query description - consider adding a doc-comment!] computing Drop impl for tcx.def_path_str(key)

adt_sized_constraint = 71

[query description - consider adding a doc-comment!] computing Sized constraints for tcx.def_path_str(key)

adt_dtorck_constraint = 72

[query description - consider adding a doc-comment!] computing drop-check constraints for tcx.def_path_str(key)

constness = 73

Returns true if this is a const fn, use the is_const_fn to know whether your crate actually sees it as const fn (e.g., the const-fn-ness might be unstable and you might not have the feature gate active).

Do not call this function manually. It is only meant to cache the base data for the is_const_fn function. Consider using is_const_fn or is_const_fn_raw instead.

asyncness = 74

[query description - consider adding a doc-comment!] checking if the function is async: tcx.def_path_str(key)

is_promotable_const_fn = 75

Returns true if calls to the function may be promoted.

This is either because the function is e.g., a tuple-struct or tuple-variant constructor, or because it has the #[rustc_promotable] attribute. The attribute should be removed in the future in favour of some form of check which figures out whether the function does not inspect the bits of any of its arguments (so is essentially just a constructor function).

coroutine_kind = 76

Returns Some(coroutine_kind) if the node pointed to by def_id is a coroutine.

coroutine_for_closure = 77

[query description - consider adding a doc-comment!] Given a coroutine-closure def id, return the def id of the coroutine returned by it

crate_variances = 78

Gets a map with the variance of every item; use variances_of instead.

variances_of = 79

Maps from the DefId of a type or region parameter to its (inferred) variance.

inferred_outlives_crate = 80

Maps from thee DefId of a type to its (inferred) outlives.

associated_item_def_ids = 81

Maps from an impl/trait or struct/variant DefId to a list of the DefIds of its associated items or fields.

associated_item = 82

Maps from a trait item to the trait item “descriptor”.

associated_items = 83

Collects the associated items defined on a trait or impl.

impl_item_implementor_ids = 84

Maps from associated items on a trait to the corresponding associated item on the impl specified by impl_id.

For example, with the following code

struct Type {}
                        // DefId
trait Trait {           // trait_id
    fn f();             // trait_f
    fn g() {}           // trait_g
}

impl Trait for Type {   // impl_id
    fn f() {}           // impl_f
    fn g() {}           // impl_g
}

The map returned for tcx.impl_item_implementor_ids(impl_id) would be { trait_f: impl_f, trait_g: impl_g }

associated_types_for_impl_traits_in_associated_fn = 85

Given fn_def_id of a trait or of an impl that implements a given trait: if fn_def_id is the def id of a function defined inside a trait, then it creates and returns the associated items that correspond to each impl trait in return position for that trait. if fn_def_id is the def id of a function defined inside an impl that implements a trait, then it creates and returns the associated items that correspond to each impl trait in return position of the implemented trait.

associated_type_for_impl_trait_in_trait = 86

Given an impl trait in trait opaque_ty_def_id, create and return the corresponding associated item.

impl_trait_header = 87

Given an impl_id, return the trait it implements along with some header information. Return None if this is an inherent impl.

issue33140_self_ty = 88

[query description - consider adding a doc-comment!] computing Self type wrt issue #33140 tcx.def_path_str(key)

inherent_impls = 89

Maps a DefId of a type to a list of its inherent impls. Contains implementations of methods that are inherent to a type. Methods in these implementations don’t need to be exported.

incoherent_impls = 90

[query description - consider adding a doc-comment!] collecting all inherent impls for {:?}

mir_unsafety_check_result = 91

The result of unsafety-checking this LocalDefId with the old checker.

check_unsafety = 92

Unsafety-check this LocalDefId.

assumed_wf_types = 93

Returns the types assumed to be well formed while “inside” of the given item.

Note that we’ve liberated the late bound regions of function signatures, so this can not be used to check whether these types are well formed.

assumed_wf_types_for_rpitit = 94

We need to store the assumed_wf_types for an RPITIT so that impls of foreign traits with return-position impl trait in traits can inherit the right wf types.

fn_sig = 95

Computes the signature of the function.

lint_mod = 96

Performs lint checking for the module.

check_unused_traits = 97

[query description - consider adding a doc-comment!] checking unused trait imports in crate

check_mod_attrs = 98

Checks the attributes in the module.

check_mod_unstable_api_usage = 99

Checks for uses of unstable APIs in the module.

check_mod_const_bodies = 100

Checks the const bodies in the module for illegal operations (e.g. if or loop).

check_mod_loops = 101

Checks the loops in the module.

check_mod_naked_functions = 102

[query description - consider adding a doc-comment!] checking naked functions in describe_as_module(key, tcx)

check_mod_privacy = 103

[query description - consider adding a doc-comment!] checking privacy in describe_as_module(key.to_local_def_id(), tcx)

check_liveness = 104

[query description - consider adding a doc-comment!] checking liveness of variables in tcx.def_path_str(key)

live_symbols_and_ignored_derived_traits = 105

Return the live symbols in the crate for dead code check.

The second return value maps from ADTs to ignored derived traits (e.g. Debug and Clone) and their respective impl (i.e., part of the derive macro)

check_mod_deathness = 106

[query description - consider adding a doc-comment!] checking deathness of variables in describe_as_module(key, tcx)

check_mod_type_wf = 107

[query description - consider adding a doc-comment!] checking that types are well-formed in describe_as_module(key, tcx)

coerce_unsized_info = 108

Caches CoerceUnsized kinds for impls on custom types.

typeck = 109

[query description - consider adding a doc-comment!] type-checking tcx.def_path_str(key)

diagnostic_only_typeck = 110

[query description - consider adding a doc-comment!] type-checking tcx.def_path_str(key)

used_trait_imports = 111

[query description - consider adding a doc-comment!] finding used_trait_imports tcx.def_path_str(key)

has_typeck_results = 112

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) has a body

coherent_trait = 113

[query description - consider adding a doc-comment!] coherence checking all impls of trait tcx.def_path_str(def_id)

mir_borrowck = 114

Borrow-checks the function body. If this is a closure, returns additional requirements that the closure’s creator must verify.

crate_inherent_impls = 115

Gets a complete map from all types to their inherent impls. Not meant to be used directly outside of coherence.

crate_inherent_impls_overlap_check = 116

Checks all types in the crate for overlap in their inherent impls. Reports errors. Not meant to be used directly outside of coherence.

orphan_check_impl = 117

Checks whether all impls in the crate pass the overlap check, returning which impls fail it. If all impls are correct, the returned slice is empty.

mir_callgraph_reachable = 118

Check whether the function has any recursion that could cause the inliner to trigger a cycle. Returns the call stack causing the cycle. The call stack does not contain the current function, just all intermediate functions.

mir_inliner_callees = 119

Obtain all the calls into other local functions

eval_to_allocation_raw = 120

Evaluates a constant and returns the computed allocation.

Do not use this directly, use the eval_to_const_value or eval_to_valtree instead.

eval_static_initializer = 121

Evaluate a static’s initializer, returning the allocation of the initializer’s memory.

eval_to_const_value_raw = 122

Evaluates const items or anonymous constants (such as enum variant explicit discriminants or array lengths) into a representation suitable for the type system and const generics.

Do not use this directly, use one of the following wrappers: tcx.const_eval_poly, tcx.const_eval_resolve, tcx.const_eval_instance, or tcx.const_eval_global_id.

eval_to_valtree = 123

Evaluate a constant and convert it to a type level constant or return None if that is not possible.

valtree_to_const_val = 124

Converts a type level constant value into ConstValue

destructure_const = 125

Destructures array, ADT or tuple constants into the constants of their fields.

lit_to_const = 126

[query description - consider adding a doc-comment!] converting literal to const

check_match = 127

[query description - consider adding a doc-comment!] match-checking tcx.def_path_str(key)

effective_visibilities = 128

Performs part of the privacy check and computes effective visibilities.

check_private_in_public = 129

[query description - consider adding a doc-comment!] checking for private elements in public interfaces

reachable_set = 130

[query description - consider adding a doc-comment!] reachability

region_scope_tree = 131

Per-body region::ScopeTree. The DefId should be the owner DefId for the body; in the case of closures, this will be redirected to the enclosing function.

mir_shims = 132

Generates a MIR body for the shim.

symbol_name = 133

The symbol_name query provides the symbol name for calling a given instance from the local crate. In particular, it will also look up the correct symbol name of instances from upstream crates.

def_kind = 134

[query description - consider adding a doc-comment!] looking up definition kind of tcx.def_path_str(def_id)

def_span = 135

Gets the span for the definition.

def_ident_span = 136

Gets the span for the identifier of the definition.

lookup_stability = 137

[query description - consider adding a doc-comment!] looking up stability of tcx.def_path_str(def_id)

lookup_const_stability = 138

[query description - consider adding a doc-comment!] looking up const stability of tcx.def_path_str(def_id)

lookup_default_body_stability = 139

[query description - consider adding a doc-comment!] looking up default body stability of tcx.def_path_str(def_id)

should_inherit_track_caller = 140

[query description - consider adding a doc-comment!] computing should_inherit_track_caller of tcx.def_path_str(def_id)

lookup_deprecation_entry = 141

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is deprecated

is_doc_hidden = 142

Determines whether an item is annotated with doc(hidden).

is_doc_notable_trait = 143

Determines whether an item is annotated with doc(notable_trait).

item_attrs = 144

Returns the attributes on the item at def_id.

Do not use this directly, use tcx.get_attrs instead.

codegen_fn_attrs = 145

[query description - consider adding a doc-comment!] computing codegen attributes of tcx.def_path_str(def_id)

asm_target_features = 146

[query description - consider adding a doc-comment!] computing target features for inline asm of tcx.def_path_str(def_id)

fn_arg_names = 147

[query description - consider adding a doc-comment!] looking up function parameter names for tcx.def_path_str(def_id)

rendered_const = 148

Gets the rendered value of the specified constant or associated constant. Used by rustdoc.

impl_parent = 149

[query description - consider adding a doc-comment!] computing specialization parent impl of tcx.def_path_str(def_id)

is_ctfe_mir_available = 150

[query description - consider adding a doc-comment!] checking if item has CTFE MIR available: tcx.def_path_str(key)

is_mir_available = 151

[query description - consider adding a doc-comment!] checking if item has MIR available: tcx.def_path_str(key)

own_existential_vtable_entries = 152

[query description - consider adding a doc-comment!] finding all existential vtable entries for trait tcx.def_path_str(key)

vtable_entries = 153

[query description - consider adding a doc-comment!] finding all vtable entries for trait tcx.def_path_str(key.def_id())

vtable_trait_upcasting_coercion_new_vptr_slot = 154

[query description - consider adding a doc-comment!] finding the slot within vtable for trait object key.1 vtable ptr during trait upcasting coercion from key.0 vtable

vtable_allocation = 155

[query description - consider adding a doc-comment!] vtable const allocation for < key.0 as key.1.map(| trait_ref | format! ("{trait_ref}")).unwrap_or("_".to_owned()) >

codegen_select_candidate = 156

[query description - consider adding a doc-comment!] computing candidate for key.1

all_local_trait_impls = 157

Return all impl blocks in the current crate.

trait_impls_of = 158

Given a trait trait_id, return all known impl blocks.

specialization_graph_of = 159

[query description - consider adding a doc-comment!] building specialization graph of trait tcx.def_path_str(trait_id)

object_safety_violations = 160

[query description - consider adding a doc-comment!] determining object safety of trait tcx.def_path_str(trait_id)

check_is_object_safe = 161

[query description - consider adding a doc-comment!] checking if trait tcx.def_path_str(trait_id) is object safe

param_env = 162

Gets the ParameterEnvironment for a given item; this environment will be in “user-facing” mode, meaning that it is suitable for type-checking etc, and it does not normalize specializable associated types. This is almost always what you want, unless you are doing MIR optimizations, in which case you might want to use reveal_all() method to change modes.

param_env_reveal_all_normalized = 163

Like param_env, but returns the ParamEnv in Reveal::All mode. Prefer this over tcx.param_env(def_id).with_reveal_all_normalized(tcx), as this method is more efficient.

is_copy_raw = 164

Trait selection queries. These are best used by invoking ty.is_copy_modulo_regions(), ty.is_copy(), etc, since that will prune the environment where possible.

is_sized_raw = 165

Query backing Ty::is_sized.

is_freeze_raw = 166

Query backing Ty::is_freeze.

is_unpin_raw = 167

Query backing Ty::is_unpin.

needs_drop_raw = 168

Query backing Ty::needs_drop.

has_significant_drop_raw = 169

Query backing Ty::has_significant_drop_raw.

has_structural_eq_impl = 170

Query backing Ty::is_structural_eq_shallow.

This is only correct for ADTs. Call is_structural_eq_shallow to handle all types correctly.

adt_drop_tys = 171

A list of types where the ADT requires drop if and only if any of those types require drop. If the ADT is known to always need drop then Err(AlwaysRequiresDrop) is returned.

adt_significant_drop_tys = 172

A list of types where the ADT requires drop if and only if any of those types has significant drop. A type marked with the attribute rustc_insignificant_dtor is considered to not be significant. A drop is significant if it is implemented by the user or does anything that will have any observable behavior (other than freeing up memory). If the ADT is known to have a significant destructor then Err(AlwaysRequiresDrop) is returned.

layout_of = 173

Computes the layout of a type. Note that this implicitly executes in “reveal all” mode, and will normalize the input type.

fn_abi_of_fn_ptr = 174

Compute a FnAbi suitable for indirect calls, i.e. to fn pointers.

NB: this doesn’t handle virtual calls - those should use fn_abi_of_instance instead, where the instance is an InstanceDef::Virtual.

fn_abi_of_instance = 175

Compute a FnAbi suitable for declaring/defining an fn instance, and for direct calls to an fn.

NB: that includes virtual calls, which are represented by “direct calls” to an InstanceDef::Virtual instance (of <dyn Trait as Trait>::fn).

dylib_dependency_formats = 176

[query description - consider adding a doc-comment!] getting dylib dependency formats of crate

dependency_formats = 177

[query description - consider adding a doc-comment!] getting the linkage format of all dependencies

is_compiler_builtins = 178

[query description - consider adding a doc-comment!] checking if the crate is_compiler_builtins

has_global_allocator = 179

[query description - consider adding a doc-comment!] checking if the crate has_global_allocator

has_alloc_error_handler = 180

[query description - consider adding a doc-comment!] checking if the crate has_alloc_error_handler

has_panic_handler = 181

[query description - consider adding a doc-comment!] checking if the crate has_panic_handler

is_profiler_runtime = 182

[query description - consider adding a doc-comment!] checking if a crate is #![profiler_runtime]

has_ffi_unwind_calls = 183

[query description - consider adding a doc-comment!] checking if tcx.def_path_str(key) contains FFI-unwind calls

required_panic_strategy = 184

[query description - consider adding a doc-comment!] getting a crate’s required panic strategy

panic_in_drop_strategy = 185

[query description - consider adding a doc-comment!] getting a crate’s configured panic-in-drop strategy

is_no_builtins = 186

[query description - consider adding a doc-comment!] getting whether a crate has #![no_builtins]

symbol_mangling_version = 187

[query description - consider adding a doc-comment!] getting a crate’s symbol mangling version

extern_crate = 188

[query description - consider adding a doc-comment!] getting crate’s ExternCrateData

specializes = 189

[query description - consider adding a doc-comment!] computing whether impls specialize one another

in_scope_traits_map = 190

[query description - consider adding a doc-comment!] getting traits in scope at a block

defaultness = 191

Returns whether the impl or associated function has the default keyword.

check_well_formed = 192

[query description - consider adding a doc-comment!] checking that tcx.def_path_str(key) is well-formed

reachable_non_generics = 193

[query description - consider adding a doc-comment!] looking up the exported symbols of a crate

is_reachable_non_generic = 194

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is an exported symbol

is_unreachable_local_definition = 195

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is reachable from outside the crate

upstream_monomorphizations = 196

The entire set of monomorphizations the local crate can safely link to because they are exported from upstream crates. Do not depend on this directly, as its value changes anytime a monomorphization gets added or removed in any upstream crate. Instead use the narrower upstream_monomorphizations_for, upstream_drop_glue_for, or, even better, Instance::upstream_monomorphization().

upstream_monomorphizations_for = 197

Returns the set of upstream monomorphizations available for the generic function identified by the given def_id. The query makes sure to make a stable selection if the same monomorphization is available in multiple upstream crates.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

upstream_drop_glue_for = 198

Returns the upstream crate that exports drop-glue for the given type (args is expected to be a single-item list containing the type one wants drop-glue for).

This is a subset of upstream_monomorphizations_for in order to increase dep-tracking granularity. Otherwise adding or removing any type with drop-glue in any upstream crate would invalidate all functions calling drop-glue of an upstream type.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

NOTE: This query could easily be extended to also support other common functions that have are large set of monomorphizations (like Clone::clone for example).

foreign_modules = 199

Returns a list of all extern blocks of a crate.

clashing_extern_declarations = 200

Lint against extern fn declarations having incompatible types.

entry_fn = 201

Identifies the entry-point (e.g., the main function) for a given crate, returning None if there is no entry point (such as for library crates).

proc_macro_decls_static = 202

Finds the rustc_proc_macro_decls item of a crate.

crate_hash = 203

[query description - consider adding a doc-comment!] looking up the hash a crate

crate_host_hash = 204

Gets the hash for the host proc macro. Used to support -Z dual-proc-macro.

extra_filename = 205

Gets the extra data to put in each output filename for a crate. For example, compiling the foo crate with extra-filename=-a creates a libfoo-b.rlib file.

crate_extern_paths = 206

Gets the paths where the crate came from in the file system.

implementations_of_trait = 207

Given a crate and a trait, look up all impls of that trait in the crate. Return (impl_id, self_ty).

crate_incoherent_impls = 208

Collects all incoherent impls for the given crate and type.

Do not call this directly, but instead use the incoherent_impls query. This query is only used to get the data necessary for that query.

native_library = 209

Get the corresponding native library from the native_libraries query

resolve_bound_vars = 210

Does lifetime resolution on items. Importantly, we can’t resolve lifetimes directly on things like trait methods, because of trait params. See rustc_resolve::late::lifetimes for details.

named_variable_map = 211

[query description - consider adding a doc-comment!] looking up a named region

is_late_bound_map = 212

[query description - consider adding a doc-comment!] testing if a region is late bound

object_lifetime_default = 213

For a given item’s generic parameter, gets the default lifetimes to be used for each parameter if a trait object were to be passed for that parameter. For example, for T in struct Foo<'a, T>, this would be 'static. For T in struct Foo<'a, T: 'a>, this would instead be 'a. This query will panic if passed something that is not a type parameter.

late_bound_vars_map = 214

[query description - consider adding a doc-comment!] looking up late bound vars

visibility = 215

Computes the visibility of the provided def_id.

If the item from the def_id doesn’t have a visibility, it will panic. For example a generic type parameter will panic if you call this method on it:

use std::fmt::Debug;

pub trait Foo<T: Debug> {}

In here, if you call visibility on T, it’ll panic.

inhabited_predicate_adt = 216

[query description - consider adding a doc-comment!] computing the uninhabited predicate of {:?}

inhabited_predicate_type = 217

Do not call this query directly: invoke Ty::inhabited_predicate instead.

dep_kind = 218

[query description - consider adding a doc-comment!] fetching what a dependency looks like

crate_name = 219

Gets the name of the crate.

module_children = 220

[query description - consider adding a doc-comment!] collecting child items of module tcx.def_path_str(def_id)

extern_mod_stmt_cnum = 221

[query description - consider adding a doc-comment!] computing crate imported by tcx.def_path_str(def_id)

lib_features = 222

[query description - consider adding a doc-comment!] calculating the lib features defined in a crate

stability_implications = 223

[query description - consider adding a doc-comment!] calculating the implications between #[unstable] features defined in a crate

intrinsic_raw = 224

Whether the function is an intrinsic

get_lang_items = 225

Returns the lang items defined in another crate by loading it from metadata.

all_diagnostic_items = 226

Returns all diagnostic items defined in all crates.

defined_lang_items = 227

Returns the lang items defined in another crate by loading it from metadata.

diagnostic_items = 228

Returns the diagnostic items defined in a crate.

missing_lang_items = 229

[query description - consider adding a doc-comment!] calculating the missing lang items in a crate

visible_parent_map = 230

The visible parent map is a map from every item to a visible parent. It prefers the shortest visible path to an item. Used for diagnostics, for example path trimming. The parents are modules, enums or traits.

trimmed_def_paths = 231

Collects the “trimmed”, shortest accessible paths to all items for diagnostics. See the provider docs for more info.

missing_extern_crate_item = 232

[query description - consider adding a doc-comment!] seeing if we’re missing an extern crate item for this crate

used_crate_source = 233

[query description - consider adding a doc-comment!] looking at the source for a crate

debugger_visualizers = 234

Returns the debugger visualizers defined for this crate. NOTE: This query has to be marked eval_always because it reads data directly from disk that is not tracked anywhere else. I.e. it represents a genuine input to the query system.

postorder_cnums = 235

[query description - consider adding a doc-comment!] generating a postorder list of CrateNums

is_private_dep = 236

Returns whether or not the crate with CrateNum ‘cnum’ is marked as a private dependency

allocator_kind = 237

[query description - consider adding a doc-comment!] getting the allocator kind for the current crate

alloc_error_handler_kind = 238

[query description - consider adding a doc-comment!] alloc error handler kind for the current crate

upvars_mentioned = 239

[query description - consider adding a doc-comment!] collecting upvars mentioned in tcx.def_path_str(def_id)

maybe_unused_trait_imports = 240

[query description - consider adding a doc-comment!] fetching potentially unused trait imports

names_imported_by_glob_use = 241

[query description - consider adding a doc-comment!] finding names imported by glob use for tcx.def_path_str(def_id)

stability_index = 242

[query description - consider adding a doc-comment!] calculating the stability index for the local crate

crates = 243

[query description - consider adding a doc-comment!] fetching all foreign CrateNum instances

used_crates = 244

[query description - consider adding a doc-comment!] fetching CrateNums for all crates loaded non-speculatively

traits = 245

A list of all traits in a crate, used by rustdoc and error reporting.

trait_impls_in_crate = 246

[query description - consider adding a doc-comment!] fetching all trait impls in a crate

exported_symbols = 247

The list of symbols exported from the given crate.

All names contained in exported_symbols(cnum) are guaranteed to correspond to a publicly visible symbol in cnum machine code.

The exported_symbols sets of different crates do not intersect.

collect_and_partition_mono_items = 248

[query description - consider adding a doc-comment!] collect_and_partition_mono_items

is_codegened_item = 249

[query description - consider adding a doc-comment!] determining whether tcx.def_path_str(def_id) needs codegen

codegen_unit = 250

[query description - consider adding a doc-comment!] getting codegen unit {sym}

unused_generic_params = 251

[query description - consider adding a doc-comment!] determining which generic parameters are unused by tcx.def_path_str(key.def_id())

backend_optimization_level = 252

[query description - consider adding a doc-comment!] optimization level used by backend

output_filenames = 253

Return the filenames where output artefacts shall be stored.

This query returns an &Arc because codegen backends need the value even after the TyCtxt has been destroyed.

normalize_canonicalized_projection_ty = 254

Do not call this query directly: Invoke normalize instead.

normalize_canonicalized_weak_ty = 255

Do not call this query directly: Invoke normalize instead.

normalize_canonicalized_inherent_projection_ty = 256

Do not call this query directly: Invoke normalize instead.

try_normalize_generic_arg_after_erasing_regions = 257

Do not call this query directly: invoke try_normalize_erasing_regions instead.

implied_outlives_bounds_compat = 258

[query description - consider adding a doc-comment!] computing implied outlives bounds for goal.value.value

implied_outlives_bounds = 259

[query description - consider adding a doc-comment!] computing implied outlives bounds v2 for goal.value.value

dropck_outlives = 260

Do not call this query directly: invoke DropckOutlives::new(dropped_ty)).fully_perform(typeck.infcx) instead.

evaluate_obligation = 261

Do not call this query directly: invoke infcx.predicate_may_hold() or infcx.predicate_must_hold() instead.

type_op_ascribe_user_type = 262

Do not call this query directly: part of the Eq type-op

type_op_eq = 263

Do not call this query directly: part of the Eq type-op

type_op_subtype = 264

Do not call this query directly: part of the Subtype type-op

type_op_prove_predicate = 265

Do not call this query directly: part of the ProvePredicate type-op

type_op_normalize_ty = 266

Do not call this query directly: part of the Normalize type-op

type_op_normalize_clause = 267

Do not call this query directly: part of the Normalize type-op

type_op_normalize_poly_fn_sig = 268

Do not call this query directly: part of the Normalize type-op

type_op_normalize_fn_sig = 269

Do not call this query directly: part of the Normalize type-op

instantiate_and_check_impossible_predicates = 270

[query description - consider adding a doc-comment!] checking impossible instantiated predicates: tcx.def_path_str(key.0)

is_impossible_associated_item = 271

[query description - consider adding a doc-comment!] checking if tcx.def_path_str(key.1) is impossible to reference within tcx.def_path_str(key.0)

method_autoderef_steps = 272

[query description - consider adding a doc-comment!] computing autoderef types for goal.value.value

supported_target_features = 273

[query description - consider adding a doc-comment!] looking up supported target features

features_query = 274

[query description - consider adding a doc-comment!] looking up enabled feature gates

crate_for_resolver = 275

[query description - consider adding a doc-comment!] the ast before macro expansion and name resolution

resolve_instance = 276

Attempt to resolve the given DefId to an Instance, for the given generics args (GenericArgsRef), returning one of:

Ok(Some(instance)) on success

Ok(None) when the GenericArgsRef are still too generic, and therefore don’t allow finding the final Instance

Err(ErrorGuaranteed) when the Instance resolution process couldn’t complete due to errors elsewhere - this is distinct from Ok(None) to avoid misleading diagnostics when an error has already been/will be emitted, for the original cause

reveal_opaque_types_in_bounds = 277

[query description - consider adding a doc-comment!] revealing opaque types in {:?}

limits = 278

[query description - consider adding a doc-comment!] looking up limits

diagnostic_hir_wf_check = 279

Performs an HIR-based well-formed check on the item with the given HirId. If we get an Unimplemented error that matches the provided Predicate, return the cause of the newly created obligation.

This is only used by error-reporting code to get a better cause (in particular, a better span) for an existing error. Therefore, it is best-effort, and may never handle all of the cases that the normal ty::Ty-based wfcheck does. This is fine, because the ty::Ty-based wfcheck is always run.

global_backend_features = 280

The list of backend features computed from CLI flags (-Ctarget-cpu, -Ctarget-feature, --target and similar).

check_validity_requirement = 281

[query description - consider adding a doc-comment!] checking validity requirement for key.1.value : key.0

compare_impl_const = 282

[query description - consider adding a doc-comment!] checking assoc const tcx.def_path_str(key.0) has the same type as trait item

deduced_param_attrs = 283

[query description - consider adding a doc-comment!] deducing parameter attributes for tcx.def_path_str(def_id)

doc_link_resolutions = 284

[query description - consider adding a doc-comment!] resolutions for documentation links for a module

doc_link_traits_in_scope = 285

[query description - consider adding a doc-comment!] traits in scope for documentation links for a module

check_tys_might_be_eq = 286

Used in super_combine_consts to ICE if the type of the two consts are definitely not going to end up being equal to eachother. This might return Ok even if the types are not equal, but will never return Err if the types might be equal.

stripped_cfg_items = 287

Get all item paths that were stripped by a #[cfg] in a particular crate. Should not be called for the local crate before the resolver outputs are created, as it is only fed there.

generics_require_sized_self = 288

[query description - consider adding a doc-comment!] check whether the item has a where Self: Sized bound

cross_crate_inlinable = 289

[query description - consider adding a doc-comment!] whether the item should be made inlinable across crates

find_field = 290

[query description - consider adding a doc-comment!] find the index of maybe nested field {ident} in tcx.def_path_str(def_id)