Trait rustc_middle::dep_graph::dep_node::DepNodeParams

pub trait DepNodeParams<Tcx>: Debug + Sizedwhere
    Tcx: DepContext,{
    // Required methods
    fn fingerprint_style() -> FingerprintStyle;
    fn recover(
        tcx: Tcx,
        dep_node: &DepNode<<Tcx as DepContext>::DepKind>
    ) -> Option<Self>;

    // Provided methods
    fn to_fingerprint(&self, _: Tcx) -> Fingerprint { ... }
    fn to_debug_str(&self, _: Tcx) -> String { ... }
}

Required Methods

fn fingerprint_style() -> FingerprintStyle

fn recover( tcx: Tcx, dep_node: &DepNode<<Tcx as DepContext>::DepKind> ) -> Option<Self>

This method tries to recover the query key from the given DepNode, something which is needed when forcing DepNodes during red-green evaluation. The query system will only call this method if fingerprint_style() is not FingerprintStyle::Opaque. It is always valid to return None here, in which case incremental compilation will treat the query as having changed instead of forcing it.

Provided Methods

fn to_fingerprint(&self, _: Tcx) -> Fingerprint

This method turns the parameters of a DepNodeConstructor into an opaque Fingerprint to be used in DepNode. Not all DepNodeParams support being turned into a Fingerprint (they don’t need to if the corresponding DepNode is anonymous).

fn to_debug_str(&self, _: Tcx) -> String