rustc_middle::hir
Module map

Structs

ItemCollector ğŸ”’

Map

ParentHirIterator ğŸ”’

An iterator that walks up the ancestor tree of a given HirId. Constructed using tcx.hir().parent_iter(hir_id).

ParentOwnerIterator

An iterator that walks up the ancestor tree of a given HirId. Constructed using tcx.hir().parent_owner_iter(hir_id).

Functions

crate_hash ğŸ”’

hir_crate_items ğŸ”’

hir_id_to_string ğŸ”’

hir_module_items ğŸ”’

upstream_crates ğŸ”’