rustc_middle::hir

Module map

This module used to contain a type called Map. That type has since been eliminated, and all its methods are now on TyCtxt. But the module name stays as map because there isn’t an obviously better name for it.

Structs

ItemCollector 🔒

ParentHirIterator 🔒

An iterator that walks up the ancestor tree of a given HirId. Constructed using tcx.hir_parent_iter(hir_id).

ParentOwnerIterator

An iterator that walks up the ancestor tree of a given HirId. Constructed using tcx.hir_parent_owner_iter(hir_id).

Functions

crate_hash 🔒

hir_crate_items 🔒

hir_module_items 🔒

upstream_crates 🔒