rustc_middle
Module infer

Modules

canonical

Canonicalization is the key to constructing a query in the middle of type inference. Ordinarily, it is not possible to store types from type inference in query keys, because they contain references to inference variables whose lifetimes are too short and so forth. Canonicalizing a value T1 using canonicalize_query produces two things:

unify_key

Structs

MemberConstraint

Requires that region must be equal to one of the regions in choice_regions. We often denote this using the syntax: