rustc_middle

Module middle

Modules

codegen_fn_attrs

debugger_visualizer

dependency_format

Type definitions for learning about the dependency formats of all upstream crates (rlibs/dylibs/oh my).

exported_symbols

lang_items

Detecting lang items.

lib_features

privacy

A pass that checks to make sure private fields and methods aren’t used outside their scopes. This pass will also generate a set of exported items which are available for use externally when compiled as a library.

region

This file declares the ScopeTree type, which describes the parent links in the region hierarchy.

resolve_bound_vars

Name resolution for lifetimes and late-bound type and const variables: type declarations.

stability

A pass that annotates every item and method with its stability level, propagating default levels lexically from parent to children ast nodes.