rustc_middle::middle

Module region

This file declares the ScopeTree type, which describes the parent links in the region hierarchy.

For more information about how MIR-based region-checking works, see the rustc dev guide.

Structs

FirstStatementIndex

Represents a subscope of block for a binding that is introduced by block.stmts[first_statement_index]. Such subscopes represent a suffix of the block. Note that each subscope does not include the initializer expression, if any, for the statement indexed by first_statement_index.

RvalueCandidate

See the rvalue_candidates field for more information on rvalue candidates in general. The lifetime field is None to indicate that certain expressions escape into ’static and should have no local cleanup scope.

Scope

Represents a statically-describable scope that can be used to bound the lifetime/region for values.

ScopeTree

The region scope tree encodes information about region relationships.

Enums

ScopeData