Module rustc_middle::mir::coverage

Metadata from source code coverage analysis and instrumentation.

Structs

BlockMarkerId

Used by CoverageKind::BlockMarker to mark blocks during THIR-to-MIR lowering, so that those blocks can be identified later.

BranchSpan

CodeRegion

ConditionId

ID of a mcdc condition. Used by llvm to check mcdc coverage.

ConditionInfo

CounterId

ID of a coverage counter. Values ascend from 0.

CoverageInfoHi

Coverage information for a function, recorded during MIR building and attached to the corresponding mir::Body. Used by the InstrumentCoverage MIR pass.

DecisionInfo

Expression

ExpressionId

ID of a coverage-counter expression. Values ascend from 0.

FunctionCoverageInfo

Stores per-function coverage information attached to a mir::Body, to be used in conjunction with the individual coverage statements injected into the function’s basic blocks.

MCDCBranchSpan

MCDCDecisionSpan

Mapping

Enums

CovTerm

Enum that can hold a constant zero value, the ID of an physical coverage counter, or the ID of a coverage-counter expression.

CoverageKind

MappingKind

Op