rustc_middle::mir

Module coverage

Metadata from source code coverage analysis and instrumentation.

Structs

BasicCoverageBlock

During the InstrumentCoverage MIR pass, a BCB is a node in the “coverage graph”, which is a refinement of the MIR control-flow graph that merges or omits some blocks that aren’t relevant to coverage.

BlockMarkerId

Used by CoverageKind::BlockMarker to mark blocks during THIR-to-MIR lowering, so that those blocks can be identified later.

BranchSpan

ConditionId

ID of a mcdc condition. Used by llvm to check mcdc coverage.

ConditionInfo

CounterId

ID of a coverage counter. Values ascend from 0.

CoverageIdsInfo

Contains information needed during codegen, obtained by inspecting the function’s MIR after MIR optimizations.

CoverageInfoHi

Coverage information for a function, recorded during MIR building and attached to the corresponding mir::Body. Used by the InstrumentCoverage MIR pass.

DecisionInfo

Expression

ExpressionId

ID of a coverage-counter expression. Values ascend from 0.

FunctionCoverageInfo

Stores per-function coverage information attached to a mir::Body, to be used in conjunction with the individual coverage statements injected into the function’s basic blocks.

MCDCBranchSpan

MCDCDecisionSpan

Mapping

NodeFlowData

Data representing a view of some underlying graph, in which each node’s successors have been merged into a single “supernode”.

Enums

CovTerm

Enum that can hold a constant zero value, the ID of an physical coverage counter, or the ID of a coverage-counter expression.

CoverageKind

MappingKind

Op

Constants

START_BCB