Struct rustc_middle::mir::interpret::Allocation

pub struct Allocation<Prov: Provenance = AllocId, Extra = (), Bytes = Box<[u8]>> {
    bytes: Bytes,
    provenance: ProvenanceMap<Prov>,
    init_mask: InitMask,
    pub align: Align,
    pub mutability: Mutability,
    pub extra: Extra,
}

This type represents an Allocation in the Miri/CTFE core engine.

Its public API is rather low-level, working directly with allocation offsets and a custom error type to account for the lack of an AllocId on this level. The Miri/CTFE core engine memory module provides higher-level access.

Fields
bytes: Bytes

The actual bytes of the allocation. Note that the bytes of a pointer represent the offset of the pointer.

provenance: ProvenanceMap<Prov>

Maps from byte addresses to extra provenance data for each pointer. Only the first byte of a pointer is inserted into the map; i.e., every entry in this map applies to pointer_size consecutive bytes starting at the given offset.

init_mask: InitMask

Denotes which part of this allocation is initialized.

align: Align

The alignment of the allocation to detect unaligned reads. (Align guarantees that this is a power of two.)

mutability: Mutability

true if the allocation is mutable. Also used by codegen to determine if a static should be put into mutable memory, which happens for static mut and static with interior mutability.

extra: Extra

Extra state for the machine.

Implementations

impl<Prov: Provenance, Bytes: AllocBytes> Allocation<Prov, (), Bytes>

pub fn from_raw_bytes( bytes: Bytes, align: Align, mutability: Mutability ) -> Self

Creates an allocation from an existing Bytes value - this is needed for miri FFI support

pub fn from_bytes<'a>( slice: impl Into<Cow<'a, [u8]>>, align: Align, mutability: Mutability ) -> Self

Creates an allocation initialized by the given bytes

pub fn from_bytes_byte_aligned_immutable<'a>( slice: impl Into<Cow<'a, [u8]>> ) -> Self

fn uninit_inner<R>( size: Size, align: Align, fail: impl FnOnce() -> R ) -> Result<Self, R>

pub fn try_uninit<'tcx>(size: Size, align: Align) -> InterpResult<'tcx, Self>

Try to create an Allocation of size bytes, failing if there is not enough memory available to the compiler to do so.

pub fn uninit(size: Size, align: Align) -> Self

Try to create an Allocation of size bytes, panics if there is not enough memory available to the compiler to do so.

Example use case: To obtain an Allocation filled with specific data, first call this function and then call write_scalar to fill in the right data.

impl<Bytes: AllocBytes> Allocation<AllocId, (), Bytes>

pub fn adjust_from_tcx<Prov: Provenance, Extra, Err>( self, cx: &impl HasDataLayout, extra: Extra, adjust_ptr: impl FnMut(Pointer<AllocId>) -> Result<Pointer<Prov>, Err> ) -> Result<Allocation<Prov, Extra, Bytes>, Err>

Adjust allocation from the ones in tcx to a custom Machine instance with a different Provenance and Extra type.

impl<Prov: Provenance, Extra, Bytes: AllocBytes> Allocation<Prov, Extra, Bytes>

Raw accessors. Provide access to otherwise private bytes.

pub fn len(&self) -> usize

pub fn size(&self) -> Size

pub fn inspect_with_uninit_and_ptr_outside_interpreter( &self, range: Range<usize> ) -> &[u8]

Looks at a slice which may contain uninitialized bytes or provenance. This differs from get_bytes_with_uninit_and_ptr in that it does no provenance checks (even on the edges) at all. This must not be used for reads affecting the interpreter execution.

pub fn init_mask(&self) -> &InitMask

Returns the mask indicating which bytes are initialized.

pub fn provenance(&self) -> &ProvenanceMap<Prov>

Returns the provenance map.

impl<Prov: Provenance, Extra, Bytes: AllocBytes> Allocation<Prov, Extra, Bytes>

Byte accessors.

pub fn base_addr(&self) -> *const u8

pub fn get_bytes_unchecked(&self, range: AllocRange) -> &[u8]

This is the entirely abstraction-violating way to just grab the raw bytes without caring about provenance or initialization.

This function also guarantees that the resulting pointer will remain stable even when new allocations are pushed to the HashMap. mem_copy_repeatedly relies on that.

pub fn get_bytes_strip_provenance( &self, cx: &impl HasDataLayout, range: AllocRange ) -> AllocResult<&[u8]>

Checks that these bytes are initialized, and then strip provenance (if possible) and return them.

It is the caller’s responsibility to check bounds and alignment beforehand. Most likely, you want to use the PlaceTy and OperandTy-based methods on InterpCx instead.

pub fn get_bytes_mut( &mut self, cx: &impl HasDataLayout, range: AllocRange ) -> AllocResult<&mut [u8]>

Just calling this already marks everything as defined and removes provenance, so be sure to actually put data there!

It is the caller’s responsibility to check bounds and alignment beforehand. Most likely, you want to use the PlaceTy and OperandTy-based methods on InterpCx instead.

pub fn get_bytes_mut_ptr( &mut self, cx: &impl HasDataLayout, range: AllocRange ) -> AllocResult<*mut [u8]>

A raw pointer variant of get_bytes_mut that avoids invalidating existing aliases into this memory.

impl<Prov: Provenance, Extra, Bytes: AllocBytes> Allocation<Prov, Extra, Bytes>

Reading and writing.

fn mark_init(&mut self, range: AllocRange, is_init: bool)

Sets the init bit for the given range.

pub fn read_scalar( &self, cx: &impl HasDataLayout, range: AllocRange, read_provenance: bool ) -> AllocResult<Scalar<Prov>>

Reads a non-ZST scalar.

If read_provenance is true, this will also read provenance; otherwise (if the machine supports that) provenance is entirely ignored.

ZSTs can’t be read because in order to obtain a Pointer, we need to check for ZSTness anyway due to integer pointers being valid for ZSTs.

It is the caller’s responsibility to check bounds and alignment beforehand. Most likely, you want to call InterpCx::read_scalar instead of this method.

pub fn write_scalar( &mut self, cx: &impl HasDataLayout, range: AllocRange, val: Scalar<Prov> ) -> AllocResult

Writes a non-ZST scalar.

ZSTs can’t be read because in order to obtain a Pointer, we need to check for ZSTness anyway due to integer pointers being valid for ZSTs.

It is the caller’s responsibility to check bounds and alignment beforehand. Most likely, you want to call InterpCx::write_scalar instead of this method.

pub fn write_uninit( &mut self, cx: &impl HasDataLayout, range: AllocRange ) -> AllocResult

Write “uninit” to the given memory range.

pub fn provenance_apply_copy(&mut self, copy: ProvenanceCopy<Prov>)

Applies a previously prepared provenance copy. The affected range, as defined in the parameters to provenance().prepare_copy is expected to be clear of provenance.

This is dangerous to use as it can violate internal Allocation invariants! It only exists to support an efficient implementation of mem_copy_repeatedly.

pub fn init_mask_apply_copy( &mut self, copy: InitCopy, range: AllocRange, repeat: u64 )

Applies a previously prepared copy of the init mask.

This is dangerous to use as it can violate internal Allocation invariants! It only exists to support an efficient implementation of mem_copy_repeatedly.