Module rustc_middle::mir::query

Values computed by queries that use MIR.

Structs

BorrowCheckResult

ClosureOutlivesRequirement

Indicates an outlives-constraint between a type or between two free regions declared on the closure.

ClosureOutlivesSubjectTy

Represents a ty::Ty for use in ClosureOutlivesSubject.

ClosureRegionRequirements

After we borrow check a closure, we are left with various requirements that we have inferred between the free regions that appear in the closure’s signature or on its field types. These requirements are then verified and proved by the closure’s creating function. This struct encodes those requirements.

ConstQualifs

The result of the mir_const_qualif query.

CoroutineLayout

The layout of coroutine state.

CoroutineSavedLocal

CoroutineSavedTy

CoverageIdsInfo

Summarizes coverage IDs inserted by the InstrumentCoverage MIR pass (for compiler option -Cinstrument-coverage), after MIR optimizations have had a chance to potentially remove some of them.

DestructuredConstant

The constituent parts of a mir constant of kind ADT or array.

UnsafetyCheckResult

UnsafetyViolation

Enums

ClosureOutlivesSubject

The subject of a ClosureOutlivesRequirement – that is, the thing that must outlive some region.

ConstraintCategory

Outlives-constraints can be categorized to determine whether and why they are interesting (for error reporting). Order of variants indicates sort order of the category, thereby influencing diagnostic output.

ReturnConstraint

UnsafetyViolationDetails

UnsafetyViolationKind

UnusedUnsafe