rustc_middle::mir::traversal
Function mono_reachable

pub fn mono_reachable<'a, 'tcx>(
    body: &'a Body<'tcx>,
    tcx: TyCtxt<'tcx>,
    instance: Instance<'tcx>,
) -> MonoReachable<'a, 'tcx> 

Traversal of a Body that tries to avoid unreachable blocks in a monomorphized Instance.

This is allowed to have false positives; blocks may be visited even if they are not actually reachable.

Such a traversal is mostly useful because it lets us skip lowering the false side of if <T as Trait>::CONST, as well as NullOp::UbChecks.