rustc_middle::query::on_disk_cache
Struct CacheEncoder

pub struct CacheEncoder<'a, 'tcx> {
    tcx: TyCtxt<'tcx>,
    encoder: FileEncoder,
    type_shorthands: FxHashMap<Ty<'tcx>, usize>,
    predicate_shorthands: FxHashMap<PredicateKind<'tcx>, usize>,
    interpret_allocs: FxIndexSet<AllocId>,
    source_map: CachingSourceMapView<'tcx>,
    file_to_file_index: FxHashMap<*const SourceFile, SourceFileIndex>,
    hygiene_context: &'a HygieneEncodeContext,
    symbol_table: FxHashMap<Symbol, usize>,
}

An encoder that can write to the incremental compilation cache.

Fields
tcx: TyCtxt<'tcx>encoder: FileEncodertype_shorthands: FxHashMap<Ty<'tcx>, usize>predicate_shorthands: FxHashMap<PredicateKind<'tcx>, usize>interpret_allocs: FxIndexSet<AllocId>source_map: CachingSourceMapView<'tcx>file_to_file_index: FxHashMap<*const SourceFile, SourceFileIndex>hygiene_context: &'a HygieneEncodeContextsymbol_table: FxHashMap<Symbol, usize>
Implementations

impl<'a, 'tcx> CacheEncoder<'a, 'tcx>

fn source_file_index(&mut self, source_file: Lrc<SourceFile>) -> SourceFileIndex

pub fn encode_tagged<T: Encodable<Self>, V: Encodable<Self>>( &mut self, tag: T, value: &V, )

Encode something with additional information that allows to do some sanity checks when decoding the data again. This method will first encode the specified tag, then the given value, then the number of bytes taken up by tag and value. On decoding, we can then verify that we get the expected tag and read the expected number of bytes.

fn finish(self) -> FileEncodeResult