Type Alias rustc_middle::query::queries::incoherent_impls::LocalKey

pub type LocalKey<'tcx> = SimplifiedType;

Aliased Type

enum LocalKey<'tcx> {

    Bool,
    Char,
    Int(IntTy),
    Uint(UintTy),
    Float(FloatTy),
    Adt(DefId),
    Foreign(DefId),
    Str,
    Array,
    Slice,
    Ref(Mutability),
    Ptr(Mutability),
    Never,
    Tuple(usize),
    MarkerTraitObject,
    Trait(DefId),
    Closure(DefId),
    Coroutine(DefId),
    CoroutineWitness(DefId),
    Function(usize),
    Placeholder,
    Error,

}

Variants

Bool

Char

Int(IntTy)

Uint(UintTy)

Float(FloatTy)

Adt(DefId)

Foreign(DefId)

Str

Array

Slice

Ref(Mutability)

Ptr(Mutability)

Never

Tuple(usize)

MarkerTraitObject

A trait object, all of whose components are markers (e.g., dyn Send + Sync).

Trait(DefId)

Closure(DefId)

Coroutine(DefId)

CoroutineWitness(DefId)

Function(usize)

Placeholder

Error

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 16 bytes

Size for each variant:

Bool: 0 bytes

Char: 0 bytes

Int: 1 byte

Uint: 1 byte

Float: 1 byte

Adt: 11 bytes

Foreign: 11 bytes

Str: 0 bytes

Array: 0 bytes

Slice: 0 bytes

Ref: 1 byte

Ptr: 1 byte

Never: 0 bytes

Tuple: 15 bytes

MarkerTraitObject: 0 bytes

Trait: 11 bytes

Closure: 11 bytes

Coroutine: 11 bytes

CoroutineWitness: 11 bytes

Function: 15 bytes

Placeholder: 0 bytes

Error: 0 bytes