Struct rustc_middle::query::QueryArenas

pub struct QueryArenas<'tcx> {

    pub trigger_delayed_bug: (),
    pub registered_tools: TypedArena<IndexSet<Ident, BuildHasherDefault<FxHasher>>>,
    pub early_lint_checks: (),
    pub resolutions: (),
    pub resolver_for_lowering_raw: (),
    pub source_span: (),
    pub hir_crate: TypedArena<Crate<'tcx>>,
    pub hir_crate_items: TypedArena<ModuleItems>,
    pub hir_module_items: TypedArena<ModuleItems>,
    pub local_def_id_to_hir_id: (),
    pub hir_owner_parent: (),
    pub opt_hir_owner_nodes: (),
    pub hir_attrs: (),
    pub const_param_default: (),
    pub type_of: (),
    pub type_of_opaque: (),
    pub type_alias_is_lazy: (),
    pub collect_return_position_impl_trait_in_trait_tys: (),
    pub is_type_alias_impl_trait: (),
    pub unsizing_params_for_adt: TypedArena<BitSet<u32>>,
    pub analysis: (),
    pub check_expectations: (),
    pub generics_of: TypedArena<Generics>,
    pub predicates_of: (),
    pub opaque_types_defined_by: (),
    pub explicit_item_bounds: (),
    pub item_bounds: (),
    pub native_libraries: TypedArena<Vec<NativeLib>>,
    pub shallow_lint_levels_on: TypedArena<ShallowLintLevelMap>,
    pub lint_expectations: TypedArena<Vec<(LintExpectationId, LintExpectation)>>,
    pub expn_that_defined: (),
    pub is_panic_runtime: (),
    pub representability: (),
    pub representability_adt_ty: (),
    pub params_in_repr: TypedArena<BitSet<u32>>,
    pub thir_body: (),
    pub thir_tree: TypedArena<String>,
    pub thir_flat: TypedArena<String>,
    pub mir_keys: TypedArena<IndexSet<LocalDefId, BuildHasherDefault<FxHasher>>>,
    pub mir_const_qualif: (),
    pub mir_built: (),
    pub mir_const: (),
    pub thir_abstract_const: (),
    pub mir_drops_elaborated_and_const_checked: (),
    pub mir_for_ctfe: (),
    pub mir_promoted: (),
    pub closure_typeinfo: (),
    pub closure_saved_names_of_captured_variables: TypedArena<IndexVec<FieldIdx, Symbol>>,
    pub mir_coroutine_witnesses: TypedArena<Option<CoroutineLayout<'tcx>>>,
    pub check_coroutine_obligations: (),
    pub optimized_mir: (),
    pub coverage_ids_info: TypedArena<CoverageIdsInfo>,
    pub promoted_mir: (),
    pub erase_regions_ty: (),
    pub wasm_import_module_map: TypedArena<UnordMap<DefId, String>>,
    pub predicates_defined_on: (),
    pub trait_explicit_predicates_and_bounds: (),
    pub explicit_predicates_of: (),
    pub inferred_outlives_of: (),
    pub super_predicates_of: (),
    pub implied_predicates_of: (),
    pub super_predicates_that_define_assoc_item: (),
    pub type_param_predicates: (),
    pub trait_def: TypedArena<TraitDef>,
    pub adt_def: (),
    pub adt_destructor: (),
    pub adt_sized_constraint: (),
    pub adt_dtorck_constraint: (),
    pub constness: (),
    pub asyncness: (),
    pub is_promotable_const_fn: (),
    pub coroutine_kind: (),
    pub coroutine_for_closure: (),
    pub crate_variances: TypedArena<CrateVariancesMap<'tcx>>,
    pub variances_of: (),
    pub inferred_outlives_crate: TypedArena<CratePredicatesMap<'tcx>>,
    pub associated_item_def_ids: (),
    pub associated_item: (),
    pub associated_items: TypedArena<AssocItems>,
    pub impl_item_implementor_ids: TypedArena<UnordMap<DefId, DefId>>,
    pub associated_types_for_impl_traits_in_associated_fn: (),
    pub associated_type_for_impl_trait_in_trait: (),
    pub impl_trait_header: (),
    pub issue33140_self_ty: (),
    pub inherent_impls: (),
    pub incoherent_impls: (),
    pub mir_unsafety_check_result: (),
    pub check_unsafety: (),
    pub assumed_wf_types: (),
    pub assumed_wf_types_for_rpitit: (),
    pub fn_sig: (),
    pub lint_mod: (),
    pub check_unused_traits: (),
    pub check_mod_attrs: (),
    pub check_mod_unstable_api_usage: (),
    pub check_mod_const_bodies: (),
    pub check_mod_loops: (),
    pub check_mod_naked_functions: (),
    pub check_mod_privacy: (),
    pub check_liveness: (),
    pub live_symbols_and_ignored_derived_traits: TypedArena<(UnordSet<LocalDefId>, UnordMap<LocalDefId, Vec<(DefId, DefId)>>)>,
    pub check_mod_deathness: (),
    pub check_mod_type_wf: (),
    pub coerce_unsized_info: (),
    pub typeck: (),
    pub diagnostic_only_typeck: (),
    pub used_trait_imports: (),
    pub has_typeck_results: (),
    pub coherent_trait: (),
    pub mir_borrowck: (),
    pub crate_inherent_impls: (),
    pub crate_inherent_impls_overlap_check: (),
    pub orphan_check_impl: (),
    pub mir_callgraph_reachable: (),
    pub mir_inliner_callees: (),
    pub eval_to_allocation_raw: (),
    pub eval_static_initializer: (),
    pub eval_to_const_value_raw: (),
    pub eval_to_valtree: (),
    pub valtree_to_const_val: (),
    pub destructure_const: (),
    pub lit_to_const: (),
    pub check_match: (),
    pub effective_visibilities: (),
    pub check_private_in_public: (),
    pub reachable_set: TypedArena<UnordSet<LocalDefId>>,
    pub region_scope_tree: (),
    pub mir_shims: TypedArena<Body<'tcx>>,
    pub symbol_name: (),
    pub def_kind: (),
    pub def_span: (),
    pub def_ident_span: (),
    pub lookup_stability: (),
    pub lookup_const_stability: (),
    pub lookup_default_body_stability: (),
    pub should_inherit_track_caller: (),
    pub lookup_deprecation_entry: (),
    pub is_doc_hidden: (),
    pub is_doc_notable_trait: (),
    pub item_attrs: (),
    pub codegen_fn_attrs: TypedArena<CodegenFnAttrs>,
    pub asm_target_features: (),
    pub fn_arg_names: (),
    pub rendered_const: TypedArena<String>,
    pub impl_parent: (),
    pub is_ctfe_mir_available: (),
    pub is_mir_available: (),
    pub own_existential_vtable_entries: (),
    pub vtable_entries: (),
    pub vtable_trait_upcasting_coercion_new_vptr_slot: (),
    pub vtable_allocation: (),
    pub codegen_select_candidate: (),
    pub all_local_trait_impls: (),
    pub trait_impls_of: TypedArena<TraitImpls>,
    pub specialization_graph_of: (),
    pub object_safety_violations: (),
    pub check_is_object_safe: (),
    pub param_env: (),
    pub param_env_reveal_all_normalized: (),
    pub is_copy_raw: (),
    pub is_sized_raw: (),
    pub is_freeze_raw: (),
    pub is_unpin_raw: (),
    pub needs_drop_raw: (),
    pub has_significant_drop_raw: (),
    pub has_structural_eq_impl: (),
    pub adt_drop_tys: (),
    pub adt_significant_drop_tys: (),
    pub layout_of: (),
    pub fn_abi_of_fn_ptr: (),
    pub fn_abi_of_instance: (),
    pub dylib_dependency_formats: (),
    pub dependency_formats: TypedArena<Arc<Vec<(CrateType, Vec<Linkage>)>>>,
    pub is_compiler_builtins: (),
    pub has_global_allocator: (),
    pub has_alloc_error_handler: (),
    pub has_panic_handler: (),
    pub is_profiler_runtime: (),
    pub has_ffi_unwind_calls: (),
    pub required_panic_strategy: (),
    pub panic_in_drop_strategy: (),
    pub is_no_builtins: (),
    pub symbol_mangling_version: (),
    pub extern_crate: (),
    pub specializes: (),
    pub in_scope_traits_map: (),
    pub defaultness: (),
    pub check_well_formed: (),
    pub reachable_non_generics: TypedArena<UnordMap<DefId, SymbolExportInfo>>,
    pub is_reachable_non_generic: (),
    pub is_unreachable_local_definition: (),
    pub upstream_monomorphizations: TypedArena<UnordMap<DefId, UnordMap<&'tcx List<GenericArg<'tcx>>, CrateNum>>>,
    pub upstream_monomorphizations_for: (),
    pub upstream_drop_glue_for: (),
    pub foreign_modules: TypedArena<IndexMap<DefId, ForeignModule, BuildHasherDefault<FxHasher>>>,
    pub clashing_extern_declarations: (),
    pub entry_fn: (),
    pub proc_macro_decls_static: (),
    pub crate_hash: (),
    pub crate_host_hash: (),
    pub extra_filename: TypedArena<String>,
    pub crate_extern_paths: TypedArena<Vec<PathBuf>>,
    pub implementations_of_trait: (),
    pub crate_incoherent_impls: (),
    pub native_library: (),
    pub resolve_bound_vars: TypedArena<ResolveBoundVars>,
    pub named_variable_map: (),
    pub is_late_bound_map: (),
    pub object_lifetime_default: (),
    pub late_bound_vars_map: (),
    pub visibility: (),
    pub inhabited_predicate_adt: (),
    pub inhabited_predicate_type: (),
    pub dep_kind: (),
    pub crate_name: (),
    pub module_children: (),
    pub extern_mod_stmt_cnum: (),
    pub lib_features: TypedArena<LibFeatures>,
    pub stability_implications: TypedArena<UnordMap<Symbol, Symbol>>,
    pub intrinsic_raw: (),
    pub get_lang_items: TypedArena<LanguageItems>,
    pub all_diagnostic_items: TypedArena<DiagnosticItems>,
    pub defined_lang_items: (),
    pub diagnostic_items: TypedArena<DiagnosticItems>,
    pub missing_lang_items: (),
    pub visible_parent_map: TypedArena<UnordMap<DefId, DefId>>,
    pub trimmed_def_paths: TypedArena<UnordMap<DefId, Symbol>>,
    pub missing_extern_crate_item: (),
    pub used_crate_source: TypedArena<Arc<CrateSource>>,
    pub debugger_visualizers: TypedArena<Vec<DebuggerVisualizerFile>>,
    pub postorder_cnums: (),
    pub is_private_dep: (),
    pub allocator_kind: (),
    pub alloc_error_handler_kind: (),
    pub upvars_mentioned: (),
    pub maybe_unused_trait_imports: (),
    pub names_imported_by_glob_use: (),
    pub stability_index: TypedArena<Index>,
    pub crates: (),
    pub used_crates: (),
    pub traits: (),
    pub trait_impls_in_crate: (),
    pub exported_symbols: (),
    pub collect_and_partition_mono_items: (),
    pub is_codegened_item: (),
    pub codegen_unit: (),
    pub unused_generic_params: (),
    pub backend_optimization_level: (),
    pub output_filenames: TypedArena<Arc<OutputFilenames>>,
    pub normalize_canonicalized_projection_ty: (),
    pub normalize_canonicalized_weak_ty: (),
    pub normalize_canonicalized_inherent_projection_ty: (),
    pub try_normalize_generic_arg_after_erasing_regions: (),
    pub implied_outlives_bounds_compat: (),
    pub implied_outlives_bounds: (),
    pub dropck_outlives: (),
    pub evaluate_obligation: (),
    pub type_op_ascribe_user_type: (),
    pub type_op_eq: (),
    pub type_op_subtype: (),
    pub type_op_prove_predicate: (),
    pub type_op_normalize_ty: (),
    pub type_op_normalize_clause: (),
    pub type_op_normalize_poly_fn_sig: (),
    pub type_op_normalize_fn_sig: (),
    pub instantiate_and_check_impossible_predicates: (),
    pub is_impossible_associated_item: (),
    pub method_autoderef_steps: (),
    pub supported_target_features: TypedArena<UnordMap<String, Option<Symbol>>>,
    pub features_query: (),
    pub crate_for_resolver: (),
    pub resolve_instance: (),
    pub reveal_opaque_types_in_bounds: (),
    pub limits: (),
    pub diagnostic_hir_wf_check: TypedArena<Option<ObligationCause<'tcx>>>,
    pub global_backend_features: TypedArena<Vec<String>>,
    pub check_validity_requirement: (),
    pub compare_impl_const: (),
    pub deduced_param_attrs: (),
    pub doc_link_resolutions: (),
    pub doc_link_traits_in_scope: (),
    pub check_tys_might_be_eq: (),
    pub stripped_cfg_items: (),
    pub generics_require_sized_self: (),
    pub cross_crate_inlinable: (),
    pub find_field: (),

}

Fields
trigger_delayed_bug: ()

This exists purely for testing the interactions between delayed bugs and incremental.

registered_tools: TypedArena<IndexSet<Ident, BuildHasherDefault<FxHasher>>>

Collects the list of all tools registered using #![register_tool].

early_lint_checks: ()

[query description - consider adding a doc-comment!] perform lints prior to macro expansion

resolutions: ()

[query description - consider adding a doc-comment!] getting the resolver outputs

resolver_for_lowering_raw: ()

[query description - consider adding a doc-comment!] getting the resolver for lowering

source_span: ()

Return the span for a definition. Contrary to def_span below, this query returns the full absolute span of the definition. This span is meant for dep-tracking rather than diagnostics. It should not be used outside of rustc_middle::hir::source_map.

hir_crate: TypedArena<Crate<'tcx>>

Represents crate as a whole (as distinct from the top-level crate module). If you call hir_crate (e.g., indirectly by calling tcx.hir().krate()), we will have to assume that any change means that you need to be recompiled. This is because the hir_crate query gives you access to all other items. To avoid this fate, do not call tcx.hir().krate(); instead, prefer wrappers like tcx.visit_all_items_in_krate().

hir_crate_items: TypedArena<ModuleItems>

All items in the crate.

hir_module_items: TypedArena<ModuleItems>

The items in a module.

This can be conveniently accessed by tcx.hir().visit_item_likes_in_module. Avoid calling this query directly.

local_def_id_to_hir_id: ()

Returns HIR ID for the given LocalDefId.

hir_owner_parent: ()

Gives access to the HIR node’s parent for the HIR owner key.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

opt_hir_owner_nodes: ()

Gives access to the HIR nodes and bodies inside key if it’s a HIR owner.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

hir_attrs: ()

Gives access to the HIR attributes inside the HIR owner key.

This can be conveniently accessed by methods on tcx.hir(). Avoid calling this query directly.

const_param_default: ()

Given the def_id of a const-generic parameter, computes the associated default const parameter. e.g. fn example<const N: usize=3> called on N would return 3.

type_of: ()

Returns the Ty of the given DefId. If the DefId points to an alias, it will “skip” this alias to return the aliased type.

type_of_opaque: ()

Specialized instance of type_of that detects cycles that are due to revealing opaque because of an auto trait bound. Unless CyclePlaceholder needs to be handled separately, call type_of instead.

type_alias_is_lazy: ()

[query description - consider adding a doc-comment!] computing whether {path} is a lazy type alias

collect_return_position_impl_trait_in_trait_tys: ()

[query description - consider adding a doc-comment!] comparing an impl and trait method signature, inferring any hidden impl Trait types in the process

is_type_alias_impl_trait: ()

[query description - consider adding a doc-comment!] determine whether the opaque is a type-alias impl trait

unsizing_params_for_adt: TypedArena<BitSet<u32>>

[query description - consider adding a doc-comment!] determining what parameters of tcx.def_path_str(key) can participate in unsizing

analysis: ()

The root query triggering all analysis passes like typeck or borrowck.

check_expectations: ()

This query checks the fulfillment of collected lint expectations. All lint emitting queries have to be done before this is executed to ensure that all expectations can be fulfilled.

This is an extra query to enable other drivers (like rustdoc) to only execute a small subset of the analysis query, while allowing lints to be expected. In rustc, this query will be executed as part of the analysis query and doesn’t have to be called a second time.

Tools can additionally pass in a tool filter. That will restrict the expectations to only trigger for lints starting with the listed tool name. This is useful for cases were not all linting code from rustc was called. With the default None all registered lints will also be checked for expectation fulfillment.

generics_of: TypedArena<Generics>

Maps from the DefId of an item (trait/struct/enum/fn) to its associated generics.

predicates_of: ()

Maps from the DefId of an item (trait/struct/enum/fn) to the predicates (where-clauses) that must be proven true in order to reference it. This is almost always the “predicates query” that you want.

predicates_of builds on predicates_defined_on – in fact, it is almost always the same as that query, except for the case of traits. For traits, predicates_of contains an additional Self: Trait<...> predicate that users don’t actually write. This reflects the fact that to invoke the trait (e.g., via Default::default) you must supply types that actually implement the trait. (However, this extra predicate gets in the way of some checks, which are intended to operate over only the actual where-clauses written by the user.)

opaque_types_defined_by: ()

[query description - consider adding a doc-comment!] computing the opaque types defined by tcx.def_path_str(key.to_def_id())

explicit_item_bounds: ()

Returns the list of bounds that can be used for SelectionCandidate::ProjectionCandidate(_) and ProjectionTyCandidate::TraitDef. Specifically this is the bounds written on the trait’s type definition, or those after the impl keyword

type X: Bound + 'lt
//      ^^^^^^^^^^^
impl Debug + Display
//   ^^^^^^^^^^^^^^^

key is the DefId of the associated type or opaque type.

Bounds from the parent (e.g. with nested impl trait) are not included.

item_bounds: ()

Elaborated version of the predicates from explicit_item_bounds.

For example:

trait MyTrait {
    type MyAType: Eq + ?Sized;
}

explicit_item_bounds returns [<Self as MyTrait>::MyAType: Eq], and item_bounds returns

[
    <Self as Trait>::MyAType: Eq,
    <Self as Trait>::MyAType: PartialEq<<Self as Trait>::MyAType>
]

Bounds from the parent (e.g. with nested impl trait) are not included.

native_libraries: TypedArena<Vec<NativeLib>>

Look up all native libraries this crate depends on. These are assembled from the following places:

extern blocks (depending on their link attributes)

the libs (-l) option

shallow_lint_levels_on: TypedArena<ShallowLintLevelMap>

[query description - consider adding a doc-comment!] looking up lint levels for tcx.def_path_str(key)

lint_expectations: TypedArena<Vec<(LintExpectationId, LintExpectation)>>

[query description - consider adding a doc-comment!] computing #[expect]ed lints in this crate

expn_that_defined: ()

[query description - consider adding a doc-comment!] getting the expansion that defined tcx.def_path_str(key)

is_panic_runtime: ()

[query description - consider adding a doc-comment!] checking if the crate is_panic_runtime

representability: ()

Checks whether a type is representable or infinitely sized

representability_adt_ty: ()

An implementation detail for the representability query

params_in_repr: TypedArena<BitSet<u32>>

Set of param indexes for type params that are in the type’s representation

thir_body: ()

Fetch the THIR for a given body. If typeck for that body failed, returns an empty Thir.

thir_tree: TypedArena<String>

Create a THIR tree for debugging.

thir_flat: TypedArena<String>

Create a list-like THIR representation for debugging.

mir_keys: TypedArena<IndexSet<LocalDefId, BuildHasherDefault<FxHasher>>>

Set of all the DefIds in this crate that have MIR associated with them. This includes all the body owners, but also things like struct constructors.

mir_const_qualif: ()

Maps DefId’s that have an associated mir::Body to the result of the MIR const-checking pass. This is the set of qualifs in the final value of a const.

mir_built: ()

Fetch the MIR for a given DefId right after it’s built - this includes unreachable code.

mir_const: ()

Fetch the MIR for a given DefId up till the point where it is ready for const qualification.

See the README for the mir module for details.

thir_abstract_const: ()

Try to build an abstract representation of the given constant.

mir_drops_elaborated_and_const_checked: ()

[query description - consider adding a doc-comment!] elaborating drops for tcx.def_path_str(key)

mir_for_ctfe: ()

[query description - consider adding a doc-comment!] caching mir of tcx.def_path_str(key) for CTFE

mir_promoted: ()

[query description - consider adding a doc-comment!] promoting constants in MIR for tcx.def_path_str(key)

closure_typeinfo: ()

[query description - consider adding a doc-comment!] finding symbols for captures of closure tcx.def_path_str(key)

closure_saved_names_of_captured_variables: TypedArena<IndexVec<FieldIdx, Symbol>>

Returns names of captured upvars for closures and coroutines.

Here are some examples:

name__field1__field2 when the upvar is captured by value.

_ref__name__field when the upvar is captured by reference.

For coroutines this only contains upvars that are shared by all states.

mir_coroutine_witnesses: TypedArena<Option<CoroutineLayout<'tcx>>>

[query description - consider adding a doc-comment!] coroutine witness types for tcx.def_path_str(key)

check_coroutine_obligations: ()

[query description - consider adding a doc-comment!] verify auto trait bounds for coroutine interior type tcx.def_path_str(key)

optimized_mir: ()

MIR after our optimization passes have run. This is MIR that is ready for codegen. This is also the only query that can fetch non-local MIR, at present.

coverage_ids_info: TypedArena<CoverageIdsInfo>

Summarizes coverage IDs inserted by the InstrumentCoverage MIR pass (for compiler option -Cinstrument-coverage), after MIR optimizations have had a chance to potentially remove some of them.

promoted_mir: ()

The DefId is the DefId of the containing MIR body. Promoteds do not have their own DefId. This function returns all promoteds in the specified body. The body references promoteds by the DefId and the mir::Promoted index. This is necessary, because after inlining a body may refer to promoteds from other bodies. In that case you still need to use the DefId of the original body.

erase_regions_ty: ()

Erases regions from ty to yield a new type. Normally you would just use tcx.erase_regions(value), however, which uses this query as a kind of cache.

wasm_import_module_map: TypedArena<UnordMap<DefId, String>>

[query description - consider adding a doc-comment!] getting wasm import module map

predicates_defined_on: ()

Maps from the DefId of an item (trait/struct/enum/fn) to the predicates (where-clauses) directly defined on it. This is equal to the explicit_predicates_of predicates plus the inferred_outlives_of predicates.

trait_explicit_predicates_and_bounds: ()

Returns everything that looks like a predicate written explicitly by the user on a trait item.

Traits are unusual, because predicates on associated types are converted into bounds on that type for backwards compatibility:

trait X where Self::U: Copy { type U; }

becomes

trait X { type U: Copy; }

explicit_predicates_of and explicit_item_bounds will then take the appropriate subsets of the predicates here.

explicit_predicates_of: ()

Returns the predicates written explicitly by the user.

inferred_outlives_of: ()

Returns the inferred outlives predicates (e.g., for struct Foo<'a, T> { x: &'a T }, this would return T: 'a).

super_predicates_of: ()

Maps from the DefId of a trait to the list of super-predicates. This is a subset of the full list of predicates. We store these in a separate map because we must evaluate them even during type conversion, often before the full predicates are available (note that supertraits have additional acyclicity requirements).

implied_predicates_of: ()

[query description - consider adding a doc-comment!] computing the implied predicates of tcx.def_path_str(key)

super_predicates_that_define_assoc_item: ()

The Option<Ident> is the name of an associated type. If it is None, then this query returns the full set of predicates. If Some<Ident>, then the query returns only the subset of super-predicates that reference traits that define the given associated type. This is used to avoid cycles in resolving types like T::Item.

type_param_predicates: ()

To avoid cycles within the predicates of a single item we compute per-type-parameter predicates for resolving T::AssocTy.

trait_def: TypedArena<TraitDef>

[query description - consider adding a doc-comment!] computing trait definition for tcx.def_path_str(key)

adt_def: ()

[query description - consider adding a doc-comment!] computing ADT definition for tcx.def_path_str(key)

adt_destructor: ()

[query description - consider adding a doc-comment!] computing Drop impl for tcx.def_path_str(key)

adt_sized_constraint: ()

[query description - consider adding a doc-comment!] computing Sized constraints for tcx.def_path_str(key)

adt_dtorck_constraint: ()

[query description - consider adding a doc-comment!] computing drop-check constraints for tcx.def_path_str(key)

constness: ()

Returns true if this is a const fn, use the is_const_fn to know whether your crate actually sees it as const fn (e.g., the const-fn-ness might be unstable and you might not have the feature gate active).

Do not call this function manually. It is only meant to cache the base data for the is_const_fn function. Consider using is_const_fn or is_const_fn_raw instead.

asyncness: ()

[query description - consider adding a doc-comment!] checking if the function is async: tcx.def_path_str(key)

is_promotable_const_fn: ()

Returns true if calls to the function may be promoted.

This is either because the function is e.g., a tuple-struct or tuple-variant constructor, or because it has the #[rustc_promotable] attribute. The attribute should be removed in the future in favour of some form of check which figures out whether the function does not inspect the bits of any of its arguments (so is essentially just a constructor function).

coroutine_kind: ()

Returns Some(coroutine_kind) if the node pointed to by def_id is a coroutine.

coroutine_for_closure: ()

[query description - consider adding a doc-comment!] Given a coroutine-closure def id, return the def id of the coroutine returned by it

crate_variances: TypedArena<CrateVariancesMap<'tcx>>

Gets a map with the variance of every item; use variances_of instead.

variances_of: ()

Maps from the DefId of a type or region parameter to its (inferred) variance.

inferred_outlives_crate: TypedArena<CratePredicatesMap<'tcx>>

Maps from thee DefId of a type to its (inferred) outlives.

associated_item_def_ids: ()

Maps from an impl/trait or struct/variant DefId to a list of the DefIds of its associated items or fields.

associated_item: ()

Maps from a trait item to the trait item “descriptor”.

associated_items: TypedArena<AssocItems>

Collects the associated items defined on a trait or impl.

impl_item_implementor_ids: TypedArena<UnordMap<DefId, DefId>>

Maps from associated items on a trait to the corresponding associated item on the impl specified by impl_id.

For example, with the following code

struct Type {}
                        // DefId
trait Trait {           // trait_id
    fn f();             // trait_f
    fn g() {}           // trait_g
}

impl Trait for Type {   // impl_id
    fn f() {}           // impl_f
    fn g() {}           // impl_g
}

The map returned for tcx.impl_item_implementor_ids(impl_id) would be { trait_f: impl_f, trait_g: impl_g }

associated_types_for_impl_traits_in_associated_fn: ()

Given fn_def_id of a trait or of an impl that implements a given trait: if fn_def_id is the def id of a function defined inside a trait, then it creates and returns the associated items that correspond to each impl trait in return position for that trait. if fn_def_id is the def id of a function defined inside an impl that implements a trait, then it creates and returns the associated items that correspond to each impl trait in return position of the implemented trait.

associated_type_for_impl_trait_in_trait: ()

Given an impl trait in trait opaque_ty_def_id, create and return the corresponding associated item.

impl_trait_header: ()

Given an impl_id, return the trait it implements along with some header information. Return None if this is an inherent impl.

issue33140_self_ty: ()

[query description - consider adding a doc-comment!] computing Self type wrt issue #33140 tcx.def_path_str(key)

inherent_impls: ()

Maps a DefId of a type to a list of its inherent impls. Contains implementations of methods that are inherent to a type. Methods in these implementations don’t need to be exported.

incoherent_impls: ()

[query description - consider adding a doc-comment!] collecting all inherent impls for {:?}

mir_unsafety_check_result: ()

The result of unsafety-checking this LocalDefId with the old checker.

check_unsafety: ()

Unsafety-check this LocalDefId.

assumed_wf_types: ()

Returns the types assumed to be well formed while “inside” of the given item.

Note that we’ve liberated the late bound regions of function signatures, so this can not be used to check whether these types are well formed.

assumed_wf_types_for_rpitit: ()

We need to store the assumed_wf_types for an RPITIT so that impls of foreign traits with return-position impl trait in traits can inherit the right wf types.

fn_sig: ()

Computes the signature of the function.

lint_mod: ()

Performs lint checking for the module.

check_unused_traits: ()

[query description - consider adding a doc-comment!] checking unused trait imports in crate

check_mod_attrs: ()

Checks the attributes in the module.

check_mod_unstable_api_usage: ()

Checks for uses of unstable APIs in the module.

check_mod_const_bodies: ()

Checks the const bodies in the module for illegal operations (e.g. if or loop).

check_mod_loops: ()

Checks the loops in the module.

check_mod_naked_functions: ()

[query description - consider adding a doc-comment!] checking naked functions in describe_as_module(key, tcx)

check_mod_privacy: ()

[query description - consider adding a doc-comment!] checking privacy in describe_as_module(key.to_local_def_id(), tcx)

check_liveness: ()

[query description - consider adding a doc-comment!] checking liveness of variables in tcx.def_path_str(key)

live_symbols_and_ignored_derived_traits: TypedArena<(UnordSet<LocalDefId>, UnordMap<LocalDefId, Vec<(DefId, DefId)>>)>

Return the live symbols in the crate for dead code check.

The second return value maps from ADTs to ignored derived traits (e.g. Debug and Clone) and their respective impl (i.e., part of the derive macro)

check_mod_deathness: ()

[query description - consider adding a doc-comment!] checking deathness of variables in describe_as_module(key, tcx)

check_mod_type_wf: ()

[query description - consider adding a doc-comment!] checking that types are well-formed in describe_as_module(key, tcx)

coerce_unsized_info: ()

Caches CoerceUnsized kinds for impls on custom types.

typeck: ()

[query description - consider adding a doc-comment!] type-checking tcx.def_path_str(key)

diagnostic_only_typeck: ()

[query description - consider adding a doc-comment!] type-checking tcx.def_path_str(key)

used_trait_imports: ()

[query description - consider adding a doc-comment!] finding used_trait_imports tcx.def_path_str(key)

has_typeck_results: ()

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) has a body

coherent_trait: ()

[query description - consider adding a doc-comment!] coherence checking all impls of trait tcx.def_path_str(def_id)

mir_borrowck: ()

Borrow-checks the function body. If this is a closure, returns additional requirements that the closure’s creator must verify.

crate_inherent_impls: ()

Gets a complete map from all types to their inherent impls. Not meant to be used directly outside of coherence.

crate_inherent_impls_overlap_check: ()

Checks all types in the crate for overlap in their inherent impls. Reports errors. Not meant to be used directly outside of coherence.

orphan_check_impl: ()

Checks whether all impls in the crate pass the overlap check, returning which impls fail it. If all impls are correct, the returned slice is empty.

mir_callgraph_reachable: ()

Check whether the function has any recursion that could cause the inliner to trigger a cycle. Returns the call stack causing the cycle. The call stack does not contain the current function, just all intermediate functions.

mir_inliner_callees: ()

Obtain all the calls into other local functions

eval_to_allocation_raw: ()

Evaluates a constant and returns the computed allocation.

Do not use this directly, use the eval_to_const_value or eval_to_valtree instead.

eval_static_initializer: ()

Evaluate a static’s initializer, returning the allocation of the initializer’s memory.

eval_to_const_value_raw: ()

Evaluates const items or anonymous constants (such as enum variant explicit discriminants or array lengths) into a representation suitable for the type system and const generics.

Do not use this directly, use one of the following wrappers: tcx.const_eval_poly, tcx.const_eval_resolve, tcx.const_eval_instance, or tcx.const_eval_global_id.

eval_to_valtree: ()

Evaluate a constant and convert it to a type level constant or return None if that is not possible.

valtree_to_const_val: ()

Converts a type level constant value into ConstValue

destructure_const: ()

Destructures array, ADT or tuple constants into the constants of their fields.

lit_to_const: ()

[query description - consider adding a doc-comment!] converting literal to const

check_match: ()

[query description - consider adding a doc-comment!] match-checking tcx.def_path_str(key)

effective_visibilities: ()

Performs part of the privacy check and computes effective visibilities.

check_private_in_public: ()

[query description - consider adding a doc-comment!] checking for private elements in public interfaces

reachable_set: TypedArena<UnordSet<LocalDefId>>

[query description - consider adding a doc-comment!] reachability

region_scope_tree: ()

Per-body region::ScopeTree. The DefId should be the owner DefId for the body; in the case of closures, this will be redirected to the enclosing function.

mir_shims: TypedArena<Body<'tcx>>

Generates a MIR body for the shim.

symbol_name: ()

The symbol_name query provides the symbol name for calling a given instance from the local crate. In particular, it will also look up the correct symbol name of instances from upstream crates.

def_kind: ()

[query description - consider adding a doc-comment!] looking up definition kind of tcx.def_path_str(def_id)

def_span: ()

Gets the span for the definition.

def_ident_span: ()

Gets the span for the identifier of the definition.

lookup_stability: ()

[query description - consider adding a doc-comment!] looking up stability of tcx.def_path_str(def_id)

lookup_const_stability: ()

[query description - consider adding a doc-comment!] looking up const stability of tcx.def_path_str(def_id)

lookup_default_body_stability: ()

[query description - consider adding a doc-comment!] looking up default body stability of tcx.def_path_str(def_id)

should_inherit_track_caller: ()

[query description - consider adding a doc-comment!] computing should_inherit_track_caller of tcx.def_path_str(def_id)

lookup_deprecation_entry: ()

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is deprecated

is_doc_hidden: ()

Determines whether an item is annotated with doc(hidden).

is_doc_notable_trait: ()

Determines whether an item is annotated with doc(notable_trait).

item_attrs: ()

Returns the attributes on the item at def_id.

Do not use this directly, use tcx.get_attrs instead.

codegen_fn_attrs: TypedArena<CodegenFnAttrs>

[query description - consider adding a doc-comment!] computing codegen attributes of tcx.def_path_str(def_id)

asm_target_features: ()

[query description - consider adding a doc-comment!] computing target features for inline asm of tcx.def_path_str(def_id)

fn_arg_names: ()

[query description - consider adding a doc-comment!] looking up function parameter names for tcx.def_path_str(def_id)

rendered_const: TypedArena<String>

Gets the rendered value of the specified constant or associated constant. Used by rustdoc.

impl_parent: ()

[query description - consider adding a doc-comment!] computing specialization parent impl of tcx.def_path_str(def_id)

is_ctfe_mir_available: ()

[query description - consider adding a doc-comment!] checking if item has CTFE MIR available: tcx.def_path_str(key)

is_mir_available: ()

[query description - consider adding a doc-comment!] checking if item has MIR available: tcx.def_path_str(key)

own_existential_vtable_entries: ()

[query description - consider adding a doc-comment!] finding all existential vtable entries for trait tcx.def_path_str(key)

vtable_entries: ()

[query description - consider adding a doc-comment!] finding all vtable entries for trait tcx.def_path_str(key.def_id())

vtable_trait_upcasting_coercion_new_vptr_slot: ()

[query description - consider adding a doc-comment!] finding the slot within vtable for trait object key.1 vtable ptr during trait upcasting coercion from key.0 vtable

vtable_allocation: ()

[query description - consider adding a doc-comment!] vtable const allocation for < key.0 as key.1.map(| trait_ref | format! ("{trait_ref}")).unwrap_or("_".to_owned()) >

codegen_select_candidate: ()

[query description - consider adding a doc-comment!] computing candidate for key.1

all_local_trait_impls: ()

Return all impl blocks in the current crate.

trait_impls_of: TypedArena<TraitImpls>

Given a trait trait_id, return all known impl blocks.

specialization_graph_of: ()

[query description - consider adding a doc-comment!] building specialization graph of trait tcx.def_path_str(trait_id)

object_safety_violations: ()

[query description - consider adding a doc-comment!] determining object safety of trait tcx.def_path_str(trait_id)

check_is_object_safe: ()

[query description - consider adding a doc-comment!] checking if trait tcx.def_path_str(trait_id) is object safe

param_env: ()

Gets the ParameterEnvironment for a given item; this environment will be in “user-facing” mode, meaning that it is suitable for type-checking etc, and it does not normalize specializable associated types. This is almost always what you want, unless you are doing MIR optimizations, in which case you might want to use reveal_all() method to change modes.

param_env_reveal_all_normalized: ()

Like param_env, but returns the ParamEnv in Reveal::All mode. Prefer this over tcx.param_env(def_id).with_reveal_all_normalized(tcx), as this method is more efficient.

is_copy_raw: ()

Trait selection queries. These are best used by invoking ty.is_copy_modulo_regions(), ty.is_copy(), etc, since that will prune the environment where possible.

is_sized_raw: ()

Query backing Ty::is_sized.

is_freeze_raw: ()

Query backing Ty::is_freeze.

is_unpin_raw: ()

Query backing Ty::is_unpin.

needs_drop_raw: ()

Query backing Ty::needs_drop.

has_significant_drop_raw: ()

Query backing Ty::has_significant_drop_raw.

has_structural_eq_impl: ()

Query backing Ty::is_structural_eq_shallow.

This is only correct for ADTs. Call is_structural_eq_shallow to handle all types correctly.

adt_drop_tys: ()

A list of types where the ADT requires drop if and only if any of those types require drop. If the ADT is known to always need drop then Err(AlwaysRequiresDrop) is returned.

adt_significant_drop_tys: ()

A list of types where the ADT requires drop if and only if any of those types has significant drop. A type marked with the attribute rustc_insignificant_dtor is considered to not be significant. A drop is significant if it is implemented by the user or does anything that will have any observable behavior (other than freeing up memory). If the ADT is known to have a significant destructor then Err(AlwaysRequiresDrop) is returned.

layout_of: ()

Computes the layout of a type. Note that this implicitly executes in “reveal all” mode, and will normalize the input type.

fn_abi_of_fn_ptr: ()

Compute a FnAbi suitable for indirect calls, i.e. to fn pointers.

NB: this doesn’t handle virtual calls - those should use fn_abi_of_instance instead, where the instance is an InstanceDef::Virtual.

fn_abi_of_instance: ()

Compute a FnAbi suitable for declaring/defining an fn instance, and for direct calls to an fn.

NB: that includes virtual calls, which are represented by “direct calls” to an InstanceDef::Virtual instance (of <dyn Trait as Trait>::fn).

dylib_dependency_formats: ()

[query description - consider adding a doc-comment!] getting dylib dependency formats of crate

dependency_formats: TypedArena<Arc<Vec<(CrateType, Vec<Linkage>)>>>

[query description - consider adding a doc-comment!] getting the linkage format of all dependencies

is_compiler_builtins: ()

[query description - consider adding a doc-comment!] checking if the crate is_compiler_builtins

has_global_allocator: ()

[query description - consider adding a doc-comment!] checking if the crate has_global_allocator

has_alloc_error_handler: ()

[query description - consider adding a doc-comment!] checking if the crate has_alloc_error_handler

has_panic_handler: ()

[query description - consider adding a doc-comment!] checking if the crate has_panic_handler

is_profiler_runtime: ()

[query description - consider adding a doc-comment!] checking if a crate is #![profiler_runtime]

has_ffi_unwind_calls: ()

[query description - consider adding a doc-comment!] checking if tcx.def_path_str(key) contains FFI-unwind calls

required_panic_strategy: ()

[query description - consider adding a doc-comment!] getting a crate’s required panic strategy

panic_in_drop_strategy: ()

[query description - consider adding a doc-comment!] getting a crate’s configured panic-in-drop strategy

is_no_builtins: ()

[query description - consider adding a doc-comment!] getting whether a crate has #![no_builtins]

symbol_mangling_version: ()

[query description - consider adding a doc-comment!] getting a crate’s symbol mangling version

extern_crate: ()

[query description - consider adding a doc-comment!] getting crate’s ExternCrateData

specializes: ()

[query description - consider adding a doc-comment!] computing whether impls specialize one another

in_scope_traits_map: ()

[query description - consider adding a doc-comment!] getting traits in scope at a block

defaultness: ()

Returns whether the impl or associated function has the default keyword.

check_well_formed: ()

[query description - consider adding a doc-comment!] checking that tcx.def_path_str(key) is well-formed

reachable_non_generics: TypedArena<UnordMap<DefId, SymbolExportInfo>>

[query description - consider adding a doc-comment!] looking up the exported symbols of a crate

is_reachable_non_generic: ()

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is an exported symbol

is_unreachable_local_definition: ()

[query description - consider adding a doc-comment!] checking whether tcx.def_path_str(def_id) is reachable from outside the crate

upstream_monomorphizations: TypedArena<UnordMap<DefId, UnordMap<&'tcx List<GenericArg<'tcx>>, CrateNum>>>

The entire set of monomorphizations the local crate can safely link to because they are exported from upstream crates. Do not depend on this directly, as its value changes anytime a monomorphization gets added or removed in any upstream crate. Instead use the narrower upstream_monomorphizations_for, upstream_drop_glue_for, or, even better, Instance::upstream_monomorphization().

upstream_monomorphizations_for: ()

Returns the set of upstream monomorphizations available for the generic function identified by the given def_id. The query makes sure to make a stable selection if the same monomorphization is available in multiple upstream crates.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

upstream_drop_glue_for: ()

Returns the upstream crate that exports drop-glue for the given type (args is expected to be a single-item list containing the type one wants drop-glue for).

This is a subset of upstream_monomorphizations_for in order to increase dep-tracking granularity. Otherwise adding or removing any type with drop-glue in any upstream crate would invalidate all functions calling drop-glue of an upstream type.

You likely want to call Instance::upstream_monomorphization() instead of invoking this query directly.

NOTE: This query could easily be extended to also support other common functions that have are large set of monomorphizations (like Clone::clone for example).

foreign_modules: TypedArena<IndexMap<DefId, ForeignModule, BuildHasherDefault<FxHasher>>>

Returns a list of all extern blocks of a crate.

clashing_extern_declarations: ()

Lint against extern fn declarations having incompatible types.

entry_fn: ()

Identifies the entry-point (e.g., the main function) for a given crate, returning None if there is no entry point (such as for library crates).

proc_macro_decls_static: ()

Finds the rustc_proc_macro_decls item of a crate.

crate_hash: ()

[query description - consider adding a doc-comment!] looking up the hash a crate

crate_host_hash: ()

Gets the hash for the host proc macro. Used to support -Z dual-proc-macro.

extra_filename: TypedArena<String>

Gets the extra data to put in each output filename for a crate. For example, compiling the foo crate with extra-filename=-a creates a libfoo-b.rlib file.

crate_extern_paths: TypedArena<Vec<PathBuf>>

Gets the paths where the crate came from in the file system.

implementations_of_trait: ()

Given a crate and a trait, look up all impls of that trait in the crate. Return (impl_id, self_ty).

crate_incoherent_impls: ()

Collects all incoherent impls for the given crate and type.

Do not call this directly, but instead use the incoherent_impls query. This query is only used to get the data necessary for that query.

native_library: ()

Get the corresponding native library from the native_libraries query

resolve_bound_vars: TypedArena<ResolveBoundVars>

Does lifetime resolution on items. Importantly, we can’t resolve lifetimes directly on things like trait methods, because of trait params. See rustc_resolve::late::lifetimes for details.

named_variable_map: ()

[query description - consider adding a doc-comment!] looking up a named region

is_late_bound_map: ()

[query description - consider adding a doc-comment!] testing if a region is late bound

object_lifetime_default: ()

For a given item’s generic parameter, gets the default lifetimes to be used for each parameter if a trait object were to be passed for that parameter. For example, for T in struct Foo<'a, T>, this would be 'static. For T in struct Foo<'a, T: 'a>, this would instead be 'a. This query will panic if passed something that is not a type parameter.

late_bound_vars_map: ()

[query description - consider adding a doc-comment!] looking up late bound vars

visibility: ()

Computes the visibility of the provided def_id.

If the item from the def_id doesn’t have a visibility, it will panic. For example a generic type parameter will panic if you call this method on it:

use std::fmt::Debug;

pub trait Foo<T: Debug> {}

In here, if you call visibility on T, it’ll panic.

inhabited_predicate_adt: ()

[query description - consider adding a doc-comment!] computing the uninhabited predicate of {:?}

inhabited_predicate_type: ()

Do not call this query directly: invoke Ty::inhabited_predicate instead.

dep_kind: ()

[query description - consider adding a doc-comment!] fetching what a dependency looks like

crate_name: ()

Gets the name of the crate.

module_children: ()

[query description - consider adding a doc-comment!] collecting child items of module tcx.def_path_str(def_id)

extern_mod_stmt_cnum: ()

[query description - consider adding a doc-comment!] computing crate imported by tcx.def_path_str(def_id)

lib_features: TypedArena<LibFeatures>

[query description - consider adding a doc-comment!] calculating the lib features defined in a crate

stability_implications: TypedArena<UnordMap<Symbol, Symbol>>

[query description - consider adding a doc-comment!] calculating the implications between #[unstable] features defined in a crate

intrinsic_raw: ()

Whether the function is an intrinsic

get_lang_items: TypedArena<LanguageItems>

Returns the lang items defined in another crate by loading it from metadata.

all_diagnostic_items: TypedArena<DiagnosticItems>

Returns all diagnostic items defined in all crates.

defined_lang_items: ()

Returns the lang items defined in another crate by loading it from metadata.

diagnostic_items: TypedArena<DiagnosticItems>

Returns the diagnostic items defined in a crate.

missing_lang_items: ()

[query description - consider adding a doc-comment!] calculating the missing lang items in a crate

visible_parent_map: TypedArena<UnordMap<DefId, DefId>>

The visible parent map is a map from every item to a visible parent. It prefers the shortest visible path to an item. Used for diagnostics, for example path trimming. The parents are modules, enums or traits.

trimmed_def_paths: TypedArena<UnordMap<DefId, Symbol>>

Collects the “trimmed”, shortest accessible paths to all items for diagnostics. See the provider docs for more info.

missing_extern_crate_item: ()

[query description - consider adding a doc-comment!] seeing if we’re missing an extern crate item for this crate

used_crate_source: TypedArena<Arc<CrateSource>>

[query description - consider adding a doc-comment!] looking at the source for a crate

debugger_visualizers: TypedArena<Vec<DebuggerVisualizerFile>>

Returns the debugger visualizers defined for this crate. NOTE: This query has to be marked eval_always because it reads data directly from disk that is not tracked anywhere else. I.e. it represents a genuine input to the query system.

postorder_cnums: ()

[query description - consider adding a doc-comment!] generating a postorder list of CrateNums

is_private_dep: ()

Returns whether or not the crate with CrateNum ‘cnum’ is marked as a private dependency

allocator_kind: ()

[query description - consider adding a doc-comment!] getting the allocator kind for the current crate

alloc_error_handler_kind: ()

[query description - consider adding a doc-comment!] alloc error handler kind for the current crate

upvars_mentioned: ()

[query description - consider adding a doc-comment!] collecting upvars mentioned in tcx.def_path_str(def_id)

maybe_unused_trait_imports: ()

[query description - consider adding a doc-comment!] fetching potentially unused trait imports

names_imported_by_glob_use: ()

[query description - consider adding a doc-comment!] finding names imported by glob use for tcx.def_path_str(def_id)

stability_index: TypedArena<Index>

[query description - consider adding a doc-comment!] calculating the stability index for the local crate

crates: ()

[query description - consider adding a doc-comment!] fetching all foreign CrateNum instances

used_crates: ()

[query description - consider adding a doc-comment!] fetching CrateNums for all crates loaded non-speculatively

traits: ()

A list of all traits in a crate, used by rustdoc and error reporting.

trait_impls_in_crate: ()

[query description - consider adding a doc-comment!] fetching all trait impls in a crate

exported_symbols: ()

The list of symbols exported from the given crate.

All names contained in exported_symbols(cnum) are guaranteed to correspond to a publicly visible symbol in cnum machine code.

The exported_symbols sets of different crates do not intersect.

collect_and_partition_mono_items: ()

[query description - consider adding a doc-comment!] collect_and_partition_mono_items

is_codegened_item: ()

[query description - consider adding a doc-comment!] determining whether tcx.def_path_str(def_id) needs codegen

codegen_unit: ()

[query description - consider adding a doc-comment!] getting codegen unit {sym}

unused_generic_params: ()

[query description - consider adding a doc-comment!] determining which generic parameters are unused by tcx.def_path_str(key.def_id())

backend_optimization_level: ()

[query description - consider adding a doc-comment!] optimization level used by backend

output_filenames: TypedArena<Arc<OutputFilenames>>

Return the filenames where output artefacts shall be stored.

This query returns an &Arc because codegen backends need the value even after the TyCtxt has been destroyed.

normalize_canonicalized_projection_ty: ()

Do not call this query directly: Invoke normalize instead.

normalize_canonicalized_weak_ty: ()

Do not call this query directly: Invoke normalize instead.

normalize_canonicalized_inherent_projection_ty: ()

Do not call this query directly: Invoke normalize instead.

try_normalize_generic_arg_after_erasing_regions: ()

Do not call this query directly: invoke try_normalize_erasing_regions instead.

implied_outlives_bounds_compat: ()

[query description - consider adding a doc-comment!] computing implied outlives bounds for goal.value.value

implied_outlives_bounds: ()

[query description - consider adding a doc-comment!] computing implied outlives bounds v2 for goal.value.value

dropck_outlives: ()

Do not call this query directly: invoke DropckOutlives::new(dropped_ty)).fully_perform(typeck.infcx) instead.

evaluate_obligation: ()

Do not call this query directly: invoke infcx.predicate_may_hold() or infcx.predicate_must_hold() instead.

type_op_ascribe_user_type: ()

Do not call this query directly: part of the Eq type-op

type_op_eq: ()

Do not call this query directly: part of the Eq type-op

type_op_subtype: ()

Do not call this query directly: part of the Subtype type-op

type_op_prove_predicate: ()

Do not call this query directly: part of the ProvePredicate type-op

type_op_normalize_ty: ()

Do not call this query directly: part of the Normalize type-op

type_op_normalize_clause: ()

Do not call this query directly: part of the Normalize type-op

type_op_normalize_poly_fn_sig: ()

Do not call this query directly: part of the Normalize type-op

type_op_normalize_fn_sig: ()

Do not call this query directly: part of the Normalize type-op

instantiate_and_check_impossible_predicates: ()

[query description - consider adding a doc-comment!] checking impossible instantiated predicates: tcx.def_path_str(key.0)

is_impossible_associated_item: ()

[query description - consider adding a doc-comment!] checking if tcx.def_path_str(key.1) is impossible to reference within tcx.def_path_str(key.0)

method_autoderef_steps: ()

[query description - consider adding a doc-comment!] computing autoderef types for goal.value.value

supported_target_features: TypedArena<UnordMap<String, Option<Symbol>>>

[query description - consider adding a doc-comment!] looking up supported target features

features_query: ()

[query description - consider adding a doc-comment!] looking up enabled feature gates

crate_for_resolver: ()

[query description - consider adding a doc-comment!] the ast before macro expansion and name resolution

resolve_instance: ()

Attempt to resolve the given DefId to an Instance, for the given generics args (GenericArgsRef), returning one of:

Ok(Some(instance)) on success

Ok(None) when the GenericArgsRef are still too generic, and therefore don’t allow finding the final Instance

Err(ErrorGuaranteed) when the Instance resolution process couldn’t complete due to errors elsewhere - this is distinct from Ok(None) to avoid misleading diagnostics when an error has already been/will be emitted, for the original cause

reveal_opaque_types_in_bounds: ()

[query description - consider adding a doc-comment!] revealing opaque types in {:?}

limits: ()

[query description - consider adding a doc-comment!] looking up limits

diagnostic_hir_wf_check: TypedArena<Option<ObligationCause<'tcx>>>

Performs an HIR-based well-formed check on the item with the given HirId. If we get an Unimplemented error that matches the provided Predicate, return the cause of the newly created obligation.

This is only used by error-reporting code to get a better cause (in particular, a better span) for an existing error. Therefore, it is best-effort, and may never handle all of the cases that the normal ty::Ty-based wfcheck does. This is fine, because the ty::Ty-based wfcheck is always run.

global_backend_features: TypedArena<Vec<String>>

The list of backend features computed from CLI flags (-Ctarget-cpu, -Ctarget-feature, --target and similar).

check_validity_requirement: ()

[query description - consider adding a doc-comment!] checking validity requirement for key.1.value : key.0

compare_impl_const: ()

[query description - consider adding a doc-comment!] checking assoc const tcx.def_path_str(key.0) has the same type as trait item

deduced_param_attrs: ()

[query description - consider adding a doc-comment!] deducing parameter attributes for tcx.def_path_str(def_id)

doc_link_resolutions: ()

[query description - consider adding a doc-comment!] resolutions for documentation links for a module

doc_link_traits_in_scope: ()

[query description - consider adding a doc-comment!] traits in scope for documentation links for a module

check_tys_might_be_eq: ()

Used in super_combine_consts to ICE if the type of the two consts are definitely not going to end up being equal to eachother. This might return Ok even if the types are not equal, but will never return Err if the types might be equal.

stripped_cfg_items: ()

Get all item paths that were stripped by a #[cfg] in a particular crate. Should not be called for the local crate before the resolver outputs are created, as it is only fed there.

generics_require_sized_self: ()

[query description - consider adding a doc-comment!] check whether the item has a where Self: Sized bound

cross_crate_inlinable: ()

[query description - consider adding a doc-comment!] whether the item should be made inlinable across crates

find_field: ()

[query description - consider adding a doc-comment!] find the index of maybe nested field {ident} in tcx.def_path_str(def_id)