Module rustc_middle::thir

THIR datatypes and definitions. See the rustc dev guide for more info.

If you compare the THIR ExprKind to hir::ExprKind, you will see it is a good bit simpler. In fact, a number of the more straight-forward MIR simplifications are already done in the lowering to THIR. For example, method calls and overloaded operators are absent: they are expected to be converted into ExprKind::Call instances.

Modules

size_asserts 🔒

visit

Macros

thir_with_elements 🔒

Structs

AdtExpr

Arm

A match arm.

ArmId

Ascription

Block

BlockId

ClosureExpr

Expr

A THIR expression.

ExprId

FieldExpr

Represents the association of a field identifier and an expression.

FieldPat

FruInfo

InlineAsmExpr

LocalVarId

Param

Description of a type-checked function parameter.

ParamId

Pat

PatRange

Stmt

StmtId

Thir

A container for a THIR body.

Enums

BindingMode

BlockSafety

BodyTy

ExprKind

Guard

A match guard.

InlineAsmOperand

LintLevel

LogicalOp

PatKind

StmtKind

Constants

UPVAR_ENV_PARAM

Type Aliases

UserTy 🔒