Module rustc_middle::traits::query

Experimental types for the trait query interface. The methods defined in this module are all based on canonicalization, which makes a canonical query by replacing unbound inference variables and regions, so that results can be reused more broadly. The providers for the queries defined here can be found in rustc_traits.

Modules

type_op

Structs

CandidateStep

DropckConstraint

A set of constraints that need to be satisfied in order for a type to be valid for destruction.

DropckOutlivesResult

MethodAutoderefBadTy

MethodAutoderefStepsResult

NoSolution

NormalizationResult

Result of the normalize_canonicalized_{{,inherent_}projection,weak}_ty queries.

Enums

OutlivesBound

Outlives bounds are relationships between generic parameters, whether they both be regions ('a: 'b) or whether types are involved (T: 'a). These relationships can be extracted from the full set of predicates we understand or also from types (in which case they are called implied bounds). They are fed to the OutlivesEnv which in turn is supplied to the region checker and other parts of the inference system.

Type Aliases

CanonicalAliasGoal

CanonicalPredicateGoal

CanonicalTyGoal

CanonicalTypeOpAscribeUserTypeGoal

CanonicalTypeOpEqGoal

CanonicalTypeOpNormalizeGoal

CanonicalTypeOpProvePredicateGoal

CanonicalTypeOpSubtypeGoal