rustc_middle::ty

Module adjustment

Structs

Adjustment

Represents coercing a value to a different type of value.

CoerceUnsizedInfo

Information for CoerceUnsized impls, storing information we have computed about the coercion.

OverloadedDeref

An overloaded autoderef step, representing a Deref(Mut)::deref(_mut) call, with the signature &'a T -> &'a U or &'a mut T -> &'a mut U. The target type is U in both cases, with the region and mutability being those shared by both the receiver and the returned reference.

PatAdjustment

Represents an implicit coercion applied to the scrutinee of a match before testing a pattern against it. Currently, this is used only for implicit dereferences.

Enums

Adjust

AllowTwoPhase

At least for initial deployment, we want to limit two-phase borrows to only a few specific cases. Right now, those are mostly “things that desugar” into method calls:

AutoBorrow

AutoBorrowMutability

CustomCoerceUnsized

PatAdjust

Represents implicit coercions of patterns’ types, rather than values’ types.

PointerCoercion