Struct rustc_middle::ty::consts::int::ScalarInt

#[repr(packed)]
pub struct ScalarInt {
    data: u128,
    size: NonZeroU8,
}

The raw bytes of a simple value.

This is a packed struct in order to allow this type to be optimally embedded in enums (like Scalar).

Fields
data: u128

The first size bytes of data are the value. Do not try to read less or more bytes than that. The remaining bytes must be 0.

size: NonZeroU8
Implementations

impl ScalarInt

pub const TRUE: ScalarInt = _

pub const FALSE: ScalarInt = _

pub fn size(self) -> Size

fn check_data(self)

Make sure the data fits in size. This is guaranteed by all constructors here, but having had this check saved us from bugs many times in the past, so keeping it around is definitely worth it.

pub fn null(size: Size) -> Self

pub fn is_null(self) -> bool

pub fn try_from_uint(i: impl Into<u128>, size: Size) -> Option<Self>

pub fn try_from_int(i: impl Into<i128>, size: Size) -> Option<Self>

pub fn assert_bits(self, target_size: Size) -> u128

pub fn to_bits(self, target_size: Size) -> Result<u128, Size>

pub fn try_to_target_usize(&self, tcx: TyCtxt<'_>) -> Result<u64, Size>

pub fn try_to_uint(self, size: Size) -> Result<u128, Size>

Tries to convert the ScalarInt to an unsigned integer of the given size. Fails if the size of the ScalarInt is not equal to size and returns the ScalarInts size in that case.

pub fn try_to_bool(self) -> Result<bool, Size>

pub fn try_to_u8(self) -> Result<u8, Size>

pub fn try_to_u16(self) -> Result<u16, Size>

Tries to convert the ScalarInt to u16. Fails if the size of the ScalarInt in not equal to Size { raw: 2 } and returns the size value of the ScalarInt in that case.

pub fn try_to_u32(self) -> Result<u32, Size>

Tries to convert the ScalarInt to u32. Fails if the size of the ScalarInt in not equal to Size { raw: 4 } and returns the size value of the ScalarInt in that case.

pub fn try_to_u64(self) -> Result<u64, Size>

Tries to convert the ScalarInt to u64. Fails if the size of the ScalarInt in not equal to Size { raw: 8 } and returns the size value of the ScalarInt in that case.

pub fn try_to_u128(self) -> Result<u128, Size>

Tries to convert the ScalarInt to u128. Fails if the size of the ScalarInt in not equal to Size { raw: 16 } and returns the size value of the ScalarInt in that case.

pub fn try_to_int(self, size: Size) -> Result<i128, Size>

Tries to convert the ScalarInt to a signed integer of the given size. Fails if the size of the ScalarInt is not equal to size and returns the ScalarInts size in that case.

pub fn try_to_i8(self) -> Result<i8, Size>

Tries to convert the ScalarInt to i8. Fails if the size of the ScalarInt is not equal to Size { raw: 1 } and returns the ScalarInts size in that case.

pub fn try_to_i16(self) -> Result<i16, Size>

Tries to convert the ScalarInt to i16. Fails if the size of the ScalarInt is not equal to Size { raw: 2 } and returns the ScalarInts size in that case.

pub fn try_to_i32(self) -> Result<i32, Size>

Tries to convert the ScalarInt to i32. Fails if the size of the ScalarInt is not equal to Size { raw: 4 } and returns the ScalarInts size in that case.

pub fn try_to_i64(self) -> Result<i64, Size>

Tries to convert the ScalarInt to i64. Fails if the size of the ScalarInt is not equal to Size { raw: 8 } and returns the ScalarInts size in that case.

pub fn try_to_i128(self) -> Result<i128, Size>

Tries to convert the ScalarInt to i128. Fails if the size of the ScalarInt is not equal to Size { raw: 16 } and returns the ScalarInts size in that case.