rustc_middle::ty::consts::kind
Trait UnevaluatedConstEvalExt

pub(crate) trait UnevaluatedConstEvalExt<'tcx> {
    // Required method
    fn prepare_for_eval(
        self,
        tcx: TyCtxt<'tcx>,
        param_env: ParamEnv<'tcx>,
    ) -> (ParamEnv<'tcx>, Self);
}

Required Methods

fn prepare_for_eval( self, tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>, ) -> (ParamEnv<'tcx>, Self)

FIXME(RalfJung): I cannot explain what this does or why it makes sense, but not doing this hurts performance.

Object Safety

This trait is not object safe.