Module rustc_middle::ty::context

Type context book-keeping.

Modules

tls

Macros

bidirectional_lang_item_map 🔒

direct_interners 🔒

nop_lift 🔒

nop_list_lift 🔒

nop_slice_lift 🔒

slice_interners 🔒

sty_debug_print 🔒

Structs

CommonConsts

CommonLifetimes

CommonTypes

CtxtInterners

CurrentGcx

This is used to get a reference to a GlobalCtxt if one is available.

DeducedParamAttrs

Parameter attributes that can only be determined by examining the body of a function instead of just its signature.

Feed

The same as TyCtxtFeed, but does not contain a TyCtxt. Use this to pass around when you have a TyCtxt elsewhere. Just an optimization to save space and not store hundreds of TyCtxtFeed in the resolver.

FreeRegionInfo

This struct contains information regarding a free parameter region, either a ReEarlyParam or ReLateParam.

GlobalCtxt

See TyCtxt for details about this type.

InternedInSet 🔒

TyCtxt

The central data structure of the compiler. It stores references to the various arenas and also houses the results of the various compiler queries that have been performed. See the rustc dev guide for more details.

TyCtxtFeed

This struct should only be created by create_def.

Constants

NUM_PREINTERNED_FRESH_FLOAT_TYS 🔒

NUM_PREINTERNED_FRESH_INT_TYS 🔒

NUM_PREINTERNED_FRESH_TYS 🔒

NUM_PREINTERNED_RE_LATE_BOUNDS_I 🔒

NUM_PREINTERNED_RE_LATE_BOUNDS_V 🔒

NUM_PREINTERNED_RE_VARS 🔒

NUM_PREINTERNED_TY_VARS 🔒

Traits

Lift

A trait implemented for all X<'a> types that can be safely and efficiently converted to X<'tcx> as long as they are part of the provided TyCtxt<'tcx>. This can be done, for example, for Ty<'tcx> or GenericArgsRef<'tcx> by looking them up in their respective interners.

Functions

_assert_tcx_fields 🔒

lang_item_to_trait_lang_item 🔒

provide

trait_lang_item_to_lang_item 🔒

Type Aliases

InternedSet 🔒