Module rustc_middle::ty::generic_args

Structs

ArgFolder 🔒

EarlyBinder

Similar to super::Binder except that it tracks early bound generics, i.e. struct Foo<T>(T) needs T instantiated immediately. This type primarily exists to avoid forgetting to call instantiate.

EarlyBinderIter

GenericArg

An entity in the Rust type system, which can be one of several kinds (types, lifetimes, and consts). To reduce memory usage, a GenericArg is an interned pointer, with the lowest 2 bits being reserved for a tag to indicate the type (Ty, Region, or Const) it points to.

IterInstantiated

IterInstantiatedCopied

UserArgs

Stores the user-given args to reach some fully qualified path (e.g., <T>::Item or <T as Trait>::Item).

UserSelfTy

Specifies the user-given self type. In the case of a path that refers to a member in an inherent impl, this self type is sometimes needed to constrain the type parameters on the impl. For example, in this code:

Enums

GenericArgKind

Constants

CONST_TAG 🔒

REGION_TAG 🔒

TAG_MASK 🔒

TYPE_TAG 🔒

Type Aliases

GenericArgs

List of generic arguments that are gonna be used to replace generic parameters.

GenericArgsRef