rustc_middle::ty

Module generic_args

Structs

GenericArg

An entity in the Rust type system, which can be one of several kinds (types, lifetimes, and consts). To reduce memory usage, a GenericArg is an interned pointer, with the lowest 2 bits being reserved for a tag to indicate the type (Ty, Region, or Const) it points to.

UserArgs

Stores the user-given args to reach some fully qualified path (e.g., <T>::Item or <T as Trait>::Item).

UserSelfTy

Specifies the user-given self type. In the case of a path that refers to a member in an inherent impl, this self type is sometimes needed to constrain the type parameters on the impl. For example, in this code:

Constants

CONST_TAG 🔒

REGION_TAG 🔒

TAG_MASK 🔒

TYPE_TAG 🔒

Traits

GenericArgPackExt 🔒

Type Aliases

GenericArgKind

GenericArgs

List of generic arguments that are gonna be used to replace generic parameters.

GenericArgsRef

TermKind