rustc_middle::ty
Module layout

Structs

LayoutCx

Enums

FnAbiError

Error produced by attempting to compute or adjust a FnAbi.

FnAbiRequest

LayoutError

SizeSkeleton

Type size “skeleton”, i.e., the only information determining a type’s size. While this is conservative, (aside from constant sizes, only pointers, newtypes thereof and null pointer optimized enums are allowed), it is enough to statically check common use cases of transmute.

ValidityRequirement

Used in check_validity_requirement to indicate the kind of initialization that is checked to be valid

Constants

MAX_SIMD_LANES

The maximum supported number of lanes in a SIMD vector.

WIDE_PTR_ADDR

The first half of a wide pointer.

WIDE_PTR_EXTRA

The second half of a wide pointer.

Traits

FloatExt

FnAbiOf

Blanket extension trait for contexts that can compute FnAbis.

FnAbiOfHelpers

Trait for contexts that want to be able to compute FnAbis. This automatically gives access to FnAbiOf, through a blanket impl.

HasParamEnv

HasTyCtxt

IntegerExt

LayoutOf

Blanket extension trait for contexts that can compute layouts of types.

LayoutOfHelpers

Trait for contexts that want to be able to compute layouts of types. This automatically gives access to LayoutOf, through a blanket impl.

MaybeResult

PrimitiveExt

Functions

fn_can_unwind

Calculates whether a function’s ABI can unwind or not.

Type Aliases

TyAndLayout