Struct rustc_middle::ty::AdtDef

pub struct AdtDef<'tcx>(pub Interned<'tcx, AdtDefData>);

Tuple Fields
0: Interned<'tcx, AdtDefData>
Implementations

impl<'tcx> AdtDef<'tcx>

pub fn did(self) -> DefId

pub fn variants(self) -> &'tcx IndexSlice<VariantIdx, VariantDef>

pub fn variant(self, idx: VariantIdx) -> &'tcx VariantDef

pub fn flags(self) -> AdtFlags

pub fn repr(self) -> ReprOptions

impl<'tcx> AdtDef<'tcx>

pub fn is_struct(self) -> bool

Returns true if this is a struct.

pub fn is_union(self) -> bool

Returns true if this is a union.

pub fn is_enum(self) -> bool

Returns true if this is an enum.

pub fn is_variant_list_non_exhaustive(self) -> bool

Returns true if the variant list of this ADT is #[non_exhaustive].

pub fn adt_kind(self) -> AdtKind

Returns the kind of the ADT.

pub fn descr(self) -> &'static str

Returns a description of this abstract data type.

pub fn variant_descr(self) -> &'static str

Returns a description of a variant of this abstract data type.

pub fn has_ctor(self) -> bool

If this function returns true, it implies that is_struct must return true.

pub fn is_fundamental(self) -> bool

Returns true if this type is #[fundamental] for the purposes of coherence checking.

pub fn is_phantom_data(self) -> bool

Returns true if this is PhantomData<T>.

pub fn is_box(self) -> bool

Returns true if this is Box<T>.

pub fn is_unsafe_cell(self) -> bool

Returns true if this is UnsafeCell<T>.

pub fn is_manually_drop(self) -> bool

Returns true if this is ManuallyDrop<T>.

pub fn has_dtor(self, tcx: TyCtxt<'tcx>) -> bool

Returns true if this type has a destructor.

pub fn has_non_const_dtor(self, tcx: TyCtxt<'tcx>) -> bool

pub fn non_enum_variant(self) -> &'tcx VariantDef

Asserts this is a struct or union and returns its unique variant.

pub fn predicates(self, tcx: TyCtxt<'tcx>) -> GenericPredicates<'tcx>

pub fn all_fields(self) -> impl Iterator<Item = &'tcx FieldDef> + Clone

Returns an iterator over all fields contained by this ADT.

pub fn is_payloadfree(self) -> bool

Whether the ADT lacks fields. Note that this includes uninhabited enums, e.g., enum Void {} is considered payload free as well.

pub fn variant_with_id(self, vid: DefId) -> &'tcx VariantDef

Return a VariantDef given a variant id.

pub fn variant_with_ctor_id(self, cid: DefId) -> &'tcx VariantDef

Return a VariantDef given a constructor id.

pub fn variant_index_with_id(self, vid: DefId) -> VariantIdx

Return the index of VariantDef given a variant id.

pub fn variant_index_with_ctor_id(self, cid: DefId) -> VariantIdx

Return the index of VariantDef given a constructor id.

pub fn variant_of_res(self, res: Res) -> &'tcx VariantDef

pub fn eval_explicit_discr( self, tcx: TyCtxt<'tcx>, expr_did: DefId ) -> Option<Discr<'tcx>>

pub fn discriminants( self, tcx: TyCtxt<'tcx> ) -> impl Iterator<Item = (VariantIdx, Discr<'tcx>)> + Captures<'tcx>

pub fn variant_range(self) -> Range<VariantIdx>

pub fn discriminant_for_variant( self, tcx: TyCtxt<'tcx>, variant_index: VariantIdx ) -> Discr<'tcx>

Computes the discriminant value used by a specific variant. Unlike discriminants, this is (amortized) constant-time, only doing at most one query for evaluating an explicit discriminant (the last one before the requested variant), assuming there are no constant-evaluation errors there.

pub fn discriminant_def_for_variant( self, variant_index: VariantIdx ) -> (Option<DefId>, u32)

Yields a DefId for the discriminant and an offset to add to it Alternatively, if there is no explicit discriminant, returns the inferred discriminant directly.

pub fn destructor(self, tcx: TyCtxt<'tcx>) -> Option<Destructor>

pub fn sized_constraint( self, tcx: TyCtxt<'tcx> ) -> EarlyBinder<&'tcx List<Ty<'tcx>>>

Returns a list of types such that Self: Sized if and only if that type is Sized, or ty::Error if this type has a recursive layout.