rustc_middle::ty
Struct Region

pub struct Region<'tcx>(pub Interned<'tcx, RegionKind<'tcx>>);

Use this rather than RegionKind, whenever possible.

Tuple Fields
0: Interned<'tcx, RegionKind<'tcx>>
Implementations

impl<'tcx> Region<'tcx>

pub fn new_early_param( tcx: TyCtxt<'tcx>, early_bound_region: EarlyParamRegion, ) -> Region<'tcx>

pub fn new_bound( tcx: TyCtxt<'tcx>, debruijn: DebruijnIndex, bound_region: BoundRegion, ) -> Region<'tcx>

pub fn new_late_param( tcx: TyCtxt<'tcx>, scope: DefId, bound_region: BoundRegionKind, ) -> Region<'tcx>

pub fn new_var(tcx: TyCtxt<'tcx>, v: RegionVid) -> Region<'tcx>

pub fn new_placeholder( tcx: TyCtxt<'tcx>, placeholder: PlaceholderRegion, ) -> Region<'tcx>

pub fn new_error(tcx: TyCtxt<'tcx>, guar: ErrorGuaranteed) -> Region<'tcx>

Constructs a RegionKind::ReError region.

pub fn new_error_misc(tcx: TyCtxt<'tcx>) -> Region<'tcx>

Constructs a RegionKind::ReError region and registers a delayed bug to ensure it gets used.

pub fn new_error_with_message<S: Into<MultiSpan>>( tcx: TyCtxt<'tcx>, span: S, msg: &'static str, ) -> Region<'tcx>

Constructs a RegionKind::ReError region and registers a delayed bug with the given msg to ensure it gets used.

pub fn new_from_kind(tcx: TyCtxt<'tcx>, kind: RegionKind<'tcx>) -> Region<'tcx>

Avoid this in favour of more specific new_* methods, where possible, to avoid the cost of the match.

impl<'tcx> Region<'tcx>

Region utilities

pub fn kind(self) -> RegionKind<'tcx>

pub fn get_name(self) -> Option<Symbol>

pub fn get_name_or_anon(self) -> Symbol

pub fn has_name(self) -> bool

Is this region named by the user?

pub fn is_error(self) -> bool

pub fn is_static(self) -> bool

pub fn is_erased(self) -> bool

pub fn is_bound(self) -> bool

pub fn is_placeholder(self) -> bool

pub fn bound_at_or_above_binder(self, index: DebruijnIndex) -> bool

pub fn type_flags(self) -> TypeFlags

pub fn is_param(self) -> bool

True for free regions other than 'static.

pub fn is_free(self) -> bool

True for free region in the current context.

This is the case for 'static and param regions.

pub fn is_var(self) -> bool

pub fn as_var(self) -> RegionVid

pub fn opt_param_def_id( self, tcx: TyCtxt<'tcx>, binding_item: DefId, ) -> Option<DefId>

Given some item binding_item, check if this region is a generic parameter introduced by it or one of the parent generics. Returns the DefId of the parameter definition if so.