Trait rustc_middle::ty::TypeVisitor

pub trait TypeVisitor<I>: Sized
where
    I: Interner,
{
    type Result: VisitorResult = ();

    // Provided methods
    fn visit_binder<T>(
        &mut self,
        t: &<I as Interner>::Binder<T>
    ) -> Self::Result
       where T: TypeVisitable<I> { ... }
    fn visit_ty(&mut self, t: <I as Interner>::Ty) -> Self::Result { ... }
    fn visit_region(&mut self, _r: <I as Interner>::Region) -> Self::Result { ... }
    fn visit_const(&mut self, c: <I as Interner>::Const) -> Self::Result { ... }
    fn visit_predicate(&mut self, p: <I as Interner>::Predicate) -> Self::Result { ... }
}

This trait is implemented for every visiting traversal. There is a visit method defined for every type of interest. Each such method has a default that recurses into the type’s fields in a non-custom fashion.

Provided Associated Types

type Result: VisitorResult = ()

Provided Methods

fn visit_binder<T>(&mut self, t: &<I as Interner>::Binder<T>) -> Self::Result
where T: TypeVisitable<I>,

fn visit_ty(&mut self, t: <I as Interner>::Ty) -> Self::Result

fn visit_region(&mut self, _r: <I as Interner>::Region) -> Self::Result

fn visit_const(&mut self, c: <I as Interner>::Const) -> Self::Result

fn visit_predicate(&mut self, p: <I as Interner>::Predicate) -> Self::Result

Object Safety

This trait is not object safe.