rustc_middle::ty

Module typeck_results

Structs

CanonicalUserTypeAnnotation

LocalSetInContext

LocalSetInContextMut

LocalTableInContext

LocalTableInContextMut

Rust2024IncompatiblePatInfo

Information on a pattern incompatible with Rust 2024, for use by the error/migration diagnostic emitted during THIR construction.

TypeckResults

UserType

UserTypeAnnotationIndex

Enums

UserTypeKind

A user-given type annotation attached to a constant. These arise from constants that are named via paths, like Foo::<A>::new and so forth.

Constants

START_INDEX

Traits

IsIdentity

Functions

invalid_hir_id_for_typeck_results 🔒

validate_hir_id_for_typeck_results 🔒

Validate that the given HirId (respectively its local_id part) can be safely used as a key in the maps of a TypeckResults. For that to be the case, the HirId must have the same owner as all the other IDs in this table (signified by hir_owner). Otherwise the HirId would be in a different frame of reference and using its local_id would result in lookup errors, or worse, in silently wrong data being stored/returned.

Type Aliases

CanonicalUserType

Canonical user type annotation.

CanonicalUserTypeAnnotations

Mapping of type annotation indices to canonical user type annotations.