Module rustc_middle::ty::util

Miscellaneous type-system utilities that are too small to deserve their own modules.

Structs

AlwaysRequiresDrop

Discr

OpaqueTypeExpander 🔒

Enums

ExplicitSelf

IgnoreRegions

Used as an input to TyCtxt::uses_unique_generic_params.

NotUniqueParam

Traits

IntTypeExt

Functions

fold_list

int_size_and_signed 🔒

is_doc_hidden

Determines whether an item is annotated with doc(hidden).

is_intrinsic

Determines whether an item is an intrinsic by Abi.

is_trivially_const_drop

needs_drop_components

Returns a list of types such that the given type needs drop if and only if any of the returned types need drop. Returns Err(AlwaysRequiresDrop) if this type always needs drop.

normalize_opaque_types

Normalizes all opaque types in the given value, replacing them with their underlying types.

provide