Trait rustc_middle::ty::visit::TypeSuperVisitable

pub trait TypeSuperVisitable<'tcx>: TypeVisitable<'tcx> {
    fn super_visit_with<V: TypeVisitor<'tcx>>(
        &self,
        visitor: &mut V
    ) -> ControlFlow<V::BreakTy>;
}

Required Methods

fn super_visit_with<V: TypeVisitor<'tcx>>(
&self,
visitor: &mut V
) -> ControlFlow<V::BreakTy>

Provides a default visit for a type of interest. This should only be called within TypeVisitor methods, when a non-custom traversal is desired for the value of the type of interest passed to that method. For example, in MyVisitor::visit_ty(ty), it is valid to call ty.super_visit_with(self), but any other visiting should be done with xyz.visit_with(self).