rustc_middle::ty
Module vtable

Enums

VtblEntry

Constants

COMMON_VTABLE_ENTRIES_ALIGN

COMMON_VTABLE_ENTRIES_DROPINPLACE

COMMON_VTABLE_ENTRIES_SIZE

Functions

vtable_allocation_provider 🔒

Retrieves an allocation that represents the contents of a vtable. Since this is a query, allocations are cached and not duplicated.

vtable_min_entries 🔒