rustc_middle::ty::walk
Function push_inner

fn push_inner<'tcx>(
    stack: &mut SmallVec<[GenericArg<'tcx>; 8]>,
    parent: GenericArg<'tcx>,
)

We push GenericArgs on the stack in reverse order so as to maintain a pre-order traversal. As of the time of this writing, the fact that the traversal is pre-order is not known to be significant to any code, but it seems like the natural order one would expect (basically, the order of the types as they are written).